Network.AWS.ElasticBeanstalk

Description

AWS Elastic Beanstalk is an easy-to-use service for deploying and scaling web applications and services developed with Java, .NET, PHP, Node.js, Python, Ruby, and Docker on familiar servers such as Apache HTTP Server, Apache Tomcat, Nginx, Passenger, and IIS 7.5/8. You can simply upload your code and Elastic Beanstalk automatically handles the deployment, from capacity provisioning, load balancing, auto-scaling to application health monitoring. At the same time, you retain full control over the AWS resources powering your application and can access the underlying resources at any time. There is no additional charge for Elastic Beanstalk - you pay only for the AWS resources needed to store and run your applications.

Documentation

module Network.AWS.ElasticBeanstalk.AbortEnvironmentUpdate

module Network.AWS.ElasticBeanstalk.CheckDNSAvailability

module Network.AWS.ElasticBeanstalk.CreateApplication

module Network.AWS.ElasticBeanstalk.CreateApplicationVersion

module Network.AWS.ElasticBeanstalk.CreateConfigurationTemplate

module Network.AWS.ElasticBeanstalk.CreateEnvironment

module Network.AWS.ElasticBeanstalk.CreateStorageLocation

module Network.AWS.ElasticBeanstalk.DeleteApplication

module Network.AWS.ElasticBeanstalk.DeleteApplicationVersion

module Network.AWS.ElasticBeanstalk.DeleteConfigurationTemplate

module Network.AWS.ElasticBeanstalk.DeleteEnvironmentConfiguration

module Network.AWS.ElasticBeanstalk.DescribeApplicationVersions

module Network.AWS.ElasticBeanstalk.DescribeApplications

module Network.AWS.ElasticBeanstalk.DescribeConfigurationOptions

module Network.AWS.ElasticBeanstalk.DescribeConfigurationSettings

module Network.AWS.ElasticBeanstalk.DescribeEnvironmentResources

module Network.AWS.ElasticBeanstalk.DescribeEnvironments

module Network.AWS.ElasticBeanstalk.DescribeEvents

module Network.AWS.ElasticBeanstalk.ListAvailableSolutionStacks

module Network.AWS.ElasticBeanstalk.RebuildEnvironment

module Network.AWS.ElasticBeanstalk.RequestEnvironmentInfo

module Network.AWS.ElasticBeanstalk.RestartAppServer

module Network.AWS.ElasticBeanstalk.RetrieveEnvironmentInfo

module Network.AWS.ElasticBeanstalk.SwapEnvironmentCNAMEs

module Network.AWS.ElasticBeanstalk.TerminateEnvironment

module Network.AWS.ElasticBeanstalk.Types

module Network.AWS.ElasticBeanstalk.UpdateApplication

module Network.AWS.ElasticBeanstalk.UpdateApplicationVersion

module Network.AWS.ElasticBeanstalk.UpdateConfigurationTemplate

module Network.AWS.ElasticBeanstalk.UpdateEnvironment

module Network.AWS.ElasticBeanstalk.ValidateConfigurationSettings