Network.AWS.ElasticBeanstalk.ComposeEnvironments

Description

Create or update a group of environments that each run a separate component of a single application. Takes a list of version labels that specify application source bundles for each of the environments to create or update. The name of each environment and other required information must be included in the source bundles in an environment manifest named env.yaml . See Compose Environments for details.

Creating a Request

composeEnvironments :: ComposeEnvironments

Creates a value of ComposeEnvironments with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ceVersionLabels - A list of version labels, specifying one or more application source bundles that belong to the target application. Each source bundle must include an environment manifest that specifies the name of the environment and the name of the solution stack to use, and optionally can specify environment links to create.

ceApplicationName - The name of the application to which the specified source bundles belong.

ceGroupName - The name of the group to which the target environments belong. Specify a group name only if the environment name defined in each target environment's manifest ends with a + (plus) character. See Environment Manifest (env.yaml) for details.

data ComposeEnvironments

Request to create or update a group of environments.

See: composeEnvironments smart constructor.

Request Lenses

ceVersionLabels :: Lens' ComposeEnvironments [Text]

A list of version labels, specifying one or more application source bundles that belong to the target application. Each source bundle must include an environment manifest that specifies the name of the environment and the name of the solution stack to use, and optionally can specify environment links to create.

ceApplicationName :: Lens' ComposeEnvironments (Maybe Text)

The name of the application to which the specified source bundles belong.

ceGroupName :: Lens' ComposeEnvironments (Maybe Text)

The name of the group to which the target environments belong. Specify a group name only if the environment name defined in each target environment's manifest ends with a + (plus) character. See Environment Manifest (env.yaml) for details.

Destructuring the Response

environmentDescriptionsMessage :: EnvironmentDescriptionsMessage

Creates a value of EnvironmentDescriptionsMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

edmNextToken - In a paginated request, the token that you can pass in a subsequent request to get the next response page.

edmEnvironments - Returns an EnvironmentDescription list.

data EnvironmentDescriptionsMessage

Result message containing a list of environment descriptions.

See: environmentDescriptionsMessage smart constructor.

Response Lenses

edmNextToken :: Lens' EnvironmentDescriptionsMessage (Maybe Text)

In a paginated request, the token that you can pass in a subsequent request to get the next response page.

edmEnvironments :: Lens' EnvironmentDescriptionsMessage [EnvironmentDescription]

Returns an EnvironmentDescription list.