Network.AWS.ElasticBeanstalk.CreateApplication

Description

Creates an application that has one configuration template named default and no application versions.

Creating a Request

createApplication

Arguments

:: Text
caApplicationName

-> CreateApplication

Creates a value of CreateApplication with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caResourceLifecycleConfig - Specify an application resource lifecycle configuration to prevent your application from accumulating too many versions.

caDescription - Describes the application.

caApplicationName - The name of the application. Constraint: This name must be unique within your account. If the specified name already exists, the action returns an InvalidParameterValue error.

data CreateApplication

Request to create an application.

See: createApplication smart constructor.

Request Lenses

caResourceLifecycleConfig :: Lens' CreateApplication (Maybe ApplicationResourceLifecycleConfig)

Specify an application resource lifecycle configuration to prevent your application from accumulating too many versions.

caDescription :: Lens' CreateApplication (Maybe Text)

Describes the application.

caApplicationName :: Lens' CreateApplication Text

The name of the application. Constraint: This name must be unique within your account. If the specified name already exists, the action returns an InvalidParameterValue error.

Destructuring the Response

applicationDescriptionMessage :: ApplicationDescriptionMessage

Creates a value of ApplicationDescriptionMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

admApplication - The ApplicationDescription of the application.

data ApplicationDescriptionMessage

Result message containing a single description of an application.

See: applicationDescriptionMessage smart constructor.

Response Lenses

admApplication :: Lens' ApplicationDescriptionMessage (Maybe ApplicationDescription)

The ApplicationDescription of the application.