Network.AWS.ElasticBeanstalk.CreateApplicationVersion

Description

Creates an application version for the specified application.

Once you create an application version with a specified Amazon S3 bucket and key location, you cannot change that Amazon S3 location. If you change the Amazon S3 location, you receive an exception when you attempt to launch an environment from the application version.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_CreateApplicationVersion.html

Request

data CreateApplicationVersion

Request constructor

createApplicationVersion

Arguments

:: Text
cavApplicationName

-> Text
cavVersionLabel

-> CreateApplicationVersion

CreateApplicationVersion constructor.

The fields accessible through corresponding lenses are:

cavApplicationName :: Text

cavAutoCreateApplication :: Maybe Bool

cavDescription :: Maybe Text

cavSourceBundle :: Maybe S3Location

cavVersionLabel :: Text

Request lenses

cavApplicationName :: Lens' CreateApplicationVersion Text

The name of the application. If no application is found with this name, and AutoCreateApplication is false, returns an InvalidParameterValue error.

cavAutoCreateApplication :: Lens' CreateApplicationVersion (Maybe Bool)

Determines how the system behaves if the specified application for this version does not already exist:

true: Automatically creates the specified application for this version if it does not already exist.

false: Returns an InvalidParameterValue if the specified application for this version does not already exist.

true : Automatically creates the specified application for this release if it does not already exist. false : Throws an InvalidParameterValue if the specified application for this release does not already exist. Default: false

Valid Values: true | false

cavDescription :: Lens' CreateApplicationVersion (Maybe Text)

Describes this version.

cavSourceBundle :: Lens' CreateApplicationVersion (Maybe S3Location)

The Amazon S3 bucket and key that identify the location of the source bundle for this version.

If data found at the Amazon S3 location exceeds the maximum allowed source bundle size, AWS Elastic Beanstalk returns an InvalidParameterValue error. The maximum size allowed is 512 MB.

Default: If not specified, AWS Elastic Beanstalk uses a sample application. If only partially specified (for example, a bucket is provided but not the key) or if no data is found at the Amazon S3 location, AWS Elastic Beanstalk returns an InvalidParameterCombination error.

cavVersionLabel :: Lens' CreateApplicationVersion Text

A label identifying this version.

Constraint: Must be unique per application. If an application version already exists with this label for the specified application, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Response

data CreateApplicationVersionResponse

Response constructor

createApplicationVersionResponse :: CreateApplicationVersionResponse

CreateApplicationVersionResponse constructor.

The fields accessible through corresponding lenses are:

cavrApplicationVersion :: Maybe ApplicationVersionDescription

Response lenses

cavrApplicationVersion :: Lens' CreateApplicationVersionResponse (Maybe ApplicationVersionDescription)

The ApplicationVersionDescription of the application version.