Network.AWS.ElasticBeanstalk.CreateConfigurationTemplate

Description

Creates a configuration template. Templates are associated with a specific application and are used to deploy different versions of the application with the same configuration settings. Related Topics DescribeConfigurationOptions> DescribeConfigurationSettings> ListAvailableSolutionStacks>.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_CreateConfigurationTemplate.html

Request

data CreateConfigurationTemplate

Request constructor

createConfigurationTemplate

Arguments

:: Text
cctApplicationName

-> Text
cctTemplateName

-> CreateConfigurationTemplate

CreateConfigurationTemplate constructor.

The fields accessible through corresponding lenses are:

cctApplicationName :: Text

cctDescription :: Maybe Text

cctEnvironmentId :: Maybe Text

cctOptionSettings :: [ConfigurationOptionSetting]

cctSolutionStackName :: Maybe Text

cctSourceConfiguration :: Maybe SourceConfiguration

cctTemplateName :: Text

Request lenses

cctApplicationName :: Lens' CreateConfigurationTemplate Text

The name of the application to associate with this configuration template. If no application is found with this name, AWS Elastic Beanstalk returns an InvalidParameterValue error.

cctDescription :: Lens' CreateConfigurationTemplate (Maybe Text)

Describes this configuration.

cctEnvironmentId :: Lens' CreateConfigurationTemplate (Maybe Text)

The ID of the environment used with this configuration template.

cctOptionSettings :: Lens' CreateConfigurationTemplate [ConfigurationOptionSetting]

If specified, AWS Elastic Beanstalk sets the specified configuration option to the requested value. The new value overrides the value obtained from the solution stack or the source configuration template.

cctSolutionStackName :: Lens' CreateConfigurationTemplate (Maybe Text)

The name of the solution stack used by this configuration. The solution stack specifies the operating system, architecture, and application server for a configuration template. It determines the set of configuration options as well as the possible and default values. Use ListAvailableSolutionStacks> to obtain a list of available solution stacks. A solution stack name or a source configuration parameter must be specified, otherwise AWS Elastic Beanstalk returns an InvalidParameterValue error. If a solution stack name is not specified and the source configuration parameter is specified, AWS Elastic Beanstalk uses the same solution stack as the source configuration template.

cctSourceConfiguration :: Lens' CreateConfigurationTemplate (Maybe SourceConfiguration)

If specified, AWS Elastic Beanstalk uses the configuration values from the specified configuration template to create a new configuration. Values specified in the OptionSettings parameter of this call overrides any values obtained from the SourceConfiguration. If no configuration template is found, returns an InvalidParameterValue error. Constraint: If both the solution stack name parameter and the source configuration parameters are specified, the solution stack of the source configuration template must match the specified solution stack name or else AWS Elastic Beanstalk returns an InvalidParameterCombination error.

cctTemplateName :: Lens' CreateConfigurationTemplate Text

The name of the configuration template. Constraint: This name must be unique per application. Default: If a configuration template already exists with this name, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Response

data CreateConfigurationTemplateResponse

Response constructor

createConfigurationTemplateResponse :: CreateConfigurationTemplateResponse

CreateConfigurationTemplateResponse constructor.

The fields accessible through corresponding lenses are:

cctrApplicationName :: Maybe Text

cctrDateCreated :: Maybe UTCTime

cctrDateUpdated :: Maybe UTCTime

cctrDeploymentStatus :: Maybe ConfigurationDeploymentStatus

cctrDescription :: Maybe Text

cctrEnvironmentName :: Maybe Text

cctrOptionSettings :: [ConfigurationOptionSetting]

cctrSolutionStackName :: Maybe Text

cctrTemplateName :: Maybe Text

Response lenses

cctrApplicationName :: Lens' CreateConfigurationTemplateResponse (Maybe Text)

The name of the application associated with this configuration set.

cctrDateCreated :: Lens' CreateConfigurationTemplateResponse (Maybe UTCTime)

The date (in UTC time) when this configuration set was created.

cctrDateUpdated :: Lens' CreateConfigurationTemplateResponse (Maybe UTCTime)

The date (in UTC time) when this configuration set was last modified.

cctrDeploymentStatus :: Lens' CreateConfigurationTemplateResponse (Maybe ConfigurationDeploymentStatus)

If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set: null: This configuration is not associated with a running environment. pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying. deployed: This is the configuration that is currently deployed to the associated running environment. failed: This is a draft configuration, that failed to successfully deploy. null: This configuration is not associated with a running environment. pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying. deployed: This is the configuration that is currently deployed to the associated running environment. failed: This is a draft configuration that failed to successfully deploy.

cctrDescription :: Lens' CreateConfigurationTemplateResponse (Maybe Text)

Describes this configuration set.

cctrEnvironmentName :: Lens' CreateConfigurationTemplateResponse (Maybe Text)

If not null, the name of the environment for this configuration set.

cctrOptionSettings :: Lens' CreateConfigurationTemplateResponse [ConfigurationOptionSetting]

A list of the configuration options and their values in this configuration set.

cctrSolutionStackName :: Lens' CreateConfigurationTemplateResponse (Maybe Text)

The name of the solution stack this configuration set uses.

cctrTemplateName :: Lens' CreateConfigurationTemplateResponse (Maybe Text)

If not null, the name of the configuration template for this configuration set.