Network.AWS.ElasticBeanstalk.CreatePlatformVersion

Description

Create a new version of your custom platform.

Creating a Request

createPlatformVersion

Arguments

:: Text
cpvPlatformName

-> Text
cpvPlatformVersion

-> S3Location
cpvPlatformDefinitionBundle

-> CreatePlatformVersion

Creates a value of CreatePlatformVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpvOptionSettings - The configuration option settings to apply to the builder environment.

cpvEnvironmentName - The name of the builder environment.

cpvPlatformName - The name of your custom platform.

cpvPlatformVersion - The number, such as 1.0.2, for the new platform version.

cpvPlatformDefinitionBundle - The location of the platform definition archive in Amazon S3.

data CreatePlatformVersion

Request to create a new platform version.

See: createPlatformVersion smart constructor.

Request Lenses

cpvOptionSettings :: Lens' CreatePlatformVersion [ConfigurationOptionSetting]

The configuration option settings to apply to the builder environment.

cpvEnvironmentName :: Lens' CreatePlatformVersion (Maybe Text)

The name of the builder environment.

cpvPlatformName :: Lens' CreatePlatformVersion Text

The name of your custom platform.

cpvPlatformVersion :: Lens' CreatePlatformVersion Text

The number, such as 1.0.2, for the new platform version.

cpvPlatformDefinitionBundle :: Lens' CreatePlatformVersion S3Location

The location of the platform definition archive in Amazon S3.

Destructuring the Response

createPlatformVersionResponse

Arguments

:: Int
cpvrsResponseStatus

-> CreatePlatformVersionResponse

Creates a value of CreatePlatformVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpvrsBuilder - The builder used to create the custom platform.

cpvrsPlatformSummary - Detailed information about the new version of the custom platform.

cpvrsResponseStatus - -- | The response status code.

data CreatePlatformVersionResponse

See: createPlatformVersionResponse smart constructor.

Response Lenses

cpvrsBuilder :: Lens' CreatePlatformVersionResponse (Maybe Builder)

The builder used to create the custom platform.

cpvrsPlatformSummary :: Lens' CreatePlatformVersionResponse (Maybe PlatformSummary)

Detailed information about the new version of the custom platform.

cpvrsResponseStatus :: Lens' CreatePlatformVersionResponse Int

- | The response status code.