Network.AWS.ElasticBeanstalk.CreateStorageLocation

Description

Creates the Amazon S3 storage location for the account.

This location is used to store user log files.

Creating a Request

createStorageLocation :: CreateStorageLocation

Creates a value of CreateStorageLocation with the minimum fields required to make a request.

data CreateStorageLocation

See: createStorageLocation smart constructor.

Destructuring the Response

createStorageLocationResponse

Arguments

:: Int
cslrsResponseStatus

-> CreateStorageLocationResponse

Creates a value of CreateStorageLocationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cslrsS3Bucket - The name of the Amazon S3 bucket created.

cslrsResponseStatus - -- | The response status code.

data CreateStorageLocationResponse

Results of a CreateStorageLocationResult call.

See: createStorageLocationResponse smart constructor.

Response Lenses

cslrsS3Bucket :: Lens' CreateStorageLocationResponse (Maybe Text)

The name of the Amazon S3 bucket created.

cslrsResponseStatus :: Lens' CreateStorageLocationResponse Int

- | The response status code.