Network.AWS.ElasticBeanstalk.DeleteApplicationVersion

Description

Deletes the specified version from the specified application.

You cannot delete an application version that is associated with a running environment.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_DeleteApplicationVersion.html

Request

data DeleteApplicationVersion

Request constructor

deleteApplicationVersion

Arguments

:: Text
davApplicationName

-> Text
davVersionLabel

-> DeleteApplicationVersion

DeleteApplicationVersion constructor.

The fields accessible through corresponding lenses are:

davApplicationName :: Text

davDeleteSourceBundle :: Maybe Bool

davVersionLabel :: Text

Request lenses

davApplicationName :: Lens' DeleteApplicationVersion Text

The name of the application to delete releases from.

davDeleteSourceBundle :: Lens' DeleteApplicationVersion (Maybe Bool)

Indicates whether to delete the associated source bundle from Amazon S3:

true: An attempt is made to delete the associated Amazon S3 source bundle specified at time of creation. false: No action is taken on the Amazon S3 source bundle specified at time of creation. Valid Values: true | false

davVersionLabel :: Lens' DeleteApplicationVersion Text

The label of the version to delete.

Response

data DeleteApplicationVersionResponse

Response constructor

deleteApplicationVersionResponse :: DeleteApplicationVersionResponse

DeleteApplicationVersionResponse constructor.