Network.AWS.ElasticBeanstalk.DeleteConfigurationTemplate

Description

Deletes the specified configuration template.

When you launch an environment using a configuration template, the environment gets a copy of the template. You can delete or modify the environment's copy of the template without affecting the running environment.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_DeleteConfigurationTemplate.html

Request

data DeleteConfigurationTemplate

Request constructor

deleteConfigurationTemplate

Arguments

:: Text
dctApplicationName

-> Text
dctTemplateName

-> DeleteConfigurationTemplate

DeleteConfigurationTemplate constructor.

The fields accessible through corresponding lenses are:

dctApplicationName :: Text

dctTemplateName :: Text

Request lenses

dctApplicationName :: Lens' DeleteConfigurationTemplate Text

The name of the application to delete the configuration template from.

dctTemplateName :: Lens' DeleteConfigurationTemplate Text

The name of the configuration template to delete.

Response

data DeleteConfigurationTemplateResponse

Response constructor

deleteConfigurationTemplateResponse :: DeleteConfigurationTemplateResponse

DeleteConfigurationTemplateResponse constructor.