Network.AWS.ElasticBeanstalk.DeleteEnvironmentConfiguration

Description

Deletes the draft configuration associated with the running environment. Updating a running environment with any configuration changes creates a draft configuration set. You can get the draft configuration using DescribeConfigurationSettings while the update is in progress or if the update fails. The DeploymentStatus for the draft configuration indicates whether the deployment is in process or has failed. The draft configuration remains in existence until it is deleted with this action. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &EnvironmentName=SampleApp &Operation=DeleteEnvironmentConfiguration &AuthParams fdf76507-f26d-11df-8a78-9f77047e0d0c.

Request

data DeleteEnvironmentConfiguration

This documentation target is not reported in the API reference.

Request constructor

mkDeleteEnvironmentConfiguration

Arguments

:: Text
decApplicationName

-> Text
decEnvironmentName

-> DeleteEnvironmentConfiguration

Smart constructor for the minimum required parameters to construct a valid DeleteEnvironmentConfiguration request.

The fields accessible through corresponding lenses are:

ApplicationName :: Text

EnvironmentName :: Text

Request lenses

decApplicationName :: Lens' DeleteEnvironmentConfiguration Text

The name of the application the environment is associated with.

decEnvironmentName :: Lens' DeleteEnvironmentConfiguration Text

The name of the environment to delete the draft configuration from.

Response

data DeleteEnvironmentConfigurationResponse

Response constructor

mkDeleteEnvironmentConfigurationResponse :: DeleteEnvironmentConfigurationResponse

Smart constructor for the minimum required parameters to construct a valid DeleteEnvironmentConfigurationResponse response.

This constructor is provided for convenience and testing purposes.