Network.AWS.ElasticBeanstalk.DescribeApplicationVersions

Description

Retrieve a list of application versions.

Creating a Request

describeApplicationVersions :: DescribeApplicationVersions

Creates a value of DescribeApplicationVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dVersionLabels - Specify a version label to show a specific application version.

dNextToken - For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request. If no NextToken is specified, the first page is retrieved.

dMaxRecords - For a paginated request. Specify a maximum number of application versions to include in each response. If no MaxRecords is specified, all available application versions are retrieved in a single response.

dApplicationName - Specify an application name to show only application versions for that application.

data DescribeApplicationVersions

Request to describe application versions.

See: describeApplicationVersions smart constructor.

Request Lenses

dVersionLabels :: Lens' DescribeApplicationVersions [Text]

Specify a version label to show a specific application version.

dNextToken :: Lens' DescribeApplicationVersions (Maybe Text)

For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request. If no NextToken is specified, the first page is retrieved.

dMaxRecords :: Lens' DescribeApplicationVersions (Maybe Natural)

For a paginated request. Specify a maximum number of application versions to include in each response. If no MaxRecords is specified, all available application versions are retrieved in a single response.

dApplicationName :: Lens' DescribeApplicationVersions (Maybe Text)

Specify an application name to show only application versions for that application.

Destructuring the Response

describeApplicationVersionsResponse

Arguments

:: Int
davrsResponseStatus

-> DescribeApplicationVersionsResponse

Creates a value of DescribeApplicationVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

davrsApplicationVersions - List of ApplicationVersionDescription objects sorted in order of creation.

davrsNextToken - In a paginated request, the token that you can pass in a subsequent request to get the next response page.

davrsResponseStatus - -- | The response status code.

data DescribeApplicationVersionsResponse

Result message wrapping a list of application version descriptions.

See: describeApplicationVersionsResponse smart constructor.

Response Lenses

davrsApplicationVersions :: Lens' DescribeApplicationVersionsResponse [ApplicationVersionDescription]

List of ApplicationVersionDescription objects sorted in order of creation.

davrsNextToken :: Lens' DescribeApplicationVersionsResponse (Maybe Text)

In a paginated request, the token that you can pass in a subsequent request to get the next response page.

davrsResponseStatus :: Lens' DescribeApplicationVersionsResponse Int

- | The response status code.