Network.AWS.ElasticBeanstalk.DescribeConfigurationOptions

Description

Describes the configuration options that are used in a particular configuration template or environment, or that a specified solution stack defines. The description includes the values the options, their default values, and an indication of the required action on a running environment if an option value is changed.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_DescribeConfigurationOptions.html

Request

data DescribeConfigurationOptions

Request constructor

describeConfigurationOptions :: DescribeConfigurationOptions

DescribeConfigurationOptions constructor.

The fields accessible through corresponding lenses are:

dcoApplicationName :: Maybe Text

dcoEnvironmentName :: Maybe Text

dcoOptions :: [OptionSpecification]

dcoSolutionStackName :: Maybe Text

dcoTemplateName :: Maybe Text

Request lenses

dcoApplicationName :: Lens' DescribeConfigurationOptions (Maybe Text)

The name of the application associated with the configuration template or environment. Only needed if you want to describe the configuration options associated with either the configuration template or environment.

dcoEnvironmentName :: Lens' DescribeConfigurationOptions (Maybe Text)

The name of the environment whose configuration options you want to describe.

dcoOptions :: Lens' DescribeConfigurationOptions [OptionSpecification]

If specified, restricts the descriptions to only the specified options.

dcoSolutionStackName :: Lens' DescribeConfigurationOptions (Maybe Text)

The name of the solution stack whose configuration options you want to describe.

dcoTemplateName :: Lens' DescribeConfigurationOptions (Maybe Text)

The name of the configuration template whose configuration options you want to describe.

Response

data DescribeConfigurationOptionsResponse

Response constructor

describeConfigurationOptionsResponse :: DescribeConfigurationOptionsResponse

DescribeConfigurationOptionsResponse constructor.

The fields accessible through corresponding lenses are:

dcorOptions :: [ConfigurationOptionDescription]

dcorSolutionStackName :: Maybe Text

Response lenses

dcorOptions :: Lens' DescribeConfigurationOptionsResponse [ConfigurationOptionDescription]

A list of ConfigurationOptionDescription.

dcorSolutionStackName :: Lens' DescribeConfigurationOptionsResponse (Maybe Text)

The name of the solution stack these configuration options belong to.