Network.AWS.ElasticBeanstalk.DescribeConfigurationSettings

Description

Returns a description of the settings for the specified configuration set, that is, either a configuration template or the configuration set associated with a running environment.

When describing the settings for the configuration set associated with a running environment, it is possible to receive two sets of setting descriptions. One is the deployed configuration set, and the other is a draft configuration of an environment that is either in the process of deployment or that failed to deploy.

Related Topics

DeleteEnvironmentConfiguration

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_DescribeConfigurationSettings.html

Request

data DescribeConfigurationSettings

Request constructor

describeConfigurationSettings

Arguments

:: Text
dcsApplicationName

-> DescribeConfigurationSettings

DescribeConfigurationSettings constructor.

The fields accessible through corresponding lenses are:

dcsApplicationName :: Text

dcsEnvironmentName :: Maybe Text

dcsTemplateName :: Maybe Text

Request lenses

dcsApplicationName :: Lens' DescribeConfigurationSettings Text

The application for the environment or configuration template.

dcsEnvironmentName :: Lens' DescribeConfigurationSettings (Maybe Text)

The name of the environment to describe.

Condition: You must specify either this or a TemplateName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

dcsTemplateName :: Lens' DescribeConfigurationSettings (Maybe Text)

The name of the configuration template to describe.

Conditional: You must specify either this parameter or an EnvironmentName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns a MissingRequiredParameter error.

Response

data DescribeConfigurationSettingsResponse

Response constructor

describeConfigurationSettingsResponse :: DescribeConfigurationSettingsResponse

DescribeConfigurationSettingsResponse constructor.

The fields accessible through corresponding lenses are:

dcsrConfigurationSettings :: [ConfigurationSettingsDescription]

Response lenses

dcsrConfigurationSettings :: Lens' DescribeConfigurationSettingsResponse [ConfigurationSettingsDescription]

A list of ConfigurationSettingsDescription.