Network.AWS.ElasticBeanstalk.DescribeEnvironmentHealth

Description

Returns information about the overall health of the specified environment. The DescribeEnvironmentHealth operation is only available with AWS Elastic Beanstalk Enhanced Health.

Creating a Request

describeEnvironmentHealth :: DescribeEnvironmentHealth

Creates a value of DescribeEnvironmentHealth with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dehEnvironmentName - Specify the environment by name. You must specify either this or an EnvironmentName, or both.

dehAttributeNames - Specify the response elements to return. To retrieve all attributes, set to All . If no attribute names are specified, returns the name of the environment.

dehEnvironmentId - Specify the environment by ID. You must specify either this or an EnvironmentName, or both.

data DescribeEnvironmentHealth

See the example below to learn how to create a request body.

See: describeEnvironmentHealth smart constructor.

Request Lenses

dehEnvironmentName :: Lens' DescribeEnvironmentHealth (Maybe Text)

Specify the environment by name. You must specify either this or an EnvironmentName, or both.

dehAttributeNames :: Lens' DescribeEnvironmentHealth [EnvironmentHealthAttribute]

Specify the response elements to return. To retrieve all attributes, set to All . If no attribute names are specified, returns the name of the environment.

dehEnvironmentId :: Lens' DescribeEnvironmentHealth (Maybe Text)

Specify the environment by ID. You must specify either this or an EnvironmentName, or both.

Destructuring the Response

describeEnvironmentHealthResponse

Arguments

:: Int
dehrsResponseStatus

-> DescribeEnvironmentHealthResponse

Creates a value of DescribeEnvironmentHealthResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dehrsStatus - The environment's operational status. Ready , Launching , Updating , Terminating , or Terminated .

dehrsCauses - Descriptions of the data that contributed to the environment's current health status.

dehrsApplicationMetrics - Application request metrics for the environment.

dehrsColor - The health color of the environment.

dehrsEnvironmentName - The environment's name.

dehrsHealthStatus - The health status of the environment. For example, Ok .

dehrsInstancesHealth - Summary health information for the instances in the environment.

dehrsRefreshedAt - The date and time that the health information was retrieved.

dehrsResponseStatus - -- | The response status code.

data DescribeEnvironmentHealthResponse

Health details for an AWS Elastic Beanstalk environment.

See: describeEnvironmentHealthResponse smart constructor.

Response Lenses

dehrsStatus :: Lens' DescribeEnvironmentHealthResponse (Maybe EnvironmentHealth)

The environment's operational status. Ready , Launching , Updating , Terminating , or Terminated .

dehrsCauses :: Lens' DescribeEnvironmentHealthResponse [Text]

Descriptions of the data that contributed to the environment's current health status.

dehrsApplicationMetrics :: Lens' DescribeEnvironmentHealthResponse (Maybe ApplicationMetrics)

Application request metrics for the environment.

dehrsColor :: Lens' DescribeEnvironmentHealthResponse (Maybe Text)

The health color of the environment.

dehrsEnvironmentName :: Lens' DescribeEnvironmentHealthResponse (Maybe Text)

The environment's name.

dehrsHealthStatus :: Lens' DescribeEnvironmentHealthResponse (Maybe Text)

The health status of the environment. For example, Ok .

dehrsInstancesHealth :: Lens' DescribeEnvironmentHealthResponse (Maybe InstanceHealthSummary)

Summary health information for the instances in the environment.

dehrsRefreshedAt :: Lens' DescribeEnvironmentHealthResponse (Maybe UTCTime)

The date and time that the health information was retrieved.

dehrsResponseStatus :: Lens' DescribeEnvironmentHealthResponse Int

- | The response status code.