Network.AWS.ElasticBeanstalk.DescribeEnvironmentManagedActions

Description

Lists an environment's upcoming and in-progress managed actions.

Creating a Request

describeEnvironmentManagedActions :: DescribeEnvironmentManagedActions

Creates a value of DescribeEnvironmentManagedActions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

demaStatus - To show only actions with a particular status, specify a status.

demaEnvironmentName - The name of the target environment.

demaEnvironmentId - The environment ID of the target environment.

data DescribeEnvironmentManagedActions

Request to list an environment's upcoming and in-progress managed actions.

See: describeEnvironmentManagedActions smart constructor.

Request Lenses

demaStatus :: Lens' DescribeEnvironmentManagedActions (Maybe ActionStatus)

To show only actions with a particular status, specify a status.

demaEnvironmentName :: Lens' DescribeEnvironmentManagedActions (Maybe Text)

The name of the target environment.

demaEnvironmentId :: Lens' DescribeEnvironmentManagedActions (Maybe Text)

The environment ID of the target environment.

Destructuring the Response

describeEnvironmentManagedActionsResponse

Arguments

:: Int
demarsResponseStatus

-> DescribeEnvironmentManagedActionsResponse

Creates a value of DescribeEnvironmentManagedActionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

demarsManagedActions - A list of upcoming and in-progress managed actions.

demarsResponseStatus - -- | The response status code.

data DescribeEnvironmentManagedActionsResponse

The result message containing a list of managed actions.

See: describeEnvironmentManagedActionsResponse smart constructor.

Response Lenses

demarsManagedActions :: Lens' DescribeEnvironmentManagedActionsResponse (Maybe (NonEmpty ManagedAction))

A list of upcoming and in-progress managed actions.

demarsResponseStatus :: Lens' DescribeEnvironmentManagedActionsResponse Int

- | The response status code.