Network.AWS.ElasticBeanstalk.DescribeEnvironmentResources

Description

Returns AWS resources for this environment. https:/elasticbeanstalk.us-east-1.amazon.com?EnvironmentId=e-hc8mvnayrx &EnvironmentName=SampleAppVersion &Operation=DescribeEnvironmentResources &AuthParams elasticbeanstalk-SampleAppVersion elasticbeanstalk-SampleAppVersion-hbAc8cSZH7 elasticbeanstalk-SampleAppVersion-us-east-1c SampleAppVersion elasticbeanstalk-SampleAppVersion-us-east-1c e1cb7b96-f287-11df-8a78-9f77047e0d0c.

Request

data DescribeEnvironmentResources

This documentation target is not reported in the API reference.

Request constructor

describeEnvironmentResources :: DescribeEnvironmentResources

Smart constructor for the minimum required parameters to construct a valid DescribeEnvironmentResources request.

The fields accessible through corresponding lenses are:

EnvironmentId :: Maybe Text

EnvironmentName :: Maybe Text

Request lenses

derEnvironmentId :: Lens' DescribeEnvironmentResources (Maybe Text)

The ID of the environment to retrieve AWS resource usage data. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

derEnvironmentName :: Lens' DescribeEnvironmentResources (Maybe Text)

The name of the environment to retrieve AWS resource usage data. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

Response

data DescribeEnvironmentResourcesResponse

Result message containing a list of environment resource descriptions.

Response constructor

describeEnvironmentResourcesResponse :: DescribeEnvironmentResourcesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeEnvironmentResourcesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

EnvironmentResources :: Maybe EnvironmentResourceDescription

Response lenses

derrEnvironmentResources :: Lens' DescribeEnvironmentResourcesResponse (Maybe EnvironmentResourceDescription)

A list of EnvironmentResourceDescription.