Network.AWS.ElasticBeanstalk.DescribeEnvironments

Description

Returns descriptions for existing environments.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_DescribeEnvironments.html

Request

data DescribeEnvironments

Request constructor

describeEnvironments :: DescribeEnvironments

DescribeEnvironments constructor.

The fields accessible through corresponding lenses are:

de1ApplicationName :: Maybe Text

de1EnvironmentIds :: [Text]

de1EnvironmentNames :: [Text]

de1IncludeDeleted :: Maybe Bool

de1IncludedDeletedBackTo :: Maybe UTCTime

de1VersionLabel :: Maybe Text

Request lenses

de1ApplicationName :: Lens' DescribeEnvironments (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application.

de1EnvironmentIds :: Lens' DescribeEnvironments [Text]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified IDs.

de1EnvironmentNames :: Lens' DescribeEnvironments [Text]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified names.

de1IncludeDeleted :: Lens' DescribeEnvironments (Maybe Bool)

Indicates whether to include deleted environments:

true: Environments that have been deleted after IncludedDeletedBackTo are displayed.

false: Do not include deleted environments.

de1IncludedDeletedBackTo :: Lens' DescribeEnvironments (Maybe UTCTime)

If specified when IncludeDeleted is set to true, then environments deleted after this date are displayed.

de1VersionLabel :: Lens' DescribeEnvironments (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application version.

Response

data DescribeEnvironmentsResponse

Response constructor

describeEnvironmentsResponse :: DescribeEnvironmentsResponse

DescribeEnvironmentsResponse constructor.

The fields accessible through corresponding lenses are:

derEnvironments :: [EnvironmentDescription]

Response lenses

derEnvironments :: Lens' DescribeEnvironmentsResponse [EnvironmentDescription]

Returns an EnvironmentDescription list.