Network.AWS.ElasticBeanstalk.DescribeEvents

Description

Returns list of event descriptions matching criteria up to the last 6 weeks. This action returns the most recent 1,000 events from the specified NextToken. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &Severity=TRACE &StartTime=2010-11-17T10%3A26%3A40Z &Operation=DescribeEvents &AuthParams Successfully completed createEnvironment activity. 2010-11-17T20:25:35.191Z New Version bb01fa74-f287-11df-8a78-9f77047e0d0c SampleApp SampleAppVersion INFO Launching a new EC2 instance: i-04a8c569 2010-11-17T20:21:30Z New Version SampleApp SampleAppVersion DEBUG At least one EC2 instance has entered the InService lifecycle state. 2010-11-17T20:20:32.008Z New Version bb01fa74-f287-11df-8a78-9f77047e0d0c SampleApp SampleAppVersion INFO Elastic Load Balancer elasticbeanstalk-SampleAppVersion has failed 0 healthy instances - Environment may not be available. 2010-11-17T20:19:28Z New Version SampleApp SampleAppVersion WARN f10d02dd-f288-11df-8a78-9f77047e0d0c.

Request

data DescribeEvents

This documentation target is not reported in the API reference.

Request constructor

mkDescribeEvents :: DescribeEvents

Smart constructor for the minimum required parameters to construct a valid DescribeEvents request.

The fields accessible through corresponding lenses are:

ApplicationName :: Maybe Text

VersionLabel :: Maybe Text

TemplateName :: Maybe Text

EnvironmentId :: Maybe Text

EnvironmentName :: Maybe Text

RequestId :: Maybe Text

Severity :: Maybe EventSeverity

StartTime :: Maybe ISO8601

EndTime :: Maybe ISO8601

MaxRecords :: Maybe Integer

NextToken :: Maybe Text

Request lenses

de1ApplicationName :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those associated with this application.

de1VersionLabel :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this application version.

de1TemplateName :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that are associated with this environment configuration.

de1EnvironmentId :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

de1EnvironmentName :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

de1RequestId :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the described events to include only those associated with this request ID.

de1Severity :: Lens' DescribeEvents (Maybe EventSeverity)

If specified, limits the events returned from this call to include only those with the specified severity or higher.

de1StartTime :: Lens' DescribeEvents (Maybe ISO8601)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur on or after this time.

de1EndTime :: Lens' DescribeEvents (Maybe ISO8601)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur up to, but not including, the EndTime.

de1MaxRecords :: Lens' DescribeEvents (Maybe Integer)

Specifies the maximum number of events that can be returned, beginning with the most recent event.

de1NextToken :: Lens' DescribeEvents (Maybe Text)

Pagination token. If specified, the events return the next batch of results.

Response

data DescribeEventsResponse

Result message wrapping a list of event descriptions.

Response constructor

mkDescribeEventsResponse :: DescribeEventsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeEventsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Events :: [EventDescription]

NextToken :: Maybe Text

Response lenses

der1rEvents :: Lens' DescribeEventsResponse [EventDescription]

A list of EventDescription.

der1rNextToken :: Lens' DescribeEventsResponse (Maybe Text)

If returned, this indicates that there are more results to obtain. Use this token in the next DescribeEvents call to get the next batch of events.