Network.AWS.ElasticBeanstalk.DescribeInstancesHealth

Description

Retrives detailed information about the health of instances in your AWS Elastic Beanstalk. This operation requires enhanced health reporting .

Creating a Request

describeInstancesHealth :: DescribeInstancesHealth

Creates a value of DescribeInstancesHealth with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dihNextToken - Specify the pagination token returned by a previous call.

dihEnvironmentName - Specify the AWS Elastic Beanstalk environment by name.

dihAttributeNames - Specifies the response elements you wish to receive. To retrieve all attributes, set to All . If no attribute names are specified, returns a list of instances.

dihEnvironmentId - Specify the AWS Elastic Beanstalk environment by ID.

data DescribeInstancesHealth

Parameters for a call to DescribeInstancesHealth .

See: describeInstancesHealth smart constructor.

Request Lenses

dihNextToken :: Lens' DescribeInstancesHealth (Maybe Text)

Specify the pagination token returned by a previous call.

dihEnvironmentName :: Lens' DescribeInstancesHealth (Maybe Text)

Specify the AWS Elastic Beanstalk environment by name.

dihAttributeNames :: Lens' DescribeInstancesHealth [InstancesHealthAttribute]

Specifies the response elements you wish to receive. To retrieve all attributes, set to All . If no attribute names are specified, returns a list of instances.

dihEnvironmentId :: Lens' DescribeInstancesHealth (Maybe Text)

Specify the AWS Elastic Beanstalk environment by ID.

Destructuring the Response

describeInstancesHealthResponse

Arguments

:: Int
dihrsResponseStatus

-> DescribeInstancesHealthResponse

Creates a value of DescribeInstancesHealthResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dihrsInstanceHealthList - Detailed health information about each instance.

dihrsNextToken - Pagination token for the next page of results, if available.

dihrsRefreshedAt - The date and time that the health information was retrieved.

dihrsResponseStatus - -- | The response status code.

data DescribeInstancesHealthResponse

Detailed health information about the Amazon EC2 instances in an AWS Elastic Beanstalk environment.

See: describeInstancesHealthResponse smart constructor.

Response Lenses

dihrsInstanceHealthList :: Lens' DescribeInstancesHealthResponse [SingleInstanceHealth]

Detailed health information about each instance.

dihrsNextToken :: Lens' DescribeInstancesHealthResponse (Maybe Text)

Pagination token for the next page of results, if available.

dihrsRefreshedAt :: Lens' DescribeInstancesHealthResponse (Maybe UTCTime)

The date and time that the health information was retrieved.

dihrsResponseStatus :: Lens' DescribeInstancesHealthResponse Int

- | The response status code.