Network.AWS.ElasticBeanstalk.ListAvailableSolutionStacks

Description

Returns a list of the available solution stack names.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_ListAvailableSolutionStacks.html

Request

data ListAvailableSolutionStacks

Request constructor

listAvailableSolutionStacks :: ListAvailableSolutionStacks

ListAvailableSolutionStacks constructor.

Response

data ListAvailableSolutionStacksResponse

Response constructor

listAvailableSolutionStacksResponse :: ListAvailableSolutionStacksResponse

ListAvailableSolutionStacksResponse constructor.

The fields accessible through corresponding lenses are:

lassrSolutionStackDetails :: [SolutionStackDescription]

lassrSolutionStacks :: [Text]

Response lenses

lassrSolutionStackDetails :: Lens' ListAvailableSolutionStacksResponse [SolutionStackDescription]

A list of available solution stacks and their SolutionStackDescription.

lassrSolutionStacks :: Lens' ListAvailableSolutionStacksResponse [Text]

A list of available solution stacks.