Network.AWS.ElasticBeanstalk.RebuildEnvironment

Description

Deletes and recreates all of the AWS resources (for example: the Auto Scaling group, load balancer, etc.) for a specified environment and forces a restart. https:/elasticbeanstalk.us-east-1.amazon.com?EnvironmentId=e-hc8mvnayrx &EnvironmentName=SampleAppVersion &Operation=RebuildEnvironment &AuthParams a7d6606e-f289-11df-8a78-9f77047e0d0c.

Request

data RebuildEnvironment

Request constructor

mkRebuildEnvironment :: RebuildEnvironment

Smart constructor for the minimum required parameters to construct a valid RebuildEnvironment request.

The fields accessible through corresponding lenses are:

EnvironmentId :: Maybe Text

EnvironmentName :: Maybe Text

Request lenses

reEnvironmentId :: Lens' RebuildEnvironment (Maybe Text)

The ID of the environment to rebuild. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

reEnvironmentName :: Lens' RebuildEnvironment (Maybe Text)

The name of the environment to rebuild. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

Response

data RebuildEnvironmentResponse

Response constructor

mkRebuildEnvironmentResponse :: RebuildEnvironmentResponse

Smart constructor for the minimum required parameters to construct a valid RebuildEnvironmentResponse response.

This constructor is provided for convenience and testing purposes.