Network.AWS.ElasticBeanstalk.RequestEnvironmentInfo

Description

Initiates a request to compile the specified type of information of the deployed environment.

Setting the InfoType to tail compiles the last lines from the application server log files of every Amazon EC2 instance in your environment.

Setting the InfoType to bundle compresses the application server log files for every Amazon EC2 instance into a '.zip' file. Legacy and .NET containers do not support bundle logs.

Use RetrieveEnvironmentInfo to obtain the set of logs.

Related Topics

RetrieveEnvironmentInfo

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_RequestEnvironmentInfo.html

Request

data RequestEnvironmentInfo

Request constructor

requestEnvironmentInfo

Arguments

:: EnvironmentInfoType
reiInfoType

-> RequestEnvironmentInfo

RequestEnvironmentInfo constructor.

The fields accessible through corresponding lenses are:

reiEnvironmentId :: Maybe Text

reiEnvironmentName :: Maybe Text

reiInfoType :: EnvironmentInfoType

Request lenses

reiEnvironmentId :: Lens' RequestEnvironmentInfo (Maybe Text)

The ID of the environment of the requested data.

If no such environment is found, RequestEnvironmentInfo returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

reiEnvironmentName :: Lens' RequestEnvironmentInfo (Maybe Text)

The name of the environment of the requested data.

If no such environment is found, RequestEnvironmentInfo returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

reiInfoType :: Lens' RequestEnvironmentInfo EnvironmentInfoType

The type of information to request.

Response

data RequestEnvironmentInfoResponse

Response constructor

requestEnvironmentInfoResponse :: RequestEnvironmentInfoResponse

RequestEnvironmentInfoResponse constructor.