Network.AWS.ElasticBeanstalk.RestartAppServer

Description

Causes the environment to restart the application container server running on each Amazon EC2 instance. https:/elasticbeanstalk.us-east-1.amazon.com?EnvironmentId=e-hc8mvnayrx &EnvironmentName=SampleAppVersion &Operation=RestartAppServer &AuthParams 90e8d1d5-f28a-11df-8a78-9f77047e0d0c.

Request

data RestartAppServer

Request constructor

restartAppServer :: RestartAppServer

Smart constructor for the minimum required parameters to construct a valid RestartAppServer request.

The fields accessible through corresponding lenses are:

EnvironmentId :: Maybe Text

EnvironmentName :: Maybe Text

Request lenses

rasEnvironmentId :: Lens' RestartAppServer (Maybe Text)

The ID of the environment to restart the server for. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

rasEnvironmentName :: Lens' RestartAppServer (Maybe Text)

The name of the environment to restart the server for. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

Response

data RestartAppServerResponse

Response constructor

restartAppServerResponse :: RestartAppServerResponse

Smart constructor for the minimum required parameters to construct a valid RestartAppServerResponse response.

This constructor is provided for convenience and testing purposes.