Network.AWS.ElasticBeanstalk.RetrieveEnvironmentInfo

Description

Retrieves the compiled information from a RequestEnvironmentInfo request.

Related Topics

RequestEnvironmentInfo

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_RetrieveEnvironmentInfo.html

Request

data RetrieveEnvironmentInfo

Request constructor

retrieveEnvironmentInfo

Arguments

:: EnvironmentInfoType
rei1InfoType

-> RetrieveEnvironmentInfo

RetrieveEnvironmentInfo constructor.

The fields accessible through corresponding lenses are:

rei1EnvironmentId :: Maybe Text

rei1EnvironmentName :: Maybe Text

rei1InfoType :: EnvironmentInfoType

Request lenses

rei1EnvironmentId :: Lens' RetrieveEnvironmentInfo (Maybe Text)

The ID of the data's environment.

If no such environment is found, returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

rei1EnvironmentName :: Lens' RetrieveEnvironmentInfo (Maybe Text)

The name of the data's environment.

If no such environment is found, returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

rei1InfoType :: Lens' RetrieveEnvironmentInfo EnvironmentInfoType

The type of information to retrieve.

Response

data RetrieveEnvironmentInfoResponse

Response constructor

retrieveEnvironmentInfoResponse :: RetrieveEnvironmentInfoResponse

RetrieveEnvironmentInfoResponse constructor.

The fields accessible through corresponding lenses are:

reirEnvironmentInfo :: [EnvironmentInfoDescription]

Response lenses

reirEnvironmentInfo :: Lens' RetrieveEnvironmentInfoResponse [EnvironmentInfoDescription]

The EnvironmentInfoDescription of the environment.