Network.AWS.ElasticBeanstalk.SwapEnvironmentCNAMEs

Description

Swaps the CNAMEs of two environments.

Creating a Request

swapEnvironmentCNAMEs :: SwapEnvironmentCNAMEs

Creates a value of SwapEnvironmentCNAMEs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

secnameDestinationEnvironmentName - The name of the destination environment. Condition: You must specify at least the DestinationEnvironmentID or the DestinationEnvironmentName . You may also specify both. You must specify the SourceEnvironmentName with the DestinationEnvironmentName .

secnameDestinationEnvironmentId - The ID of the destination environment. Condition: You must specify at least the DestinationEnvironmentID or the DestinationEnvironmentName . You may also specify both. You must specify the SourceEnvironmentId with the DestinationEnvironmentId .

secnameSourceEnvironmentName - The name of the source environment. Condition: You must specify at least the SourceEnvironmentID or the SourceEnvironmentName . You may also specify both. If you specify the SourceEnvironmentName , you must specify the DestinationEnvironmentName .

secnameSourceEnvironmentId - The ID of the source environment. Condition: You must specify at least the SourceEnvironmentID or the SourceEnvironmentName . You may also specify both. If you specify the SourceEnvironmentId , you must specify the DestinationEnvironmentId .

data SwapEnvironmentCNAMEs

Swaps the CNAMEs of two environments.

See: swapEnvironmentCNAMEs smart constructor.

Request Lenses

secnameDestinationEnvironmentName :: Lens' SwapEnvironmentCNAMEs (Maybe Text)

The name of the destination environment. Condition: You must specify at least the DestinationEnvironmentID or the DestinationEnvironmentName . You may also specify both. You must specify the SourceEnvironmentName with the DestinationEnvironmentName .

secnameDestinationEnvironmentId :: Lens' SwapEnvironmentCNAMEs (Maybe Text)

The ID of the destination environment. Condition: You must specify at least the DestinationEnvironmentID or the DestinationEnvironmentName . You may also specify both. You must specify the SourceEnvironmentId with the DestinationEnvironmentId .

secnameSourceEnvironmentName :: Lens' SwapEnvironmentCNAMEs (Maybe Text)

The name of the source environment. Condition: You must specify at least the SourceEnvironmentID or the SourceEnvironmentName . You may also specify both. If you specify the SourceEnvironmentName , you must specify the DestinationEnvironmentName .

secnameSourceEnvironmentId :: Lens' SwapEnvironmentCNAMEs (Maybe Text)

The ID of the source environment. Condition: You must specify at least the SourceEnvironmentID or the SourceEnvironmentName . You may also specify both. If you specify the SourceEnvironmentId , you must specify the DestinationEnvironmentId .

Destructuring the Response

swapEnvironmentCNAMEsResponse :: SwapEnvironmentCNAMEsResponse

Creates a value of SwapEnvironmentCNAMEsResponse with the minimum fields required to make a request.

data SwapEnvironmentCNAMEsResponse

See: swapEnvironmentCNAMEsResponse smart constructor.