Network.AWS.ElasticBeanstalk.TerminateEnvironment

Description

Terminates the specified environment.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_TerminateEnvironment.html

Request

data TerminateEnvironment

Request constructor

terminateEnvironment :: TerminateEnvironment

TerminateEnvironment constructor.

The fields accessible through corresponding lenses are:

teEnvironmentId :: Maybe Text

teEnvironmentName :: Maybe Text

teTerminateResources :: Maybe Bool

Request lenses

teEnvironmentId :: Lens' TerminateEnvironment (Maybe Text)

The ID of the environment to terminate.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

teEnvironmentName :: Lens' TerminateEnvironment (Maybe Text)

The name of the environment to terminate.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

teTerminateResources :: Lens' TerminateEnvironment (Maybe Bool)

Indicates whether the associated AWS resources should shut down when the environment is terminated:

true: (default) The user AWS resources (for example, the Auto Scaling group, LoadBalancer, etc.) are terminated along with the environment.

false: The environment is removed from the AWS Elastic Beanstalk but the AWS resources continue to operate.

true: The specified environment as well as the associated AWS resources, such as Auto Scaling group and LoadBalancer, are terminated. false: AWS Elastic Beanstalk resource management is removed from the environment, but the AWS resources continue to operate. For more information, see the AWSElastic Beanstalk User Guide.

Default: true

Valid Values: true | false

Response

data TerminateEnvironmentResponse

Response constructor

terminateEnvironmentResponse :: TerminateEnvironmentResponse

TerminateEnvironmentResponse constructor.

The fields accessible through corresponding lenses are:

terAbortableOperationInProgress :: Maybe Bool

terApplicationName :: Maybe Text

terCNAME :: Maybe Text

terDateCreated :: Maybe UTCTime

terDateUpdated :: Maybe UTCTime

terDescription :: Maybe Text

terEndpointURL :: Maybe Text

terEnvironmentId :: Maybe Text

terEnvironmentName :: Maybe Text

terHealth :: Maybe EnvironmentHealth

terResources :: Maybe EnvironmentResourcesDescription

terSolutionStackName :: Maybe Text

terStatus :: Maybe EnvironmentStatus

terTemplateName :: Maybe Text

terTier :: Maybe EnvironmentTier

terVersionLabel :: Maybe Text

Response lenses

terAbortableOperationInProgress :: Lens' TerminateEnvironmentResponse (Maybe Bool)

Lists in-progress environment updates and application version deployments that you can cancel.

terApplicationName :: Lens' TerminateEnvironmentResponse (Maybe Text)

The name of the application associated with this environment.

terCNAME :: Lens' TerminateEnvironmentResponse (Maybe Text)

The URL to the CNAME for this environment.

terDateCreated :: Lens' TerminateEnvironmentResponse (Maybe UTCTime)

The creation date for this environment.

terDateUpdated :: Lens' TerminateEnvironmentResponse (Maybe UTCTime)

The last modified date for this environment.

terDescription :: Lens' TerminateEnvironmentResponse (Maybe Text)

Describes this environment.

terEndpointURL :: Lens' TerminateEnvironmentResponse (Maybe Text)

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

terEnvironmentId :: Lens' TerminateEnvironmentResponse (Maybe Text)

The ID of this environment.

terEnvironmentName :: Lens' TerminateEnvironmentResponse (Maybe Text)

The name of this environment.

terHealth :: Lens' TerminateEnvironmentResponse (Maybe EnvironmentHealth)

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment:

Red : Indicates the environment is not working.

Yellow: Indicates that something is wrong, the application might not be available, but the instances appear running.

Green: Indicates the environment is healthy and fully functional.

Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. Green: Indicates the environment is healthy and fully functional. Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey

terResources :: Lens' TerminateEnvironmentResponse (Maybe EnvironmentResourcesDescription)

The description of the AWS resources used by this environment.

terSolutionStackName :: Lens' TerminateEnvironmentResponse (Maybe Text)

The name of the SolutionStack deployed with this environment.

terStatus :: Lens' TerminateEnvironmentResponse (Maybe EnvironmentStatus)

The current operational status of the environment:

Launching: Environment is in the process of initial deployment. Updating: Environment is in the process of updating its configuration settings or application version. Ready: Environment is available to have an action performed on it, such as update or terminate. Terminating: Environment is in the shut-down process. Terminated: Environment is not running.

terTemplateName :: Lens' TerminateEnvironmentResponse (Maybe Text)

The name of the configuration template used to originally launch this environment.

terTier :: Lens' TerminateEnvironmentResponse (Maybe EnvironmentTier)

Describes the current tier of this environment.

terVersionLabel :: Lens' TerminateEnvironmentResponse (Maybe Text)

The application version deployed in this environment.