Network.AWS.ElasticBeanstalk.Types

Service

data ElasticBeanstalk

Version 2010-12-01 of the Amazon Elastic Beanstalk service.

Error

data RESTError :: *

XML

ns :: Text

ApplicationDescription

data ApplicationDescription

applicationDescription :: ApplicationDescription

ApplicationDescription constructor.

The fields accessible through corresponding lenses are:

adApplicationName :: Maybe Text

adConfigurationTemplates :: [Text]

adDateCreated :: Maybe UTCTime

adDateUpdated :: Maybe UTCTime

adDescription :: Maybe Text

adVersions :: [Text]

adApplicationName :: Lens' ApplicationDescription (Maybe Text)

The name of the application.

adConfigurationTemplates :: Lens' ApplicationDescription [Text]

The names of the configuration templates associated with this application.

adDateCreated :: Lens' ApplicationDescription (Maybe UTCTime)

The date when the application was created.

adDateUpdated :: Lens' ApplicationDescription (Maybe UTCTime)

The date when the application was last modified.

adDescription :: Lens' ApplicationDescription (Maybe Text)

User-defined description of the application.

adVersions :: Lens' ApplicationDescription [Text]

The names of the versions for this application.

EventSeverity

data EventSeverity

Constructors

Debug
DEBUG

Error
ERROR

Fatal
FATAL

Info
INFO

Trace
TRACE

Warn
WARN

Tag

data Tag

tag :: Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Maybe Text

tagValue :: Maybe Text

tagKey :: Lens' Tag (Maybe Text)

The key of the tag.

tagValue :: Lens' Tag (Maybe Text)

The value of the tag.

EventDescription

data EventDescription

eventDescription :: EventDescription

EventDescription constructor.

The fields accessible through corresponding lenses are:

edApplicationName :: Maybe Text

edEnvironmentName :: Maybe Text

edEventDate :: Maybe UTCTime

edMessage :: Maybe Text

edRequestId :: Maybe Text

edSeverity :: Maybe EventSeverity

edTemplateName :: Maybe Text

edVersionLabel :: Maybe Text

edApplicationName :: Lens' EventDescription (Maybe Text)

The application associated with the event.

edEnvironmentName :: Lens' EventDescription (Maybe Text)

The name of the environment associated with this event.

edEventDate :: Lens' EventDescription (Maybe UTCTime)

The date when the event occurred.

edMessage :: Lens' EventDescription (Maybe Text)

The event message.

edRequestId :: Lens' EventDescription (Maybe Text)

The web service request ID for the activity of this event.

edSeverity :: Lens' EventDescription (Maybe EventSeverity)

The severity level of this event.

edTemplateName :: Lens' EventDescription (Maybe Text)

The name of the configuration associated with this event.

edVersionLabel :: Lens' EventDescription (Maybe Text)

The release label for the application version associated with this event.

LaunchConfiguration

data LaunchConfiguration

launchConfiguration :: LaunchConfiguration

LaunchConfiguration constructor.

The fields accessible through corresponding lenses are:

lcName :: Maybe Text

lcName :: Lens' LaunchConfiguration (Maybe Text)

The name of the launch configuration.

ApplicationVersionDescriptionMessage

data ApplicationVersionDescriptionMessage

applicationVersionDescriptionMessage :: ApplicationVersionDescriptionMessage

ApplicationVersionDescriptionMessage constructor.

The fields accessible through corresponding lenses are:

avdmApplicationVersion :: Maybe ApplicationVersionDescription

avdmApplicationVersion :: Lens' ApplicationVersionDescriptionMessage (Maybe ApplicationVersionDescription)

The ApplicationVersionDescription of the application version.

AutoScalingGroup

data AutoScalingGroup

autoScalingGroup :: AutoScalingGroup

AutoScalingGroup constructor.

The fields accessible through corresponding lenses are:

asgName :: Maybe Text

asgName :: Lens' AutoScalingGroup (Maybe Text)

The name of the AutoScalingGroup .

ConfigurationDeploymentStatus

data ConfigurationDeploymentStatus

Constructors

Deployed
deployed

Failed
failed

Pending
pending

ConfigurationOptionSetting

data ConfigurationOptionSetting

configurationOptionSetting :: ConfigurationOptionSetting

ConfigurationOptionSetting constructor.

The fields accessible through corresponding lenses are:

cosNamespace :: Maybe Text

cosOptionName :: Maybe Text

cosValue :: Maybe Text

cosNamespace :: Lens' ConfigurationOptionSetting (Maybe Text)

A unique namespace identifying the option's associated AWS resource.

cosOptionName :: Lens' ConfigurationOptionSetting (Maybe Text)

The name of the configuration option.

cosValue :: Lens' ConfigurationOptionSetting (Maybe Text)

The current value for the configuration option.

ConfigurationOptionValueType

data ConfigurationOptionValueType

Constructors

List'
List

Scalar
Scalar

ConfigurationSettingsDescription

data ConfigurationSettingsDescription

configurationSettingsDescription :: ConfigurationSettingsDescription

ConfigurationSettingsDescription constructor.

The fields accessible through corresponding lenses are:

csdApplicationName :: Maybe Text

csdDateCreated :: Maybe UTCTime

csdDateUpdated :: Maybe UTCTime

csdDeploymentStatus :: Maybe ConfigurationDeploymentStatus

csdDescription :: Maybe Text

csdEnvironmentName :: Maybe Text

csdOptionSettings :: [ConfigurationOptionSetting]

csdSolutionStackName :: Maybe Text

csdTemplateName :: Maybe Text

csdApplicationName :: Lens' ConfigurationSettingsDescription (Maybe Text)

The name of the application associated with this configuration set.

csdDateCreated :: Lens' ConfigurationSettingsDescription (Maybe UTCTime)

The date (in UTC time) when this configuration set was created.

csdDateUpdated :: Lens' ConfigurationSettingsDescription (Maybe UTCTime)

The date (in UTC time) when this configuration set was last modified.

csdDeploymentStatus :: Lens' ConfigurationSettingsDescription (Maybe ConfigurationDeploymentStatus)

If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set:

null: This configuration is not associated with a running environment.

pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying.

deployed: This is the configuration that is currently deployed to the associated running environment.

failed: This is a draft configuration, that failed to successfully deploy.

null: This configuration is not associated with a running environment. pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying. deployed: This is the configuration that is currently deployed to the associated running environment. failed: This is a draft configuration that failed to successfully deploy.

csdDescription :: Lens' ConfigurationSettingsDescription (Maybe Text)

Describes this configuration set.

csdEnvironmentName :: Lens' ConfigurationSettingsDescription (Maybe Text)

If not null, the name of the environment for this configuration set.

csdOptionSettings :: Lens' ConfigurationSettingsDescription [ConfigurationOptionSetting]

A list of the configuration options and their values in this configuration set.

csdSolutionStackName :: Lens' ConfigurationSettingsDescription (Maybe Text)

The name of the solution stack this configuration set uses.

csdTemplateName :: Lens' ConfigurationSettingsDescription (Maybe Text)

If not null, the name of the configuration template for this configuration set.

ApplicationVersionDescription

data ApplicationVersionDescription

applicationVersionDescription :: ApplicationVersionDescription

ApplicationVersionDescription constructor.

The fields accessible through corresponding lenses are:

avdApplicationName :: Maybe Text

avdDateCreated :: Maybe UTCTime

avdDateUpdated :: Maybe UTCTime

avdDescription :: Maybe Text

avdSourceBundle :: Maybe S3Location

avdVersionLabel :: Maybe Text

avdApplicationName :: Lens' ApplicationVersionDescription (Maybe Text)

The name of the application associated with this release.

avdDateCreated :: Lens' ApplicationVersionDescription (Maybe UTCTime)

The creation date of the application version.

avdDateUpdated :: Lens' ApplicationVersionDescription (Maybe UTCTime)

The last modified date of the application version.

avdDescription :: Lens' ApplicationVersionDescription (Maybe Text)

The description of this application version.

avdSourceBundle :: Lens' ApplicationVersionDescription (Maybe S3Location)

The location where the source bundle is located for this version.

avdVersionLabel :: Lens' ApplicationVersionDescription (Maybe Text)

A label uniquely identifying the version for the associated application.

OptionSpecification

data OptionSpecification

optionSpecification :: OptionSpecification

OptionSpecification constructor.

The fields accessible through corresponding lenses are:

osNamespace :: Maybe Text

osOptionName :: Maybe Text

osNamespace :: Lens' OptionSpecification (Maybe Text)

A unique namespace identifying the option's associated AWS resource.

osOptionName :: Lens' OptionSpecification (Maybe Text)

The name of the configuration option.

EnvironmentResourceDescription

data EnvironmentResourceDescription

environmentResourceDescription :: EnvironmentResourceDescription

EnvironmentResourceDescription constructor.

The fields accessible through corresponding lenses are:

erdAutoScalingGroups :: [AutoScalingGroup]

erdEnvironmentName :: Maybe Text

erdInstances :: [Instance]

erdLaunchConfigurations :: [LaunchConfiguration]

erdLoadBalancers :: [LoadBalancer]

erdQueues :: [Queue]

erdTriggers :: [Trigger]

erdAutoScalingGroups :: Lens' EnvironmentResourceDescription [AutoScalingGroup]

The AutoScalingGroups used by this environment.

erdEnvironmentName :: Lens' EnvironmentResourceDescription (Maybe Text)

The name of the environment.

erdInstances :: Lens' EnvironmentResourceDescription [Instance]

The Amazon EC2 instances used by this environment.

erdLaunchConfigurations :: Lens' EnvironmentResourceDescription [LaunchConfiguration]

The Auto Scaling launch configurations in use by this environment.

erdLoadBalancers :: Lens' EnvironmentResourceDescription [LoadBalancer]

The LoadBalancers in use by this environment.

erdQueues :: Lens' EnvironmentResourceDescription [Queue]

The queues used by this environment.

erdTriggers :: Lens' EnvironmentResourceDescription [Trigger]

The AutoScaling triggers in use by this environment.

Queue

data Queue

queue :: Queue

Queue constructor.

The fields accessible through corresponding lenses are:

qName :: Maybe Text

qURL :: Maybe Text

qName :: Lens' Queue (Maybe Text)

The name of the queue.

qURL :: Lens' Queue (Maybe Text)

The URL of the queue.

EnvironmentStatus

data EnvironmentStatus

Constructors

Launching
Launching

Ready
Ready

Terminated
Terminated

Terminating
Terminating

Updating
Updating

LoadBalancerDescription

data LoadBalancerDescription

loadBalancerDescription :: LoadBalancerDescription

LoadBalancerDescription constructor.

The fields accessible through corresponding lenses are:

lbdDomain :: Maybe Text

lbdListeners :: [Listener]

lbdLoadBalancerName :: Maybe Text

lbdDomain :: Lens' LoadBalancerDescription (Maybe Text)

The domain name of the LoadBalancer.

lbdListeners :: Lens' LoadBalancerDescription [Listener]

A list of Listeners used by the LoadBalancer.

lbdLoadBalancerName :: Lens' LoadBalancerDescription (Maybe Text)

The name of the LoadBalancer.

ApplicationDescriptionMessage

data ApplicationDescriptionMessage

applicationDescriptionMessage :: ApplicationDescriptionMessage

ApplicationDescriptionMessage constructor.

The fields accessible through corresponding lenses are:

admApplication :: Maybe ApplicationDescription

admApplication :: Lens' ApplicationDescriptionMessage (Maybe ApplicationDescription)

The ApplicationDescription of the application.

EnvironmentTier

data EnvironmentTier

environmentTier :: EnvironmentTier

EnvironmentTier constructor.

The fields accessible through corresponding lenses are:

etName :: Maybe Text

etType :: Maybe Text

etVersion :: Maybe Text

etName :: Lens' EnvironmentTier (Maybe Text)

The name of this environment tier.

etType :: Lens' EnvironmentTier (Maybe Text)

The type of this environment tier.

etVersion :: Lens' EnvironmentTier (Maybe Text)

The version of this environment tier.

LoadBalancer

data LoadBalancer

loadBalancer :: LoadBalancer

LoadBalancer constructor.

The fields accessible through corresponding lenses are:

lbName :: Maybe Text

lbName :: Lens' LoadBalancer (Maybe Text)

The name of the LoadBalancer.

EnvironmentResourcesDescription

data EnvironmentResourcesDescription

environmentResourcesDescription :: EnvironmentResourcesDescription

EnvironmentResourcesDescription constructor.

The fields accessible through corresponding lenses are:

erdLoadBalancer :: Maybe LoadBalancerDescription

erdLoadBalancer :: Lens' EnvironmentResourcesDescription (Maybe LoadBalancerDescription)

Describes the LoadBalancer.

OptionRestrictionRegex

data OptionRestrictionRegex

optionRestrictionRegex :: OptionRestrictionRegex

OptionRestrictionRegex constructor.

The fields accessible through corresponding lenses are:

orrLabel :: Maybe Text

orrPattern :: Maybe Text

orrLabel :: Lens' OptionRestrictionRegex (Maybe Text)

A unique name representing this regular expression.

orrPattern :: Lens' OptionRestrictionRegex (Maybe Text)

The regular expression pattern that a string configuration option value with this restriction must match.

ConfigurationOptionDescription

data ConfigurationOptionDescription

configurationOptionDescription :: ConfigurationOptionDescription

ConfigurationOptionDescription constructor.

The fields accessible through corresponding lenses are:

codChangeSeverity :: Maybe Text

codDefaultValue :: Maybe Text

codMaxLength :: Maybe Int

codMaxValue :: Maybe Int

codMinValue :: Maybe Int

codName :: Maybe Text

codNamespace :: Maybe Text

codRegex :: Maybe OptionRestrictionRegex

codUserDefined :: Maybe Bool

codValueOptions :: [Text]

codValueType :: Maybe ConfigurationOptionValueType

codChangeSeverity :: Lens' ConfigurationOptionDescription (Maybe Text)

An indication of which action is required if the value for this configuration option changes:

NoInterruption - There is no interruption to the environment or application availability.

RestartEnvironment - The environment is restarted, all AWS resources are deleted and recreated, and the environment is unavailable during the process.

RestartApplicationServer - The environment is available the entire time. However, a short application outage occurs when the application servers on the running Amazon EC2 instances are restarted.

NoInterruption : There is no interruption to the environment or application availability. RestartEnvironment : The environment is entirely restarted, all AWS resources are deleted and recreated, and the environment is unavailable during the process. RestartApplicationServer : The environment is available the entire time. However, a short application outage occurs when the application servers on the running Amazon EC2 instances are restarted.

codDefaultValue :: Lens' ConfigurationOptionDescription (Maybe Text)

The default value for this configuration option.

codMaxLength :: Lens' ConfigurationOptionDescription (Maybe Int)

If specified, the configuration option must be a string value no longer than this value.

codMaxValue :: Lens' ConfigurationOptionDescription (Maybe Int)

If specified, the configuration option must be a numeric value less than this value.

codMinValue :: Lens' ConfigurationOptionDescription (Maybe Int)

If specified, the configuration option must be a numeric value greater than this value.

codName :: Lens' ConfigurationOptionDescription (Maybe Text)

The name of the configuration option.

codNamespace :: Lens' ConfigurationOptionDescription (Maybe Text)

A unique namespace identifying the option's associated AWS resource.

codRegex :: Lens' ConfigurationOptionDescription (Maybe OptionRestrictionRegex)

If specified, the configuration option must be a string value that satisfies this regular expression.

codUserDefined :: Lens' ConfigurationOptionDescription (Maybe Bool)

An indication of whether the user defined this configuration option:

true : This configuration option was defined by the user. It is a valid choice for specifying this as an Option to Remove when updating configuration settings.

false : This configuration was not defined by the user.

true : This configuration option was defined by the user. It is a valid choice for specifying if this as an 'Option to Remove' when updating configuration settings.

false : This configuration was not defined by the user. Constraint: You can remove only UserDefined options from a configuration.

Valid Values: true | false

codValueOptions :: Lens' ConfigurationOptionDescription [Text]

If specified, values for the configuration option are selected from this list.

codValueType :: Lens' ConfigurationOptionDescription (Maybe ConfigurationOptionValueType)

An indication of which type of values this option has and whether it is allowable to select one or more than one of the possible values:

Scalar : Values for this option are a single selection from the possible values, or a unformatted string or numeric value governed by the MINMAXRegex constraints:

List : Values for this option are multiple selections of the possible values.

Boolean : Values for this option are either true or false .

Scalar : Values for this option are a single selection from the possible values, or an unformatted string, or numeric value governed by the 'MIN/MAX/Regex' constraints. List : Values for this option are multiple selections from the possible values. Boolean : Values for this option are either true or false .

SourceConfiguration

data SourceConfiguration

sourceConfiguration :: SourceConfiguration

SourceConfiguration constructor.

The fields accessible through corresponding lenses are:

scApplicationName :: Maybe Text

scTemplateName :: Maybe Text

scApplicationName :: Lens' SourceConfiguration (Maybe Text)

The name of the application associated with the configuration.

scTemplateName :: Lens' SourceConfiguration (Maybe Text)

The name of the configuration template.

EnvironmentInfoDescription

data EnvironmentInfoDescription

environmentInfoDescription :: EnvironmentInfoDescription

EnvironmentInfoDescription constructor.

The fields accessible through corresponding lenses are:

eidEc2InstanceId :: Maybe Text

eidInfoType :: Maybe EnvironmentInfoType

eidMessage :: Maybe Text

eidSampleTimestamp :: Maybe UTCTime

eidEc2InstanceId :: Lens' EnvironmentInfoDescription (Maybe Text)

The Amazon EC2 Instance ID for this information.

eidInfoType :: Lens' EnvironmentInfoDescription (Maybe EnvironmentInfoType)

The type of information retrieved.

eidMessage :: Lens' EnvironmentInfoDescription (Maybe Text)

The retrieved information.

eidSampleTimestamp :: Lens' EnvironmentInfoDescription (Maybe UTCTime)

The time stamp when this information was retrieved.

S3Location

data S3Location

s3Location :: S3Location

S3Location constructor.

The fields accessible through corresponding lenses are:

slS3Bucket :: Maybe Text

slS3Key :: Maybe Text

slS3Bucket :: Lens' S3Location (Maybe Text)

The Amazon S3 bucket where the data is located.

slS3Key :: Lens' S3Location (Maybe Text)

The Amazon S3 key where the data is located.

ValidationMessage

data ValidationMessage

validationMessage :: ValidationMessage

ValidationMessage constructor.

The fields accessible through corresponding lenses are:

vmMessage :: Maybe Text

vmNamespace :: Maybe Text

vmOptionName :: Maybe Text

vmSeverity :: Maybe ValidationSeverity

vmMessage :: Lens' ValidationMessage (Maybe Text)

A message describing the error or warning.

vmNamespace :: Lens' ValidationMessage (Maybe Text)

vmOptionName :: Lens' ValidationMessage (Maybe Text)

vmSeverity :: Lens' ValidationMessage (Maybe ValidationSeverity)

An indication of the severity of this message:

error: This message indicates that this is not a valid setting for an option.

warning: This message is providing information you should take into account.

error: This message indicates that this is not a valid setting for an option. warning: This message is providing information you should take into account.

ValidationSeverity

data ValidationSeverity

Constructors

VSError
error

VSWarning
warning

Trigger

data Trigger

trigger :: Trigger

Trigger constructor.

The fields accessible through corresponding lenses are:

tName :: Maybe Text

tName :: Lens' Trigger (Maybe Text)

The name of the trigger.

EnvironmentInfoType

data EnvironmentInfoType

Constructors

Bundle
bundle

Tail'
tail

EnvironmentDescription

data EnvironmentDescription

environmentDescription :: EnvironmentDescription

EnvironmentDescription constructor.

The fields accessible through corresponding lenses are:

ed1AbortableOperationInProgress :: Maybe Bool

ed1ApplicationName :: Maybe Text

ed1CNAME :: Maybe Text

ed1DateCreated :: Maybe UTCTime

ed1DateUpdated :: Maybe UTCTime

ed1Description :: Maybe Text

ed1EndpointURL :: Maybe Text

ed1EnvironmentId :: Maybe Text

ed1EnvironmentName :: Maybe Text

ed1Health :: Maybe EnvironmentHealth

ed1Resources :: Maybe EnvironmentResourcesDescription

ed1SolutionStackName :: Maybe Text

ed1Status :: Maybe EnvironmentStatus

ed1TemplateName :: Maybe Text

ed1Tier :: Maybe EnvironmentTier

ed1VersionLabel :: Maybe Text

ed1AbortableOperationInProgress :: Lens' EnvironmentDescription (Maybe Bool)

Lists in-progress environment updates and application version deployments that you can cancel.

ed1ApplicationName :: Lens' EnvironmentDescription (Maybe Text)

The name of the application associated with this environment.

ed1CNAME :: Lens' EnvironmentDescription (Maybe Text)

The URL to the CNAME for this environment.

ed1DateCreated :: Lens' EnvironmentDescription (Maybe UTCTime)

The creation date for this environment.

ed1DateUpdated :: Lens' EnvironmentDescription (Maybe UTCTime)

The last modified date for this environment.

ed1Description :: Lens' EnvironmentDescription (Maybe Text)

Describes this environment.

ed1EndpointURL :: Lens' EnvironmentDescription (Maybe Text)

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

ed1EnvironmentId :: Lens' EnvironmentDescription (Maybe Text)

The ID of this environment.

ed1EnvironmentName :: Lens' EnvironmentDescription (Maybe Text)

The name of this environment.

ed1Health :: Lens' EnvironmentDescription (Maybe EnvironmentHealth)

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment:

Red : Indicates the environment is not working.

Yellow: Indicates that something is wrong, the application might not be available, but the instances appear running.

Green: Indicates the environment is healthy and fully functional.

Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. Green: Indicates the environment is healthy and fully functional. Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey

ed1Resources :: Lens' EnvironmentDescription (Maybe EnvironmentResourcesDescription)

The description of the AWS resources used by this environment.

ed1SolutionStackName :: Lens' EnvironmentDescription (Maybe Text)

The name of the SolutionStack deployed with this environment.

ed1Status :: Lens' EnvironmentDescription (Maybe EnvironmentStatus)

The current operational status of the environment:

Launching: Environment is in the process of initial deployment. Updating: Environment is in the process of updating its configuration settings or application version. Ready: Environment is available to have an action performed on it, such as update or terminate. Terminating: Environment is in the shut-down process. Terminated: Environment is not running.

ed1TemplateName :: Lens' EnvironmentDescription (Maybe Text)

The name of the configuration template used to originally launch this environment.

ed1Tier :: Lens' EnvironmentDescription (Maybe EnvironmentTier)

Describes the current tier of this environment.

ed1VersionLabel :: Lens' EnvironmentDescription (Maybe Text)

The application version deployed in this environment.

Listener

data Listener

listener :: Listener

Listener constructor.

The fields accessible through corresponding lenses are:

lPort :: Maybe Int

lProtocol :: Maybe Text

lPort :: Lens' Listener (Maybe Int)

The port that is used by the Listener.

lProtocol :: Lens' Listener (Maybe Text)

The protocol that is used by the Listener.

EnvironmentHealth

data EnvironmentHealth

Constructors

Green
Green

Grey
Grey

Red
Red

Yellow
Yellow

Instance

data Instance

instance' :: Instance

Instance constructor.

The fields accessible through corresponding lenses are:

iId :: Maybe Text

iId :: Lens' Instance (Maybe Text)

The ID of the Amazon EC2 instance.

SolutionStackDescription

data SolutionStackDescription

solutionStackDescription :: SolutionStackDescription

SolutionStackDescription constructor.

The fields accessible through corresponding lenses are:

ssdPermittedFileTypes :: [Text]

ssdSolutionStackName :: Maybe Text

ssdPermittedFileTypes :: Lens' SolutionStackDescription [Text]

The permitted file types allowed for a solution stack.

ssdSolutionStackName :: Lens' SolutionStackDescription (Maybe Text)

The name of the solution stack.