Network.AWS.ElasticBeanstalk.UpdateApplicationVersion

Description

Updates the specified application version to have the specified properties.

If a property (for example, description) is not provided, the value remains unchanged. To clear properties, specify an empty string.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_UpdateApplicationVersion.html

Request

data UpdateApplicationVersion

Request constructor

updateApplicationVersion

Arguments

:: Text
uavApplicationName

-> Text
uavVersionLabel

-> UpdateApplicationVersion

UpdateApplicationVersion constructor.

The fields accessible through corresponding lenses are:

uavApplicationName :: Text

uavDescription :: Maybe Text

uavVersionLabel :: Text

Request lenses

uavApplicationName :: Lens' UpdateApplicationVersion Text

The name of the application associated with this version.

If no application is found with this name, UpdateApplication returns an InvalidParameterValue error.

uavDescription :: Lens' UpdateApplicationVersion (Maybe Text)

A new description for this release.

uavVersionLabel :: Lens' UpdateApplicationVersion Text

The name of the version to update.

If no application version is found with this label, UpdateApplication returns an InvalidParameterValue error.

Response

data UpdateApplicationVersionResponse

Response constructor

updateApplicationVersionResponse :: UpdateApplicationVersionResponse

UpdateApplicationVersionResponse constructor.

The fields accessible through corresponding lenses are:

uavrApplicationVersion :: Maybe ApplicationVersionDescription

Response lenses

uavrApplicationVersion :: Lens' UpdateApplicationVersionResponse (Maybe ApplicationVersionDescription)

The ApplicationVersionDescription of the application version.