Network.AWS.ElasticBeanstalk.UpdateConfigurationTemplate

Description

Updates the specified configuration template to have the specified properties or configuration option values.

If a property (for example, ApplicationName) is not provided, its value remains unchanged. To clear such properties, specify an empty string. Related Topics

DescribeConfigurationOptions

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_UpdateConfigurationTemplate.html

Request

data UpdateConfigurationTemplate

Request constructor

updateConfigurationTemplate

Arguments

:: Text
uctApplicationName

-> Text
uctTemplateName

-> UpdateConfigurationTemplate

UpdateConfigurationTemplate constructor.

The fields accessible through corresponding lenses are:

uctApplicationName :: Text

uctDescription :: Maybe Text

uctOptionSettings :: [ConfigurationOptionSetting]

uctOptionsToRemove :: [OptionSpecification]

uctTemplateName :: Text

Request lenses

uctApplicationName :: Lens' UpdateConfigurationTemplate Text

The name of the application associated with the configuration template to update.

If no application is found with this name, UpdateConfigurationTemplate returns an InvalidParameterValue error.

uctDescription :: Lens' UpdateConfigurationTemplate (Maybe Text)

A new description for the configuration.

uctOptionSettings :: Lens' UpdateConfigurationTemplate [ConfigurationOptionSetting]

A list of configuration option settings to update with the new specified option value.

uctOptionsToRemove :: Lens' UpdateConfigurationTemplate [OptionSpecification]

A list of configuration options to remove from the configuration set.

Constraint: You can remove only UserDefined configuration options.

uctTemplateName :: Lens' UpdateConfigurationTemplate Text

The name of the configuration template to update.

If no configuration template is found with this name, UpdateConfigurationTemplate returns an InvalidParameterValue error.

Response

data UpdateConfigurationTemplateResponse

Response constructor

updateConfigurationTemplateResponse :: UpdateConfigurationTemplateResponse

UpdateConfigurationTemplateResponse constructor.

The fields accessible through corresponding lenses are:

uctrApplicationName :: Maybe Text

uctrDateCreated :: Maybe UTCTime

uctrDateUpdated :: Maybe UTCTime

uctrDeploymentStatus :: Maybe ConfigurationDeploymentStatus

uctrDescription :: Maybe Text

uctrEnvironmentName :: Maybe Text

uctrOptionSettings :: [ConfigurationOptionSetting]

uctrSolutionStackName :: Maybe Text

uctrTemplateName :: Maybe Text

Response lenses

uctrApplicationName :: Lens' UpdateConfigurationTemplateResponse (Maybe Text)

The name of the application associated with this configuration set.

uctrDateCreated :: Lens' UpdateConfigurationTemplateResponse (Maybe UTCTime)

The date (in UTC time) when this configuration set was created.

uctrDateUpdated :: Lens' UpdateConfigurationTemplateResponse (Maybe UTCTime)

The date (in UTC time) when this configuration set was last modified.

uctrDeploymentStatus :: Lens' UpdateConfigurationTemplateResponse (Maybe ConfigurationDeploymentStatus)

If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set:

null: This configuration is not associated with a running environment.

pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying.

deployed: This is the configuration that is currently deployed to the associated running environment.

failed: This is a draft configuration, that failed to successfully deploy.

null: This configuration is not associated with a running environment. pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying. deployed: This is the configuration that is currently deployed to the associated running environment. failed: This is a draft configuration that failed to successfully deploy.

uctrDescription :: Lens' UpdateConfigurationTemplateResponse (Maybe Text)

Describes this configuration set.

uctrEnvironmentName :: Lens' UpdateConfigurationTemplateResponse (Maybe Text)

If not null, the name of the environment for this configuration set.

uctrOptionSettings :: Lens' UpdateConfigurationTemplateResponse [ConfigurationOptionSetting]

A list of the configuration options and their values in this configuration set.

uctrSolutionStackName :: Lens' UpdateConfigurationTemplateResponse (Maybe Text)

The name of the solution stack this configuration set uses.

uctrTemplateName :: Lens' UpdateConfigurationTemplateResponse (Maybe Text)

If not null, the name of the configuration template for this configuration set.