Network.AWS.ElasticBeanstalk.UpdateEnvironment

Description

Updates the environment description, deploys a new application version, updates the configuration settings to an entirely new configuration template, or updates select configuration option values in the running environment.

Attempting to update both the release and configuration is not allowed and AWS Elastic Beanstalk returns an InvalidParameterCombination error.

When updating the configuration settings to a new template or individual settings, a draft configuration is created and DescribeConfigurationSettings for this environment returns two setting descriptions with different DeploymentStatus values.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_UpdateEnvironment.html

Request

data UpdateEnvironment

Request constructor

updateEnvironment :: UpdateEnvironment

UpdateEnvironment constructor.

The fields accessible through corresponding lenses are:

ueDescription :: Maybe Text

ueEnvironmentId :: Maybe Text

ueEnvironmentName :: Maybe Text

ueOptionSettings :: [ConfigurationOptionSetting]

ueOptionsToRemove :: [OptionSpecification]

ueSolutionStackName :: Maybe Text

ueTemplateName :: Maybe Text

ueTier :: Maybe EnvironmentTier

ueVersionLabel :: Maybe Text

Request lenses

ueDescription :: Lens' UpdateEnvironment (Maybe Text)

If this parameter is specified, AWS Elastic Beanstalk updates the description of this environment.

ueEnvironmentId :: Lens' UpdateEnvironment (Maybe Text)

The ID of the environment to update.

If no environment with this ID exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

ueEnvironmentName :: Lens' UpdateEnvironment (Maybe Text)

The name of the environment to update. If no environment with this name exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

ueOptionSettings :: Lens' UpdateEnvironment [ConfigurationOptionSetting]

If specified, AWS Elastic Beanstalk updates the configuration set associated with the running environment and sets the specified configuration options to the requested value.

ueOptionsToRemove :: Lens' UpdateEnvironment [OptionSpecification]

A list of custom user-defined configuration options to remove from the configuration set for this environment.

ueSolutionStackName :: Lens' UpdateEnvironment (Maybe Text)

This specifies the platform version that the environment will run after the environment is updated.

ueTemplateName :: Lens' UpdateEnvironment (Maybe Text)

If this parameter is specified, AWS Elastic Beanstalk deploys this configuration template to the environment. If no such configuration template is found, AWS Elastic Beanstalk returns an InvalidParameterValue error.

ueTier :: Lens' UpdateEnvironment (Maybe EnvironmentTier)

This specifies the tier to use to update the environment.

Condition: You can only update the tier version for an environment. If you change the name of the type, AWS Elastic Beanstalk returns InvalidParameterValue error.

ueVersionLabel :: Lens' UpdateEnvironment (Maybe Text)

If this parameter is specified, AWS Elastic Beanstalk deploys the named application version to the environment. If no such application version is found, returns an InvalidParameterValue error.

Response

data UpdateEnvironmentResponse

Response constructor

updateEnvironmentResponse :: UpdateEnvironmentResponse

UpdateEnvironmentResponse constructor.

The fields accessible through corresponding lenses are:

uerAbortableOperationInProgress :: Maybe Bool

uerApplicationName :: Maybe Text

uerCNAME :: Maybe Text

uerDateCreated :: Maybe UTCTime

uerDateUpdated :: Maybe UTCTime

uerDescription :: Maybe Text

uerEndpointURL :: Maybe Text

uerEnvironmentId :: Maybe Text

uerEnvironmentName :: Maybe Text

uerHealth :: Maybe EnvironmentHealth

uerResources :: Maybe EnvironmentResourcesDescription

uerSolutionStackName :: Maybe Text

uerStatus :: Maybe EnvironmentStatus

uerTemplateName :: Maybe Text

uerTier :: Maybe EnvironmentTier

uerVersionLabel :: Maybe Text

Response lenses

uerAbortableOperationInProgress :: Lens' UpdateEnvironmentResponse (Maybe Bool)

Lists in-progress environment updates and application version deployments that you can cancel.

uerApplicationName :: Lens' UpdateEnvironmentResponse (Maybe Text)

The name of the application associated with this environment.

uerCNAME :: Lens' UpdateEnvironmentResponse (Maybe Text)

The URL to the CNAME for this environment.

uerDateCreated :: Lens' UpdateEnvironmentResponse (Maybe UTCTime)

The creation date for this environment.

uerDateUpdated :: Lens' UpdateEnvironmentResponse (Maybe UTCTime)

The last modified date for this environment.

uerDescription :: Lens' UpdateEnvironmentResponse (Maybe Text)

Describes this environment.

uerEndpointURL :: Lens' UpdateEnvironmentResponse (Maybe Text)

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

uerEnvironmentId :: Lens' UpdateEnvironmentResponse (Maybe Text)

The ID of this environment.

uerEnvironmentName :: Lens' UpdateEnvironmentResponse (Maybe Text)

The name of this environment.

uerHealth :: Lens' UpdateEnvironmentResponse (Maybe EnvironmentHealth)

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment:

Red : Indicates the environment is not working.

Yellow: Indicates that something is wrong, the application might not be available, but the instances appear running.

Green: Indicates the environment is healthy and fully functional.

Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. Green: Indicates the environment is healthy and fully functional. Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey

uerResources :: Lens' UpdateEnvironmentResponse (Maybe EnvironmentResourcesDescription)

The description of the AWS resources used by this environment.

uerSolutionStackName :: Lens' UpdateEnvironmentResponse (Maybe Text)

The name of the SolutionStack deployed with this environment.

uerStatus :: Lens' UpdateEnvironmentResponse (Maybe EnvironmentStatus)

The current operational status of the environment:

Launching: Environment is in the process of initial deployment. Updating: Environment is in the process of updating its configuration settings or application version. Ready: Environment is available to have an action performed on it, such as update or terminate. Terminating: Environment is in the shut-down process. Terminated: Environment is not running.

uerTemplateName :: Lens' UpdateEnvironmentResponse (Maybe Text)

The name of the configuration template used to originally launch this environment.

uerTier :: Lens' UpdateEnvironmentResponse (Maybe EnvironmentTier)

Describes the current tier of this environment.

uerVersionLabel :: Lens' UpdateEnvironmentResponse (Maybe Text)

The application version deployed in this environment.