Network.AWS.ElasticBeanstalk.V2010_12_01.CreateApplication

Description

Creates an application that has one configuration template named default and no application versions. The &lt;code&gt;default&lt;/code&gt; configuration template is for a 32-bit version of the Amazon Linux operating system running the Tomcat 6 application container. &lt;/note&gt;"> https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &Description=Sample%20Description &Operation=CreateApplication &AuthParams Sample Description SampleApp 2010-11-16T23:09:20.256Z 2010-11-16T23:09:20.256Z Default 8b00e053-f1d6-11df-8a78-9f77047e0d0c.

Request

data CreateApplication

This documentation target is not reported in the API reference.

Request constructor

mkCreateApplication

Arguments

:: Text
caApplicationName

-> CreateApplication

Smart constructor for the minimum required parameters to construct a valid CreateApplication request.

The fields accessible through corresponding lenses are:

ApplicationName :: Text

Description :: Maybe Text

Request lenses

caApplicationName :: Lens' CreateApplication Text

The name of the application. Constraint: This name must be unique within your account. If the specified name already exists, the action returns an InvalidParameterValue error.

caDescription :: Lens' CreateApplication (Maybe Text)

Describes the application.

Response

data CreateApplicationResponse

Result message containing a single description of an application.

Response constructor

mkCreateApplicationResponse :: CreateApplicationResponse

Smart constructor for the minimum required parameters to construct a valid CreateApplicationResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Application :: Maybe ApplicationDescription

Response lenses

carApplication :: Lens' CreateApplicationResponse (Maybe ApplicationDescription)

The ApplicationDescription of the application.