Network.AWS.ElasticBeanstalk.V2010_12_01.DeleteConfigurationTemplate

Description

Deletes the specified configuration template. When you launch an environment using a configuration template, the environment gets a copy of the template. You can delete or modify the environment's copy of the template without affecting the running environment. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &TemplateName=SampleAppTemplate &Operation=DeleteConfigurationTemplate &AuthParams af9cf1b6-f25e-11df-8a78-9f77047e0d0c.

Request

data DeleteConfigurationTemplate

This documentation target is not reported in the API reference.

Request constructor

mkDeleteConfigurationTemplate

Arguments

:: Text
dctApplicationName

-> Text
dctTemplateName

-> DeleteConfigurationTemplate

Smart constructor for the minimum required parameters to construct a valid DeleteConfigurationTemplate request.

Request lenses

dctApplicationName :: Lens' DeleteConfigurationTemplate Text

The name of the application to delete the configuration template from.

dctTemplateName :: Lens' DeleteConfigurationTemplate Text

The name of the configuration template to delete.

Response

data DeleteConfigurationTemplateResponse