Network.AWS.ElasticBeanstalk.V2010_12_01.DeleteEnvironmentConfiguration

Description

Deletes the draft configuration associated with the running environment. Updating a running environment with any configuration changes creates a draft configuration set. You can get the draft configuration using DescribeConfigurationSettings while the update is in progress or if the update fails. The DeploymentStatus for the draft configuration indicates whether the deployment is in process or has failed. The draft configuration remains in existence until it is deleted with this action. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &EnvironmentName=SampleApp &Operation=DeleteEnvironmentConfiguration &AuthParams fdf76507-f26d-11df-8a78-9f77047e0d0c.

Documentation

data DeleteEnvironmentConfiguration

Constructors

DeleteEnvironmentConfiguration

Fields

_decmApplicationName :: Text

The name of the application the environment is associated with.

_decmEnvironmentName :: Text

The name of the environment to delete the draft configuration from.

decmEnvironmentName :: Lens' DeleteEnvironmentConfiguration Text

decmApplicationName :: Lens' DeleteEnvironmentConfiguration Text

data DeleteEnvironmentConfigurationResponse

Constructors

DeleteEnvironmentConfigurationResponse