Network.AWS.ElasticBeanstalk.V2010_12_01.DescribeApplicationVersions

Description

Returns descriptions for existing application versions. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &Operation=DescribeApplicationVersions &AuthParams amazonaws.com sample.war Version1 description SampleApp 2010-11-17T03:21:59.161Z 2010-11-17T03:21:59.161Z 773cd80a-f26c-11df-8a78-9f77047e0d0c.

Documentation

describeApplicationVersions :: DescribeApplicationVersions

Minimum specification for a DescribeApplicationVersions request.

data DescribeApplicationVersions

Constructors

DescribeApplicationVersions

Fields

_davnApplicationName :: Maybe Text

If specified, AWS Elastic Beanstalk restricts the returned descriptions to only include ones that are associated with the specified application.

_davnVersionLabels :: [Text]

If specified, restricts the returned descriptions to only include ones that have the specified version labels.

davnVersionLabels :: Lens' DescribeApplicationVersions [Text]

davnApplicationName :: Lens' DescribeApplicationVersions (Maybe Text)

data DescribeApplicationVersionsResponse

Constructors

DescribeApplicationVersionsResponse

Fields

_avdoApplicationVersions :: [ApplicationVersionDescription]

A list of ApplicationVersionDescription .

avdoApplicationVersions :: Iso' DescribeApplicationVersionsResponse [ApplicationVersionDescription]