Network.AWS.ElasticBeanstalk.V2010_12_01.DescribeApplications

Description

Returns the descriptions of existing applications. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationNames.member.1=SampleApplication &Operation=DescribeApplications &AuthParams Sample Description SampleApplication 2010-11-16T20:20:51.974Z 2010-11-16T20:20:51.974Z Default 577c70ff-f1d7-11df-8a78-9f77047e0d0c.

Request

data DescribeApplications

This documentation target is not reported in the API reference.

Request constructor

mkDescribeApplications :: DescribeApplications

Smart constructor for the minimum required parameters to construct a valid DescribeApplications request.

Request lenses

da1ApplicationNames :: Lens' DescribeApplications [Text]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to only include those with the specified names.

Response

data DescribeApplicationsResponse

Result message containing a list of application descriptions.

Response lenses

darApplications :: Lens' DescribeApplicationsResponse [ApplicationDescription]

This parameter contains a list of ApplicationDescription.