Network.AWS.ElasticBeanstalk.V2010_12_01.DescribeEnvironments

Description

Returns descriptions for existing environments. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &IncludeDeleted=true &IncludedDeletedBackTo=2008-11-05T06%3A00%3A00Z &Operation=DescribeEnvironments &AuthParams Version1 Available SampleApp elasticbeanstalk-SampleApp-1394386994.us-east-1.elb.amazonaws.com SampleApp-jxb293wg7n.elasticbeanstalk.amazonaws.com Green e-icsgecu3wf 2010-11-17T04:01:40.668Z 32bit Amazon Linux running Tomcat 7 EnvDescrip SampleApp 2010-11-17T03:59:33.520Z 44790c68-f260-11df-8a78-9f77047e0d0c.

Request

data DescribeEnvironments

This documentation target is not reported in the API reference.

Request constructor

mkDescribeEnvironments :: DescribeEnvironments

Smart constructor for the minimum required parameters to construct a valid DescribeEnvironments request.

Request lenses

deApplicationName :: Lens' DescribeEnvironments (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application.

deVersionLabel :: Lens' DescribeEnvironments (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application version.

deEnvironmentIds :: Lens' DescribeEnvironments [Text]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified IDs.

deEnvironmentNames :: Lens' DescribeEnvironments [Text]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified names.

deIncludeDeleted :: Lens' DescribeEnvironments (Maybe Bool)

Indicates whether to include deleted environments: true: Environments that have been deleted after IncludedDeletedBackTo are displayed. false: Do not include deleted environments.

deIncludedDeletedBackTo :: Lens' DescribeEnvironments (Maybe ISO8601)

If specified when IncludeDeleted is set to true, then environments deleted after this date are displayed.

Response

data DescribeEnvironmentsResponse

Result message containing a list of environment descriptions.

Response lenses

der1Environments :: Lens' DescribeEnvironmentsResponse [EnvironmentDescription]

Returns an EnvironmentDescription list.