Network.AWS.ElasticBeanstalk.V2010_12_01.ListAvailableSolutionStacks

Description

Returns a list of the available solution stack names. https:/elasticbeanstalk.us-east-1.amazon.com?Operation=ListAvailableSolutionStacks &AuthParams 64bit Amazon Linux running Tomcat 6 32bit Amazon Linux running Tomcat 6 64bit Amazon Linux running Tomcat 7 32bit Amazon Linux running Tomcat 7 f21e2a92-f1fc-11df-8a78-9f77047e0d0c.

Documentation

data ListAvailableSolutionStacks

Constructors

ListAvailableSolutionStacks

data ListAvailableSolutionStacksResponse

Constructors

ListAvailableSolutionStacksResponse

Fields

_lassrmSolutionStackDetails :: [SolutionStackDescription]

A list of available solution stacks and their SolutionStackDescription.

_lassrmSolutionStacks :: [Text]

A list of available solution stacks.

lassrmSolutionStacks :: Lens' ListAvailableSolutionStacksResponse [Text]

lassrmSolutionStackDetails :: Lens' ListAvailableSolutionStacksResponse [SolutionStackDescription]