Network.AWS.ElasticBeanstalk.V2010_12_01.TerminateEnvironment

Description

Terminates the specified environment. https:/elasticbeanstalk.us-east-1.amazon.com?EnvironmentId=e-icsgecu3wf &EnvironmentName=SampleApp &TerminateResources=true &Operation=TerminateEnvironment &AuthParams Version1 Terminating SampleApp elasticbeanstalk-SampleApp-1394386994.us-east-1.elb.amazonaws.com SampleApp-jxb293wg7n.elasticbeanstalk.amazonaws.com Grey e-icsgecu3wf 2010-11-17T17:10:41.976Z 32bit Amazon Linux running Tomcat 7 EnvDescrip SampleApp 2010-11-17T03:59:33.520Z 9b71af21-f26d-11df-8a78-9f77047e0d0c.

Request

data TerminateEnvironment

This documentation target is not reported in the API reference.

Request constructor

mkTerminateEnvironment :: TerminateEnvironment

Smart constructor for the minimum required parameters to construct a valid TerminateEnvironment request.

The fields accessible through corresponding lenses are:

EnvironmentId :: Maybe Text

EnvironmentName :: Maybe Text

TerminateResources :: Maybe Bool

Request lenses

teEnvironmentId :: Lens' TerminateEnvironment (Maybe Text)

The ID of the environment to terminate. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

teEnvironmentName :: Lens' TerminateEnvironment (Maybe Text)

The name of the environment to terminate. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

teTerminateResources :: Lens' TerminateEnvironment (Maybe Bool)

Indicates whether the associated AWS resources should shut down when the environment is terminated: true: (default) The user AWS resources (for example, the Auto Scaling group, LoadBalancer, etc.) are terminated along with the environment. false: The environment is removed from the AWS Elastic Beanstalk but the AWS resources continue to operate. true: The specified environment as well as the associated AWS resources, such as Auto Scaling group and LoadBalancer, are terminated. false: AWS Elastic Beanstalk resource management is removed from the environment, but the AWS resources continue to operate. For more information, see the AWS Elastic Beanstalk User Guide. Default: true Valid Values: true | false.

Response

data TerminateEnvironmentResponse

Describes the properties of an environment.

Response constructor

mkTerminateEnvironmentResponse :: TerminateEnvironmentResponse

Smart constructor for the minimum required parameters to construct a valid TerminateEnvironmentResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

EnvironmentName :: Maybe Text

EnvironmentId :: Maybe Text

ApplicationName :: Maybe Text

VersionLabel :: Maybe Text

SolutionStackName :: Maybe Text

TemplateName :: Maybe Text

Description :: Maybe Text

EndpointURL :: Maybe Text

CNAME :: Maybe Text

DateCreated :: Maybe ISO8601

DateUpdated :: Maybe ISO8601

Status :: Maybe EnvironmentStatus

Health :: Maybe EnvironmentHealth

Resources :: Maybe EnvironmentResourcesDescription

Tier :: Maybe EnvironmentTier

Response lenses

terEnvironmentName :: Lens' TerminateEnvironmentResponse (Maybe Text)

The name of this environment.

terEnvironmentId :: Lens' TerminateEnvironmentResponse (Maybe Text)

The ID of this environment.

terApplicationName :: Lens' TerminateEnvironmentResponse (Maybe Text)

The name of the application associated with this environment.

terVersionLabel :: Lens' TerminateEnvironmentResponse (Maybe Text)

The application version deployed in this environment.

terSolutionStackName :: Lens' TerminateEnvironmentResponse (Maybe Text)

The name of the SolutionStack deployed with this environment.

terTemplateName :: Lens' TerminateEnvironmentResponse (Maybe Text)

The name of the configuration template used to originally launch this environment.

terDescription :: Lens' TerminateEnvironmentResponse (Maybe Text)

Describes this environment.

terEndpointURL :: Lens' TerminateEnvironmentResponse (Maybe Text)

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

terCNAME :: Lens' TerminateEnvironmentResponse (Maybe Text)

The URL to the CNAME for this environment.

terDateCreated :: Lens' TerminateEnvironmentResponse (Maybe ISO8601)

The creation date for this environment.

terDateUpdated :: Lens' TerminateEnvironmentResponse (Maybe ISO8601)

The last modified date for this environment.

terStatus :: Lens' TerminateEnvironmentResponse (Maybe EnvironmentStatus)

The current operational status of the environment: Launching: Environment is in the process of initial deployment. Updating: Environment is in the process of updating its configuration settings or application version. Ready: Environment is available to have an action performed on it, such as update or terminate. Terminating: Environment is in the shut-down process. Terminated: Environment is not running.

terHealth :: Lens' TerminateEnvironmentResponse (Maybe EnvironmentHealth)

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment: Red : Indicates the environment is not working. Yellow: Indicates that something is wrong, the application might not be available, but the instances appear running. Green: Indicates the environment is healthy and fully functional. Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. Green: Indicates the environment is healthy and fully functional. Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey.

terResources :: Lens' TerminateEnvironmentResponse (Maybe EnvironmentResourcesDescription)

The description of the AWS resources used by this environment.

terTier :: Lens' TerminateEnvironmentResponse (Maybe EnvironmentTier)

Describes the current tier of this environment.