Network.AWS.ElasticBeanstalk.V2010_12_01.Types

Description

AWS Elastic Beanstalk is an easy-to-use service for deploying and scaling web applications and services developed with Java, .NET, PHP, Node.js, Python, Ruby, and Docker on familiar servers such as Apache HTTP Server, Apache Tomcat, Nginx, Passenger, and IIS 7.5/8. You can simply upload your code and Elastic Beanstalk automatically handles the deployment, from capacity provisioning, load balancing, auto-scaling to application health monitoring. At the same time, you retain full control over the AWS resources powering your application and can access the underlying resources at any time. There is no additional charge for Elastic Beanstalk - you pay only for the AWS resources needed to store and run your applications.

Documentation

data ElasticBeanstalk

Supported version (2010-12-01) of the AWS Elastic Beanstalk service.

xmlOptions :: Tagged a XMLOptions

data ConfigurationDeploymentStatus

If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set: null: This configuration is not associated with a running environment. pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying. deployed: This is the configuration that is currently deployed to the associated running environment. failed: This is a draft configuration, that failed to successfully deploy. null: This configuration is not associated with a running environment. pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying. deployed: This is the configuration that is currently deployed to the associated running environment. failed: This is a draft configuration that failed to successfully deploy.

Constructors

ConfigurationDeploymentStatusDeployed
deployed

ConfigurationDeploymentStatusFailed
failed

ConfigurationDeploymentStatusPending
pending

data ConfigurationOptionValueType

An indication of which type of values this option has and whether it is allowable to select one or more than one of the possible values: Scalar : Values for this option are a single selection from the possible values, or a unformatted string or numeric value governed by the MINMAXRegex constraints: List : Values for this option are multiple selections of the possible values. Boolean : Values for this option are either true or false . Scalar : Values for this option are a single selection from the possible values, or an unformatted string, or numeric value governed by the MINMAXRegex constraints. List : Values for this option are multiple selections from the possible values. Boolean : Values for this option are either true or false .

Constructors

ConfigurationOptionValueTypeList
List

ConfigurationOptionValueTypeScalar
Scalar

data EnvironmentHealth

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment: Red : Indicates the environment is not working. Yellow: Indicates that something is wrong, the application might not be available, but the instances appear running. Green: Indicates the environment is healthy and fully functional. Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. Green: Indicates the environment is healthy and fully functional. Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey.

Constructors

EnvironmentHealthGreen
Green

EnvironmentHealthGrey
Grey

EnvironmentHealthRed
Red

EnvironmentHealthYellow
Yellow

data EnvironmentInfoType

The type of information to request.

Constructors

EnvironmentInfoTypeTail
tail

data EnvironmentStatus

The current operational status of the environment: Launching: Environment is in the process of initial deployment. Updating: Environment is in the process of updating its configuration settings or application version. Ready: Environment is available to have an action performed on it, such as update or terminate. Terminating: Environment is in the shut-down process. Terminated: Environment is not running.

Constructors

EnvironmentStatusLaunching
Launching

EnvironmentStatusReady
Ready

EnvironmentStatusTerminated
Terminated

EnvironmentStatusTerminating
Terminating

EnvironmentStatusUpdating
Updating

data EventSeverity

If specified, limits the events returned from this call to include only those with the specified severity or higher.

Constructors

EventSeverityDebug
DEBUG

EventSeverityError
ERROR

EventSeverityFatal
FATAL

EventSeverityInfo
INFO

EventSeverityTrace
TRACE

EventSeverityWarn
WARN

data ValidationSeverity

An indication of the severity of this message: error: This message indicates that this is not a valid setting for an option. warning: This message is providing information you should take into account. error: This message indicates that this is not a valid setting for an option. warning: This message is providing information you should take into account.

Constructors

ValidationSeverityError
error

ValidationSeverityWarning
warning

newtype AutoScalingGroup

Describes an Auto Scaling launch configuration.

Constructors

AutoScalingGroup

Fields

_asgName :: Maybe Text

The name of the AutoScalingGroup .

newtype EnvironmentResourcesDescription

The description of the AWS resources used by this environment.

Constructors

EnvironmentResourcesDescription

Fields

_erdLoadBalancer :: Maybe LoadBalancerDescription

Describes the LoadBalancer.

newtype Instance

The description of an Amazon EC2 instance.

Constructors

Instance

Fields

_ieId :: Maybe Text

The ID of the Amazon EC2 instance.

newtype LaunchConfiguration

Describes an Auto Scaling launch configuration.

Constructors

LaunchConfiguration

Fields

_lcName :: Maybe Text

The name of the launch configuration.

newtype LoadBalancer

Describes a LoadBalancer.

Constructors

LoadBalancer

Fields

_lbName :: Maybe Text

The name of the LoadBalancer.

newtype Trigger

Describes a trigger.

Constructors

Trigger

Fields

_trName :: Maybe Text

The name of the trigger.

data ApplicationDescription

Describes the properties of an application.

Constructors

ApplicationDescription

Fields

_adVersions :: [Text]

The names of the versions for this application.

_adDateUpdated :: Maybe ISO8601

The date when the application was last modified.

_adDateCreated :: Maybe ISO8601

The date when the application was created.

_adApplicationName :: Maybe Text

The name of the application.

_adConfigurationTemplates :: [Text]

The names of the configuration templates associated with this application.

_adDescription :: Maybe Text

User-defined description of the application.

data ApplicationVersionDescription

The ApplicationVersionDescription of the application version.

Constructors

ApplicationVersionDescription

Fields

_avdSourceBundle :: Maybe S3Location

The location where the source bundle is located for this version.

_avdDateUpdated :: Maybe ISO8601

The last modified date of the application version.

_avdDateCreated :: Maybe ISO8601

The creation date of the application version.

_avdVersionLabel :: Maybe Text

A label uniquely identifying the version for the associated application.

_avdApplicationName :: Maybe Text

The name of the application associated with this release.

_avdDescription :: Maybe Text

The description of this application version.

data ConfigurationOptionDescription

Describes the possible values for a configuration option.

Constructors

ConfigurationOptionDescription

Fields

_coeMaxValue :: Maybe Integer

If specified, the configuration option must be a numeric value less than this value.

_coeRegex :: Maybe OptionRestrictionRegex

If specified, the configuration option must be a string value that satisfies this regular expression.

_coeMaxLength :: Maybe Integer

If specified, the configuration option must be a string value no longer than this value.

_coeUserDefined :: Maybe Bool

An indication of whether the user defined this configuration option: true : This configuration option was defined by the user. It is a valid choice for specifying this as an Option to Remove when updating configuration settings. false : This configuration was not defined by the user. true : This configuration option was defined by the user. It is a valid choice for specifying if this as an Option to Remove when updating configuration settings. false : This configuration was not defined by the user. Constraint: You can remove only UserDefined options from a configuration. Valid Values: true | false.

_coeNamespace :: Maybe Text

A unique namespace identifying the option's associated AWS resource.

_coeValueOptions :: [Text]

If specified, values for the configuration option are selected from this list.

_coeName :: Maybe Text

The name of the configuration option.

_coeChangeSeverity :: Maybe Text

An indication of which action is required if the value for this configuration option changes: NoInterruption - There is no interruption to the environment or application availability. RestartEnvironment - The environment is restarted, all AWS resources are deleted and recreated, and the environment is unavailable during the process. RestartApplicationServer - The environment is available the entire time. However, a short application outage occurs when the application servers on the running Amazon EC2 instances are restarted. NoInterruption : There is no interruption to the environment or application availability. RestartEnvironment : The environment is entirely restarted, all AWS resources are deleted and recreated, and the environment is unavailable during the process. RestartApplicationServer : The environment is available the entire time. However, a short application outage occurs when the application servers on the running Amazon EC2 instances are restarted.

_coeDefaultValue :: Maybe Text

The default value for this configuration option.

_coeValueType :: Maybe ConfigurationOptionValueType

An indication of which type of values this option has and whether it is allowable to select one or more than one of the possible values: Scalar : Values for this option are a single selection from the possible values, or a unformatted string or numeric value governed by the MINMAXRegex constraints: List : Values for this option are multiple selections of the possible values. Boolean : Values for this option are either true or false . Scalar : Values for this option are a single selection from the possible values, or an unformatted string, or numeric value governed by the MINMAXRegex constraints. List : Values for this option are multiple selections from the possible values. Boolean : Values for this option are either true or false .

_coeMinValue :: Maybe Integer

If specified, the configuration option must be a numeric value greater than this value.

data ConfigurationOptionSetting

A specification identifying an individual configuration option along with its current value. For a list of possible option values, go to Option Values in the AWS Elastic Beanstalk Developer Guide.

Constructors

ConfigurationOptionSetting

Fields

_cosOptionName :: Maybe Text

The name of the configuration option.

_cosNamespace :: Maybe Text

A unique namespace identifying the option's associated AWS resource.

_cosValue :: Maybe Text

The current value for the configuration option.

data ConfigurationSettingsDescription

Describes the settings for a configuration set.

Constructors

ConfigurationSettingsDescription

Fields

_csgTemplateName :: Maybe Text

If not null, the name of the configuration template for this configuration set.

_csgOptionSettings :: [ConfigurationOptionSetting]

A list of the configuration options and their values in this configuration set.

_csgDateUpdated :: Maybe ISO8601

The date (in UTC time) when this configuration set was last modified.

_csgDateCreated :: Maybe ISO8601

The date (in UTC time) when this configuration set was created.

_csgEnvironmentName :: Maybe Text

If not null, the name of the environment for this configuration set.

_csgApplicationName :: Maybe Text

The name of the application associated with this configuration set.

_csgDeploymentStatus :: Maybe ConfigurationDeploymentStatus

If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set: null: This configuration is not associated with a running environment. pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying. deployed: This is the configuration that is currently deployed to the associated running environment. failed: This is a draft configuration, that failed to successfully deploy. null: This configuration is not associated with a running environment. pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying. deployed: This is the configuration that is currently deployed to the associated running environment. failed: This is a draft configuration that failed to successfully deploy.

_csgSolutionStackName :: Maybe Text

The name of the solution stack this configuration set uses.

_csgDescription :: Maybe Text

Describes this configuration set.

data EnvironmentDescription

Describes the properties of an environment.

Constructors

EnvironmentDescription

Fields

_eeenStatus :: Maybe EnvironmentStatus

The current operational status of the environment: Launching: Environment is in the process of initial deployment. Updating: Environment is in the process of updating its configuration settings or application version. Ready: Environment is available to have an action performed on it, such as update or terminate. Terminating: Environment is in the shut-down process. Terminated: Environment is not running.

_eeenCNAME :: Maybe Text

The URL to the CNAME for this environment.

_eeenTemplateName :: Maybe Text

The name of the configuration template used to originally launch this environment.

_eeenEndpointURL :: Maybe Text

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

_eeenResources :: Maybe EnvironmentResourcesDescription

The description of the AWS resources used by this environment.

_eeenDateUpdated :: Maybe ISO8601

The last modified date for this environment.

_eeenDateCreated :: Maybe ISO8601

The creation date for this environment.

_eeenHealth :: Maybe EnvironmentHealth

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment: Red : Indicates the environment is not working. Yellow: Indicates that something is wrong, the application might not be available, but the instances appear running. Green: Indicates the environment is healthy and fully functional. Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. Green: Indicates the environment is healthy and fully functional. Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey.

_eeenVersionLabel :: Maybe Text

The application version deployed in this environment.

_eeenTier :: Maybe EnvironmentTier

Describes the current tier of this environment.

_eeenEnvironmentName :: Maybe Text

The name of this environment.

_eeenApplicationName :: Maybe Text

The name of the application associated with this environment.

_eeenSolutionStackName :: Maybe Text

The name of the SolutionStack deployed with this environment.

_eeenEnvironmentId :: Maybe Text

The ID of this environment.

_eeenDescription :: Maybe Text

Describes this environment.

data EnvironmentInfoDescription

The information retrieved from the Amazon EC2 instances.

Constructors

EnvironmentInfoDescription

Fields

_eidSampleTimestamp :: Maybe ISO8601

The time stamp when this information was retrieved.

_eidEc2InstanceId :: Maybe Text

The Amazon EC2 Instance ID for this information.

_eidInfoType :: Maybe EnvironmentInfoType

The type of information retrieved.

_eidMessage :: Maybe Text

The retrieved information.

data EnvironmentResourceDescription

A list of EnvironmentResourceDescription.

Constructors

EnvironmentResourceDescription

Fields

_erfQueues :: [Queue]

The queues used by this environment.

_erfTriggers :: [Trigger]

The AutoScaling triggers in use by this environment.

_erfLoadBalancers :: [LoadBalancer]

The LoadBalancers in use by this environment.

_erfEnvironmentName :: Maybe Text

The name of the environment.

_erfInstances :: [Instance]

The Amazon EC2 instances used by this environment.

_erfLaunchConfigurations :: [LaunchConfiguration]

The Auto Scaling launch configurations in use by this environment.

_erfAutoScalingGroups :: [AutoScalingGroup]

The AutoScalingGroups used by this environment.

data EnvironmentTier

This specifies the tier to use to update the environment. Condition: You can only update the tier version for an environment. If you change the name of the type, AWS Elastic Beanstalk returns InvalidParameterValue error.

Constructors

EnvironmentTier

Fields

_etName :: Maybe Text

The name of this environment tier.

_etVersion :: Maybe Text

The version of this environment tier.

_etType :: Maybe Text

The type of this environment tier.

data EventDescription

Describes an event.

Constructors

EventDescription

Fields

_efRequestId :: Maybe Text

The web service request ID for the activity of this event.

_efTemplateName :: Maybe Text

The name of the configuration associated with this event.

_efSeverity :: Maybe EventSeverity

The severity level of this event.

_efVersionLabel :: Maybe Text

The release label for the application version associated with this event.

_efEnvironmentName :: Maybe Text

The name of the environment associated with this event.

_efApplicationName :: Maybe Text

The application associated with the event.

_efEventDate :: Maybe ISO8601

The date when the event occurred.

_efMessage :: Maybe Text

The event message.

data Listener

Describes the properties of a Listener for the LoadBalancer.

Constructors

Listener

Fields

_lProtocol :: Maybe Text

The protocol that is used by the Listener.

_lPort :: Maybe Integer

The port that is used by the Listener.

data LoadBalancerDescription

Describes the LoadBalancer.

Constructors

LoadBalancerDescription

Fields

_lbdLoadBalancerName :: Maybe Text

The name of the LoadBalancer.

_lbdDomain :: Maybe Text

The domain name of the LoadBalancer.

_lbdListeners :: [Listener]

A list of Listeners used by the LoadBalancer.

data OptionRestrictionRegex

If specified, the configuration option must be a string value that satisfies this regular expression.

Constructors

OptionRestrictionRegex

Fields

_orrPattern :: Maybe Text

The regular expression pattern that a string configuration option value with this restriction must match.

_orrLabel :: Maybe Text

A unique name representing this regular expression.

data OptionSpecification

A specification identifying an individual configuration option.

Constructors

OptionSpecification

Fields

_osOptionName :: Maybe Text

The name of the configuration option.

_osNamespace :: Maybe Text

A unique namespace identifying the option's associated AWS resource.

data Queue

Describes a queue.

Constructors

Queue

Fields

_qeURL :: Maybe Text

The URL of the queue.

_qeName :: Maybe Text

The name of the queue.

data S3Location

The Amazon S3 bucket and key that identify the location of the source bundle for this version. If data found at the Amazon S3 location exceeds the maximum allowed source bundle size, AWS Elastic Beanstalk returns an InvalidParameterValue error. The maximum size allowed is 512 MB. Default: If not specified, AWS Elastic Beanstalk uses a sample application. If only partially specified (for example, a bucket is provided but not the key) or if no data is found at the Amazon S3 location, AWS Elastic Beanstalk returns an InvalidParameterCombination error.

Constructors

S3Location

Fields

_snS3Key :: Maybe Text

The Amazon S3 key where the data is located.

_snS3Bucket :: Maybe Text

The Amazon S3 bucket where the data is located.

data SolutionStackDescription

Describes the solution stack.

Constructors

SolutionStackDescription

Fields

_ssdPermittedFileTypes :: [Text]

The permitted file types allowed for a solution stack.

_ssdSolutionStackName :: Maybe Text

The name of the solution stack.

data SourceConfiguration

If specified, AWS Elastic Beanstalk uses the configuration values from the specified configuration template to create a new configuration. Values specified in the OptionSettings parameter of this call overrides any values obtained from the SourceConfiguration. If no configuration template is found, returns an InvalidParameterValue error. Constraint: If both the solution stack name parameter and the source configuration parameters are specified, the solution stack of the source configuration template must match the specified solution stack name or else AWS Elastic Beanstalk returns an InvalidParameterCombination error.

Constructors

SourceConfiguration

Fields

_seTemplateName :: Maybe Text

The name of the configuration template.

_seApplicationName :: Maybe Text

The name of the application associated with the configuration.

data Tag

Describes a tag applied to a resource in an environment.

Constructors

Tag

Fields

_tgValue :: Maybe Text

The value of the tag.

_tgKey :: Maybe Text

The key of the tag.

data ValidationMessage

An error or warning for a desired configuration option value.

Constructors

ValidationMessage

Fields

_vyOptionName :: Maybe Text

_vySeverity :: Maybe ValidationSeverity

An indication of the severity of this message: error: This message indicates that this is not a valid setting for an option. warning: This message is providing information you should take into account. error: This message indicates that this is not a valid setting for an option. warning: This message is providing information you should take into account.

_vyNamespace :: Maybe Text

_vyMessage :: Maybe Text

A message describing the error or warning.

asgName :: Iso' AutoScalingGroup (Maybe Text)

erdLoadBalancer :: Iso' EnvironmentResourcesDescription (Maybe LoadBalancerDescription)

ieId :: Iso' Instance (Maybe Text)

lcName :: Iso' LaunchConfiguration (Maybe Text)

lbName :: Iso' LoadBalancer (Maybe Text)

trName :: Iso' Trigger (Maybe Text)

adVersions :: Lens' ApplicationDescription [Text]

adDescription :: Lens' ApplicationDescription (Maybe Text)

adDateUpdated :: Lens' ApplicationDescription (Maybe ISO8601)

adDateCreated :: Lens' ApplicationDescription (Maybe ISO8601)

adConfigurationTemplates :: Lens' ApplicationDescription [Text]

adApplicationName :: Lens' ApplicationDescription (Maybe Text)

avdVersionLabel :: Lens' ApplicationVersionDescription (Maybe Text)

avdSourceBundle :: Lens' ApplicationVersionDescription (Maybe S3Location)

avdDescription :: Lens' ApplicationVersionDescription (Maybe Text)

avdDateUpdated :: Lens' ApplicationVersionDescription (Maybe ISO8601)

avdDateCreated :: Lens' ApplicationVersionDescription (Maybe ISO8601)

avdApplicationName :: Lens' ApplicationVersionDescription (Maybe Text)

coeValueType :: Lens' ConfigurationOptionDescription (Maybe ConfigurationOptionValueType)

coeValueOptions :: Lens' ConfigurationOptionDescription [Text]

coeUserDefined :: Lens' ConfigurationOptionDescription (Maybe Bool)

coeRegex :: Lens' ConfigurationOptionDescription (Maybe OptionRestrictionRegex)

coeNamespace :: Lens' ConfigurationOptionDescription (Maybe Text)

coeName :: Lens' ConfigurationOptionDescription (Maybe Text)

coeMinValue :: Lens' ConfigurationOptionDescription (Maybe Integer)

coeMaxValue :: Lens' ConfigurationOptionDescription (Maybe Integer)

coeMaxLength :: Lens' ConfigurationOptionDescription (Maybe Integer)

coeDefaultValue :: Lens' ConfigurationOptionDescription (Maybe Text)

coeChangeSeverity :: Lens' ConfigurationOptionDescription (Maybe Text)

cosValue :: Lens' ConfigurationOptionSetting (Maybe Text)

cosOptionName :: Lens' ConfigurationOptionSetting (Maybe Text)

cosNamespace :: Lens' ConfigurationOptionSetting (Maybe Text)

csgTemplateName :: Lens' ConfigurationSettingsDescription (Maybe Text)

csgSolutionStackName :: Lens' ConfigurationSettingsDescription (Maybe Text)

csgOptionSettings :: Lens' ConfigurationSettingsDescription [ConfigurationOptionSetting]

csgEnvironmentName :: Lens' ConfigurationSettingsDescription (Maybe Text)

csgDescription :: Lens' ConfigurationSettingsDescription (Maybe Text)

csgDeploymentStatus :: Lens' ConfigurationSettingsDescription (Maybe ConfigurationDeploymentStatus)

csgDateUpdated :: Lens' ConfigurationSettingsDescription (Maybe ISO8601)

csgDateCreated :: Lens' ConfigurationSettingsDescription (Maybe ISO8601)

csgApplicationName :: Lens' ConfigurationSettingsDescription (Maybe Text)

eeenVersionLabel :: Lens' EnvironmentDescription (Maybe Text)

eeenTier :: Lens' EnvironmentDescription (Maybe EnvironmentTier)

eeenTemplateName :: Lens' EnvironmentDescription (Maybe Text)

eeenStatus :: Lens' EnvironmentDescription (Maybe EnvironmentStatus)

eeenSolutionStackName :: Lens' EnvironmentDescription (Maybe Text)

eeenResources :: Lens' EnvironmentDescription (Maybe EnvironmentResourcesDescription)

eeenHealth :: Lens' EnvironmentDescription (Maybe EnvironmentHealth)

eeenEnvironmentName :: Lens' EnvironmentDescription (Maybe Text)

eeenEnvironmentId :: Lens' EnvironmentDescription (Maybe Text)

eeenEndpointURL :: Lens' EnvironmentDescription (Maybe Text)

eeenDescription :: Lens' EnvironmentDescription (Maybe Text)

eeenDateUpdated :: Lens' EnvironmentDescription (Maybe ISO8601)

eeenDateCreated :: Lens' EnvironmentDescription (Maybe ISO8601)

eeenCNAME :: Lens' EnvironmentDescription (Maybe Text)

eeenApplicationName :: Lens' EnvironmentDescription (Maybe Text)

eidSampleTimestamp :: Lens' EnvironmentInfoDescription (Maybe ISO8601)

eidMessage :: Lens' EnvironmentInfoDescription (Maybe Text)

eidInfoType :: Lens' EnvironmentInfoDescription (Maybe EnvironmentInfoType)

eidEc2InstanceId :: Lens' EnvironmentInfoDescription (Maybe Text)

erfTriggers :: Lens' EnvironmentResourceDescription [Trigger]

erfQueues :: Lens' EnvironmentResourceDescription [Queue]

erfLoadBalancers :: Lens' EnvironmentResourceDescription [LoadBalancer]

erfLaunchConfigurations :: Lens' EnvironmentResourceDescription [LaunchConfiguration]

erfInstances :: Lens' EnvironmentResourceDescription [Instance]

erfEnvironmentName :: Lens' EnvironmentResourceDescription (Maybe Text)

erfAutoScalingGroups :: Lens' EnvironmentResourceDescription [AutoScalingGroup]

etVersion :: Lens' EnvironmentTier (Maybe Text)

etType :: Lens' EnvironmentTier (Maybe Text)

etName :: Lens' EnvironmentTier (Maybe Text)

efVersionLabel :: Lens' EventDescription (Maybe Text)

efTemplateName :: Lens' EventDescription (Maybe Text)

efSeverity :: Lens' EventDescription (Maybe EventSeverity)

efRequestId :: Lens' EventDescription (Maybe Text)

efMessage :: Lens' EventDescription (Maybe Text)

efEventDate :: Lens' EventDescription (Maybe ISO8601)

efEnvironmentName :: Lens' EventDescription (Maybe Text)

efApplicationName :: Lens' EventDescription (Maybe Text)

lProtocol :: Lens' Listener (Maybe Text)

lPort :: Lens' Listener (Maybe Integer)

lbdLoadBalancerName :: Lens' LoadBalancerDescription (Maybe Text)

lbdListeners :: Lens' LoadBalancerDescription [Listener]

lbdDomain :: Lens' LoadBalancerDescription (Maybe Text)

orrPattern :: Lens' OptionRestrictionRegex (Maybe Text)

orrLabel :: Lens' OptionRestrictionRegex (Maybe Text)

osOptionName :: Lens' OptionSpecification (Maybe Text)

osNamespace :: Lens' OptionSpecification (Maybe Text)

qeURL :: Lens' Queue (Maybe Text)

qeName :: Lens' Queue (Maybe Text)

snS3Key :: Lens' S3Location (Maybe Text)

snS3Bucket :: Lens' S3Location (Maybe Text)

ssdSolutionStackName :: Lens' SolutionStackDescription (Maybe Text)

ssdPermittedFileTypes :: Lens' SolutionStackDescription [Text]

seTemplateName :: Lens' SourceConfiguration (Maybe Text)

seApplicationName :: Lens' SourceConfiguration (Maybe Text)

tgValue :: Lens' Tag (Maybe Text)

tgKey :: Lens' Tag (Maybe Text)

vySeverity :: Lens' ValidationMessage (Maybe ValidationSeverity)

vyOptionName :: Lens' ValidationMessage (Maybe Text)

vyNamespace :: Lens' ValidationMessage (Maybe Text)

vyMessage :: Lens' ValidationMessage (Maybe Text)