Network.AWS.ElasticBeanstalk.V2010_12_01.UpdateApplication

Description

Updates the specified application to have the specified properties. If a property (for example, description) is not provided, the value remains unchanged. To clear these properties, specify an empty string. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &Description=Another%20Description &Operation=UpdateApplication &AuthParams New Version Another Description SampleApp 2010-11-17T19:26:20.410Z 2010-11-17T20:42:54.611Z Default 40be666b-f28b-11df-8a78-9f77047e0d0c.

Request

data UpdateApplication

This documentation target is not reported in the API reference.

Request constructor

mkUpdateApplication

Arguments

:: Text
uaApplicationName

-> UpdateApplication

Smart constructor for the minimum required parameters to construct a valid UpdateApplication request.

Request lenses

uaApplicationName :: Lens' UpdateApplication Text

The name of the application to update. If no such application is found, UpdateApplication returns an InvalidParameterValue error.

uaDescription :: Lens' UpdateApplication (Maybe Text)

A new description for the application. Default: If not specified, AWS Elastic Beanstalk does not update the description.

Response

data UpdateApplicationResponse

Result message containing a single description of an application.

Response lenses

uarsApplication :: Lens' UpdateApplicationResponse (Maybe ApplicationDescription)

The ApplicationDescription of the application.