Network.AWS.ElasticBeanstalk.V2010_12_01.UpdateApplicationVersion

Description

Updates the specified application version to have the specified properties. If a property (for example, description) is not provided, the value remains unchanged. To clear properties, specify an empty string. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &VersionLabel=New%20Version &Description=New%20Release%20Description &Operation=UpdateApplicationVersion &AuthParams awsemr sample.war New Version New Release Description SampleApp 2010-11-17T19:26:20.699Z 2010-11-17T20:48:16.632Z 00b10aa1-f28c-11df-8a78-9f77047e0d0c.

Documentation

updateApplicationVersion

Arguments

:: Text
_uavmApplicationName

-> Text
_uavmVersionLabel

-> UpdateApplicationVersion

Minimum specification for a UpdateApplicationVersion request.

data UpdateApplicationVersion

Constructors

UpdateApplicationVersion

Fields

_uavmApplicationName :: Text

The name of the application associated with this version. If no application is found with this name, UpdateApplication returns an InvalidParameterValue error.

_uavmVersionLabel :: Text

The name of the version to update. If no application version is found with this label, UpdateApplication returns an InvalidParameterValue error.

_uavmDescription :: Maybe Text

A new description for this release.

uavmVersionLabel :: Lens' UpdateApplicationVersion Text

uavmDescription :: Lens' UpdateApplicationVersion (Maybe Text)

uavmApplicationName :: Lens' UpdateApplicationVersion Text

data UpdateApplicationVersionResponse

Constructors

UpdateApplicationVersionResponse

Fields

_avdnApplicationVersion :: Maybe ApplicationVersionDescription

The ApplicationVersionDescription of the application version.

avdnApplicationVersion :: Iso' UpdateApplicationVersionResponse (Maybe ApplicationVersionDescription)