Network.AWS.ElasticBeanstalk.V2010_12_01.UpdateEnvironment

Description

Updates the environment description, deploys a new application version, updates the configuration settings to an entirely new configuration template, or updates select configuration option values in the running environment. Attempting to update both the release and configuration is not allowed and AWS Elastic Beanstalk returns an InvalidParameterCombination error. When updating the configuration settings to a new template or individual settings, a draft configuration is created and DescribeConfigurationSettings for this environment returns two setting descriptions with different DeploymentStatus values. https:/elasticbeanstalk.us-east-1.amazon.com?EnvironmentId=e-hc8mvnayrx &EnvironmentName=SampleAppVersion &TemplateName=default &OptionsToRemove.member.1.Namespace=aws%3Aautoscaling%3Atrigger &OptionsToRemove.member.1.OptionName=MeasureName &Operation=UpdateEnvironment &AuthParams New Version Deploying SampleApp elasticbeanstalk-SampleAppVersion-246126201.us-east-1.elb.amazonaws.com SampleApp.elasticbeanstalk.amazonaws.com Grey e-hc8mvnayrx 2010-11-17T21:05:55.251Z 32bit Amazon Linux running Tomcat 7 SampleAppDescription SampleAppVersion 2010-11-17T20:17:42.339Z 7705f0bc-f28e-11df-8a78-9f77047e0d0c.

Request

data UpdateEnvironment

This documentation target is not reported in the API reference.

Request constructor

mkUpdateEnvironment :: UpdateEnvironment

Smart constructor for the minimum required parameters to construct a valid UpdateEnvironment request.

Request lenses

ueEnvironmentId :: Lens' UpdateEnvironment (Maybe Text)

The ID of the environment to update. If no environment with this ID exists, AWS Elastic Beanstalk returns an InvalidParameterValue error. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

ueEnvironmentName :: Lens' UpdateEnvironment (Maybe Text)

The name of the environment to update. If no environment with this name exists, AWS Elastic Beanstalk returns an InvalidParameterValue error. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

ueDescription :: Lens' UpdateEnvironment (Maybe Text)

If this parameter is specified, AWS Elastic Beanstalk updates the description of this environment.

ueTier :: Lens' UpdateEnvironment (Maybe EnvironmentTier)

This specifies the tier to use to update the environment. Condition: You can only update the tier version for an environment. If you change the name of the type, AWS Elastic Beanstalk returns InvalidParameterValue error.

ueVersionLabel :: Lens' UpdateEnvironment (Maybe Text)

If this parameter is specified, AWS Elastic Beanstalk deploys the named application version to the environment. If no such application version is found, returns an InvalidParameterValue error.

ueTemplateName :: Lens' UpdateEnvironment (Maybe Text)

If this parameter is specified, AWS Elastic Beanstalk deploys this configuration template to the environment. If no such configuration template is found, AWS Elastic Beanstalk returns an InvalidParameterValue error.

ueOptionSettings :: Lens' UpdateEnvironment [ConfigurationOptionSetting]

If specified, AWS Elastic Beanstalk updates the configuration set associated with the running environment and sets the specified configuration options to the requested value.

ueOptionsToRemove :: Lens' UpdateEnvironment [OptionSpecification]

A list of custom user-defined configuration options to remove from the configuration set for this environment.

Response

data UpdateEnvironmentResponse

Describes the properties of an environment.

Response lenses

uersEnvironmentName :: Lens' UpdateEnvironmentResponse (Maybe Text)

The name of this environment.

uersEnvironmentId :: Lens' UpdateEnvironmentResponse (Maybe Text)

The ID of this environment.

uersApplicationName :: Lens' UpdateEnvironmentResponse (Maybe Text)

The name of the application associated with this environment.

uersVersionLabel :: Lens' UpdateEnvironmentResponse (Maybe Text)

The application version deployed in this environment.

uersSolutionStackName :: Lens' UpdateEnvironmentResponse (Maybe Text)

The name of the SolutionStack deployed with this environment.

uersTemplateName :: Lens' UpdateEnvironmentResponse (Maybe Text)

The name of the configuration template used to originally launch this environment.

uersDescription :: Lens' UpdateEnvironmentResponse (Maybe Text)

Describes this environment.

uersEndpointURL :: Lens' UpdateEnvironmentResponse (Maybe Text)

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

uersCNAME :: Lens' UpdateEnvironmentResponse (Maybe Text)

The URL to the CNAME for this environment.

uersDateCreated :: Lens' UpdateEnvironmentResponse (Maybe ISO8601)

The creation date for this environment.

uersDateUpdated :: Lens' UpdateEnvironmentResponse (Maybe ISO8601)

The last modified date for this environment.

uersStatus :: Lens' UpdateEnvironmentResponse (Maybe EnvironmentStatus)

The current operational status of the environment: Launching: Environment is in the process of initial deployment. Updating: Environment is in the process of updating its configuration settings or application version. Ready: Environment is available to have an action performed on it, such as update or terminate. Terminating: Environment is in the shut-down process. Terminated: Environment is not running.

uersHealth :: Lens' UpdateEnvironmentResponse (Maybe EnvironmentHealth)

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment: Red : Indicates the environment is not working. Yellow: Indicates that something is wrong, the application might not be available, but the instances appear running. Green: Indicates the environment is healthy and fully functional. Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. Green: Indicates the environment is healthy and fully functional. Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey.

uersResources :: Lens' UpdateEnvironmentResponse (Maybe EnvironmentResourcesDescription)

The description of the AWS resources used by this environment.

uersTier :: Lens' UpdateEnvironmentResponse (Maybe EnvironmentTier)

Describes the current tier of this environment.