Network.AWS.ElasticBeanstalk.V2010_12_01.ValidateConfigurationSettings

Description

Takes a set of configuration settings and either a configuration template or environment, and determines whether those values are valid. This action returns a list of messages indicating any errors or warnings associated with the selection of option values. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &EnvironmentName=SampleAppVersion &OptionSettings.member.1.Namespace=aws%3Aautoscaling%3Atrigger &OptionSettings.member.1.OptionName=LowerThreshold &OptionSettings.member.1.Value=1000000 &Operation=ValidateConfigurationSettings &AuthParams 06f1cfff-f28f-11df-8a78-9f77047e0d0c.

Request

data ValidateConfigurationSettings

A list of validation messages for a specified configuration template.

Request constructor

mkValidateConfigurationSettings

Arguments

:: Text
vcsApplicationName

-> [ConfigurationOptionSetting]
vcsOptionSettings

-> ValidateConfigurationSettings

Smart constructor for the minimum required parameters to construct a valid ValidateConfigurationSettings request.

Request lenses

vcsApplicationName :: Lens' ValidateConfigurationSettings Text

The name of the application that the configuration template or environment belongs to.

vcsTemplateName :: Lens' ValidateConfigurationSettings (Maybe Text)

The name of the configuration template to validate the settings against. Condition: You cannot specify both this and an environment name.

vcsEnvironmentName :: Lens' ValidateConfigurationSettings (Maybe Text)

The name of the environment to validate the settings against. Condition: You cannot specify both this and a configuration template name.

vcsOptionSettings :: Lens' ValidateConfigurationSettings [ConfigurationOptionSetting]

A list of the options and desired values to evaluate.

Response

data ValidateConfigurationSettingsResponse

Provides a list of validation messages.

Response constructor

mkValidateConfigurationSettingsResponse :: ValidateConfigurationSettingsResponse

Smart constructor for the minimum required parameters to construct a valid ValidateConfigurationSettingsResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

vcsrMessages :: Lens' ValidateConfigurationSettingsResponse [ValidationMessage]

A list of ValidationMessage.