Network.AWS.ElasticBeanstalk.ValidateConfigurationSettings

Description

Takes a set of configuration settings and either a configuration template or environment, and determines whether those values are valid.

This action returns a list of messages indicating any errors or warnings associated with the selection of option values.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_ValidateConfigurationSettings.html

Request

data ValidateConfigurationSettings

Request constructor

validateConfigurationSettings

Arguments

:: Text
vcsApplicationName

-> ValidateConfigurationSettings

ValidateConfigurationSettings constructor.

The fields accessible through corresponding lenses are:

vcsApplicationName :: Text

vcsEnvironmentName :: Maybe Text

vcsOptionSettings :: [ConfigurationOptionSetting]

vcsTemplateName :: Maybe Text

Request lenses

vcsApplicationName :: Lens' ValidateConfigurationSettings Text

The name of the application that the configuration template or environment belongs to.

vcsEnvironmentName :: Lens' ValidateConfigurationSettings (Maybe Text)

The name of the environment to validate the settings against.

Condition: You cannot specify both this and a configuration template name.

vcsOptionSettings :: Lens' ValidateConfigurationSettings [ConfigurationOptionSetting]

A list of the options and desired values to evaluate.

vcsTemplateName :: Lens' ValidateConfigurationSettings (Maybe Text)

The name of the configuration template to validate the settings against.

Condition: You cannot specify both this and an environment name.

Response

data ValidateConfigurationSettingsResponse

Response constructor

validateConfigurationSettingsResponse :: ValidateConfigurationSettingsResponse

ValidateConfigurationSettingsResponse constructor.

The fields accessible through corresponding lenses are:

vcsrMessages :: [ValidationMessage]

Response lenses

vcsrMessages :: Lens' ValidateConfigurationSettingsResponse [ValidationMessage]

A list of ValidationMessage.