Network.AWS.ElasticBeanstalk.CreateApplicationVersion

Description

Creates an application version for the specified application. You can create an application version from a source bundle in Amazon S3, a commit in AWS CodeCommit, or the output of an AWS CodeBuild build as follows:

Specify a commit in an AWS CodeCommit repository with SourceBuildInformation .

Specify a build in an AWS CodeBuild with SourceBuildInformation and BuildConfiguration .

Specify a source bundle in S3 with SourceBundle

Omit both SourceBuildInformation and SourceBundle to use the default sample application.

Creating a Request

createApplicationVersion

Arguments

:: Text
cavApplicationName

-> Text
cavVersionLabel

-> CreateApplicationVersion

Creates a value of CreateApplicationVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cavProcess - Preprocesses and validates the environment manifest and configuration files in the source bundle. Validating configuration files can identify issues prior to deploying the application version to an environment.

cavSourceBundle - The Amazon S3 bucket and key that identify the location of the source bundle for this version. Specify a source bundle in S3 or a commit in an AWS CodeCommit repository (with SourceBuildInformation ), but not both. If neither SourceBundle nor SourceBuildInformation are provided, Elastic Beanstalk uses a sample application.

cavAutoCreateApplication - Set to true to create an application with the specified name if it doesn't already exist.

cavSourceBuildInformation - Specify a commit in an AWS CodeCommit Git repository to use as the source code for the application version.

cavDescription - Describes this version.

cavBuildConfiguration - Settings for an AWS CodeBuild build.

cavApplicationName - The name of the application. If no application is found with this name, and AutoCreateApplication is false , returns an InvalidParameterValue error.

cavVersionLabel - A label identifying this version. Constraint: Must be unique per application. If an application version already exists with this label for the specified application, AWS Elastic Beanstalk returns an InvalidParameterValue error.

data CreateApplicationVersion

See: createApplicationVersion smart constructor.

Request Lenses

cavProcess :: Lens' CreateApplicationVersion (Maybe Bool)

Preprocesses and validates the environment manifest and configuration files in the source bundle. Validating configuration files can identify issues prior to deploying the application version to an environment.

cavSourceBundle :: Lens' CreateApplicationVersion (Maybe S3Location)

The Amazon S3 bucket and key that identify the location of the source bundle for this version. Specify a source bundle in S3 or a commit in an AWS CodeCommit repository (with SourceBuildInformation ), but not both. If neither SourceBundle nor SourceBuildInformation are provided, Elastic Beanstalk uses a sample application.

cavAutoCreateApplication :: Lens' CreateApplicationVersion (Maybe Bool)

Set to true to create an application with the specified name if it doesn't already exist.

cavSourceBuildInformation :: Lens' CreateApplicationVersion (Maybe SourceBuildInformation)

Specify a commit in an AWS CodeCommit Git repository to use as the source code for the application version.

cavDescription :: Lens' CreateApplicationVersion (Maybe Text)

Describes this version.

cavBuildConfiguration :: Lens' CreateApplicationVersion (Maybe BuildConfiguration)

Settings for an AWS CodeBuild build.

cavApplicationName :: Lens' CreateApplicationVersion Text

The name of the application. If no application is found with this name, and AutoCreateApplication is false , returns an InvalidParameterValue error.

cavVersionLabel :: Lens' CreateApplicationVersion Text

A label identifying this version. Constraint: Must be unique per application. If an application version already exists with this label for the specified application, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Destructuring the Response

applicationVersionDescriptionMessage :: ApplicationVersionDescriptionMessage

Creates a value of ApplicationVersionDescriptionMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avdmApplicationVersion - The ApplicationVersionDescription of the application version.

data ApplicationVersionDescriptionMessage

Result message wrapping a single description of an application version.

See: applicationVersionDescriptionMessage smart constructor.

Response Lenses

avdmApplicationVersion :: Lens' ApplicationVersionDescriptionMessage (Maybe ApplicationVersionDescription)

The ApplicationVersionDescription of the application version.