Network.AWS.ElasticBeanstalk.CreateEnvironment

Description

Launches an environment for the specified application using the specified configuration.

Creating a Request

createEnvironment

Arguments

:: Text
cApplicationName

-> CreateEnvironment

Creates a value of CreateEnvironment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cCNAMEPrefix - If specified, the environment attempts to use this value as the prefix for the CNAME. If not specified, the CNAME is generated automatically by appending a random alphanumeric string to the environment name.

cTemplateName - The name of the configuration template to use in deployment. If no configuration template is found with this name, AWS Elastic Beanstalk returns an InvalidParameterValue error.

cOptionsToRemove - A list of custom user-defined configuration options to remove from the configuration set for this new environment.

cOptionSettings - If specified, AWS Elastic Beanstalk sets the specified configuration options to the requested value in the configuration set for the new environment. These override the values obtained from the solution stack or the configuration template.

cVersionLabel - The name of the application version to deploy. If the specified application has no associated application versions, AWS Elastic Beanstalk UpdateEnvironment returns an InvalidParameterValue error. Default: If not specified, AWS Elastic Beanstalk attempts to launch the sample application in the container.

cPlatformARN - The ARN of the platform.

cTier - This specifies the tier to use for creating this environment.

cEnvironmentName - A unique name for the deployment environment. Used in the application URL. Constraint: Must be from 4 to 40 characters in length. The name can contain only letters, numbers, and hyphens. It cannot start or end with a hyphen. This name must be unique within a region in your account. If the specified name already exists in the region, AWS Elastic Beanstalk returns an InvalidParameterValue error. Default: If the CNAME parameter is not specified, the environment name becomes part of the CNAME, and therefore part of the visible URL for your application.

cSolutionStackName - This is an alternative to specifying a template name. If specified, AWS Elastic Beanstalk sets the configuration values to the default values associated with the specified solution stack.

cGroupName - The name of the group to which the target environment belongs. Specify a group name only if the environment's name is specified in an environment manifest and not with the environment name parameter. See Environment Manifest (env.yaml) for details.

cDescription - Describes this environment.

cTags - This specifies the tags applied to resources in the environment.

cApplicationName - The name of the application that contains the version to be deployed. If no application is found with this name, CreateEnvironment returns an InvalidParameterValue error.

data CreateEnvironment

See: createEnvironment smart constructor.

Request Lenses

cCNAMEPrefix :: Lens' CreateEnvironment (Maybe Text)

If specified, the environment attempts to use this value as the prefix for the CNAME. If not specified, the CNAME is generated automatically by appending a random alphanumeric string to the environment name.

cTemplateName :: Lens' CreateEnvironment (Maybe Text)

The name of the configuration template to use in deployment. If no configuration template is found with this name, AWS Elastic Beanstalk returns an InvalidParameterValue error.

cOptionsToRemove :: Lens' CreateEnvironment [OptionSpecification]

A list of custom user-defined configuration options to remove from the configuration set for this new environment.

cOptionSettings :: Lens' CreateEnvironment [ConfigurationOptionSetting]

If specified, AWS Elastic Beanstalk sets the specified configuration options to the requested value in the configuration set for the new environment. These override the values obtained from the solution stack or the configuration template.

cVersionLabel :: Lens' CreateEnvironment (Maybe Text)

The name of the application version to deploy. If the specified application has no associated application versions, AWS Elastic Beanstalk UpdateEnvironment returns an InvalidParameterValue error. Default: If not specified, AWS Elastic Beanstalk attempts to launch the sample application in the container.

cPlatformARN :: Lens' CreateEnvironment (Maybe Text)

The ARN of the platform.

cTier :: Lens' CreateEnvironment (Maybe EnvironmentTier)

This specifies the tier to use for creating this environment.

cEnvironmentName :: Lens' CreateEnvironment (Maybe Text)

A unique name for the deployment environment. Used in the application URL. Constraint: Must be from 4 to 40 characters in length. The name can contain only letters, numbers, and hyphens. It cannot start or end with a hyphen. This name must be unique within a region in your account. If the specified name already exists in the region, AWS Elastic Beanstalk returns an InvalidParameterValue error. Default: If the CNAME parameter is not specified, the environment name becomes part of the CNAME, and therefore part of the visible URL for your application.

cSolutionStackName :: Lens' CreateEnvironment (Maybe Text)

This is an alternative to specifying a template name. If specified, AWS Elastic Beanstalk sets the configuration values to the default values associated with the specified solution stack.

cGroupName :: Lens' CreateEnvironment (Maybe Text)

The name of the group to which the target environment belongs. Specify a group name only if the environment's name is specified in an environment manifest and not with the environment name parameter. See Environment Manifest (env.yaml) for details.

cDescription :: Lens' CreateEnvironment (Maybe Text)

Describes this environment.

cTags :: Lens' CreateEnvironment [Tag]

This specifies the tags applied to resources in the environment.

cApplicationName :: Lens' CreateEnvironment Text

The name of the application that contains the version to be deployed. If no application is found with this name, CreateEnvironment returns an InvalidParameterValue error.

Destructuring the Response

environmentDescription :: EnvironmentDescription

Creates a value of EnvironmentDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eStatus - The current operational status of the environment: * Launching : Environment is in the process of initial deployment. * Updating : Environment is in the process of updating its configuration settings or application version. * Ready : Environment is available to have an action performed on it, such as update or terminate. * Terminating : Environment is in the shut-down process. * Terminated : Environment is not running.

eCNAME - The URL to the CNAME for this environment.

eTemplateName - The name of the configuration template used to originally launch this environment.

eAbortableOperationInProgress - Indicates if there is an in-progress environment configuration update or application version deployment that you can cancel. true: There is an update in progress. false: There are no updates currently in progress.

eEndpointURL - For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

eResources - The description of the AWS resources used by this environment.

eDateUpdated - The last modified date for this environment.

eDateCreated - The creation date for this environment.

eHealth - Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment: * Red : Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. * Yellow : Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. * Green : Indicates the environment is healthy and fully functional. * Grey : Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey

eVersionLabel - The application version deployed in this environment.

ePlatformARN - The ARN of the platform.

eTier - Describes the current tier of this environment.

eEnvironmentName - The name of this environment.

eApplicationName - The name of the application associated with this environment.

eEnvironmentARN - The environment's Amazon Resource Name (ARN), which can be used in other API reuqests that require an ARN.

eSolutionStackName - The name of the SolutionStack deployed with this environment.

eEnvironmentId - The ID of this environment.

eHealthStatus - Returns the health status of the application running in your environment. For more information, see Health Colors and Statuses .

eEnvironmentLinks - A list of links to other environments in the same group.

eDescription - Describes this environment.

data EnvironmentDescription

Describes the properties of an environment.

See: environmentDescription smart constructor.

Response Lenses

eStatus :: Lens' EnvironmentDescription (Maybe EnvironmentStatus)

The current operational status of the environment: * Launching : Environment is in the process of initial deployment. * Updating : Environment is in the process of updating its configuration settings or application version. * Ready : Environment is available to have an action performed on it, such as update or terminate. * Terminating : Environment is in the shut-down process. * Terminated : Environment is not running.

eCNAME :: Lens' EnvironmentDescription (Maybe Text)

The URL to the CNAME for this environment.

eTemplateName :: Lens' EnvironmentDescription (Maybe Text)

The name of the configuration template used to originally launch this environment.

eAbortableOperationInProgress :: Lens' EnvironmentDescription (Maybe Bool)

Indicates if there is an in-progress environment configuration update or application version deployment that you can cancel. true: There is an update in progress. false: There are no updates currently in progress.

eEndpointURL :: Lens' EnvironmentDescription (Maybe Text)

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

eResources :: Lens' EnvironmentDescription (Maybe EnvironmentResourcesDescription)

The description of the AWS resources used by this environment.

eDateUpdated :: Lens' EnvironmentDescription (Maybe UTCTime)

The last modified date for this environment.

eDateCreated :: Lens' EnvironmentDescription (Maybe UTCTime)

The creation date for this environment.

eHealth :: Lens' EnvironmentDescription (Maybe EnvironmentHealth)

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment: * Red : Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. * Yellow : Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. * Green : Indicates the environment is healthy and fully functional. * Grey : Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey

eVersionLabel :: Lens' EnvironmentDescription (Maybe Text)

The application version deployed in this environment.

ePlatformARN :: Lens' EnvironmentDescription (Maybe Text)

The ARN of the platform.

eTier :: Lens' EnvironmentDescription (Maybe EnvironmentTier)

Describes the current tier of this environment.

eEnvironmentName :: Lens' EnvironmentDescription (Maybe Text)

The name of this environment.

eApplicationName :: Lens' EnvironmentDescription (Maybe Text)

The name of the application associated with this environment.

eEnvironmentARN :: Lens' EnvironmentDescription (Maybe Text)

The environment's Amazon Resource Name (ARN), which can be used in other API reuqests that require an ARN.

eSolutionStackName :: Lens' EnvironmentDescription (Maybe Text)

The name of the SolutionStack deployed with this environment.

eEnvironmentId :: Lens' EnvironmentDescription (Maybe Text)

The ID of this environment.

eHealthStatus :: Lens' EnvironmentDescription (Maybe EnvironmentHealthStatus)

Returns the health status of the application running in your environment. For more information, see Health Colors and Statuses .

eEnvironmentLinks :: Lens' EnvironmentDescription [EnvironmentLink]

A list of links to other environments in the same group.

eDescription :: Lens' EnvironmentDescription (Maybe Text)

Describes this environment.