Network.AWS.ElasticBeanstalk.DeleteConfigurationTemplate

Description

Deletes the specified configuration template.

Creating a Request

deleteConfigurationTemplate

Arguments

:: Text
dctApplicationName

-> Text
dctTemplateName

-> DeleteConfigurationTemplate

Creates a value of DeleteConfigurationTemplate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dctApplicationName - The name of the application to delete the configuration template from.

dctTemplateName - The name of the configuration template to delete.

data DeleteConfigurationTemplate

Request to delete a configuration template.

See: deleteConfigurationTemplate smart constructor.

Request Lenses

dctApplicationName :: Lens' DeleteConfigurationTemplate Text

The name of the application to delete the configuration template from.

dctTemplateName :: Lens' DeleteConfigurationTemplate Text

The name of the configuration template to delete.

Destructuring the Response

deleteConfigurationTemplateResponse :: DeleteConfigurationTemplateResponse

Creates a value of DeleteConfigurationTemplateResponse with the minimum fields required to make a request.

data DeleteConfigurationTemplateResponse

See: deleteConfigurationTemplateResponse smart constructor.