Network.AWS.ElasticBeanstalk.DeleteEnvironmentConfiguration

Description

Deletes the draft configuration associated with the running environment.

Updating a running environment with any configuration changes creates a draft configuration set. You can get the draft configuration using DescribeConfigurationSettings while the update is in progress or if the update fails. The DeploymentStatus for the draft configuration indicates whether the deployment is in process or has failed. The draft configuration remains in existence until it is deleted with this action.

Creating a Request

deleteEnvironmentConfiguration

Arguments

:: Text
decApplicationName

-> Text
decEnvironmentName

-> DeleteEnvironmentConfiguration

Creates a value of DeleteEnvironmentConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

decApplicationName - The name of the application the environment is associated with.

decEnvironmentName - The name of the environment to delete the draft configuration from.

data DeleteEnvironmentConfiguration

Request to delete a draft environment configuration.

See: deleteEnvironmentConfiguration smart constructor.

Request Lenses

decApplicationName :: Lens' DeleteEnvironmentConfiguration Text

The name of the application the environment is associated with.

decEnvironmentName :: Lens' DeleteEnvironmentConfiguration Text

The name of the environment to delete the draft configuration from.

Destructuring the Response

deleteEnvironmentConfigurationResponse :: DeleteEnvironmentConfigurationResponse

Creates a value of DeleteEnvironmentConfigurationResponse with the minimum fields required to make a request.

data DeleteEnvironmentConfigurationResponse

See: deleteEnvironmentConfigurationResponse smart constructor.