Network.AWS.ElasticBeanstalk.DeleteEnvironmentConfiguration

Description

Deletes the draft configuration associated with the running environment.

Updating a running environment with any configuration changes creates a draft configuration set. You can get the draft configuration using DescribeConfigurationSettings while the update is in progress or if the update fails. The DeploymentStatus for the draft configuration indicates whether the deployment is in process or has failed. The draft configuration remains in existence until it is deleted with this action.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_DeleteEnvironmentConfiguration.html

Request

data DeleteEnvironmentConfiguration

Request constructor

deleteEnvironmentConfiguration

Arguments

:: Text
decApplicationName

-> Text
decEnvironmentName

-> DeleteEnvironmentConfiguration

DeleteEnvironmentConfiguration constructor.

The fields accessible through corresponding lenses are:

decApplicationName :: Text

decEnvironmentName :: Text

Request lenses

decApplicationName :: Lens' DeleteEnvironmentConfiguration Text

The name of the application the environment is associated with.

decEnvironmentName :: Lens' DeleteEnvironmentConfiguration Text

The name of the environment to delete the draft configuration from.

Response

data DeleteEnvironmentConfigurationResponse

Response constructor

deleteEnvironmentConfigurationResponse :: DeleteEnvironmentConfigurationResponse

DeleteEnvironmentConfigurationResponse constructor.