Network.AWS.ElasticBeanstalk.DescribeConfigurationOptions

Description

Describes the configuration options that are used in a particular configuration template or environment, or that a specified solution stack defines. The description includes the values the options, their default values, and an indication of the required action on a running environment if an option value is changed.

Creating a Request

describeConfigurationOptions :: DescribeConfigurationOptions

Creates a value of DescribeConfigurationOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcoTemplateName - The name of the configuration template whose configuration options you want to describe.

dcoPlatformARN - The ARN of the custom platform.

dcoEnvironmentName - The name of the environment whose configuration options you want to describe.

dcoApplicationName - The name of the application associated with the configuration template or environment. Only needed if you want to describe the configuration options associated with either the configuration template or environment.

dcoSolutionStackName - The name of the solution stack whose configuration options you want to describe.

dcoOptions - If specified, restricts the descriptions to only the specified options.

data DescribeConfigurationOptions

Result message containing a list of application version descriptions.

See: describeConfigurationOptions smart constructor.

Request Lenses

dcoTemplateName :: Lens' DescribeConfigurationOptions (Maybe Text)

The name of the configuration template whose configuration options you want to describe.

dcoPlatformARN :: Lens' DescribeConfigurationOptions (Maybe Text)

The ARN of the custom platform.

dcoEnvironmentName :: Lens' DescribeConfigurationOptions (Maybe Text)

The name of the environment whose configuration options you want to describe.

dcoApplicationName :: Lens' DescribeConfigurationOptions (Maybe Text)

The name of the application associated with the configuration template or environment. Only needed if you want to describe the configuration options associated with either the configuration template or environment.

dcoSolutionStackName :: Lens' DescribeConfigurationOptions (Maybe Text)

The name of the solution stack whose configuration options you want to describe.

dcoOptions :: Lens' DescribeConfigurationOptions [OptionSpecification]

If specified, restricts the descriptions to only the specified options.

Destructuring the Response

describeConfigurationOptionsResponse

Arguments

:: Int
dcorsResponseStatus

-> DescribeConfigurationOptionsResponse

Creates a value of DescribeConfigurationOptionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcorsPlatformARN - The ARN of the platform.

dcorsSolutionStackName - The name of the solution stack these configuration options belong to.

dcorsOptions - A list of ConfigurationOptionDescription .

dcorsResponseStatus - -- | The response status code.

data DescribeConfigurationOptionsResponse

Describes the settings for a specified configuration set.

See: describeConfigurationOptionsResponse smart constructor.

Response Lenses

dcorsPlatformARN :: Lens' DescribeConfigurationOptionsResponse (Maybe Text)

The ARN of the platform.

dcorsSolutionStackName :: Lens' DescribeConfigurationOptionsResponse (Maybe Text)

The name of the solution stack these configuration options belong to.

dcorsOptions :: Lens' DescribeConfigurationOptionsResponse [ConfigurationOptionDescription]

A list of ConfigurationOptionDescription .

dcorsResponseStatus :: Lens' DescribeConfigurationOptionsResponse Int

- | The response status code.