Network.AWS.ElasticBeanstalk.DescribeConfigurationSettings

Description

Returns a description of the settings for the specified configuration set, that is, either a configuration template or the configuration set associated with a running environment.

When describing the settings for the configuration set associated with a running environment, it is possible to receive two sets of setting descriptions. One is the deployed configuration set, and the other is a draft configuration of an environment that is either in the process of deployment or that failed to deploy.

Related Topics

DeleteEnvironmentConfiguration

Creating a Request

describeConfigurationSettings

Arguments

:: Text
dcsApplicationName

-> DescribeConfigurationSettings

Creates a value of DescribeConfigurationSettings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsTemplateName - The name of the configuration template to describe. Conditional: You must specify either this parameter or an EnvironmentName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns a MissingRequiredParameter error.

dcsEnvironmentName - The name of the environment to describe. Condition: You must specify either this or a TemplateName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

dcsApplicationName - The application for the environment or configuration template.

data DescribeConfigurationSettings

Result message containing all of the configuration settings for a specified solution stack or configuration template.

See: describeConfigurationSettings smart constructor.

Request Lenses

dcsTemplateName :: Lens' DescribeConfigurationSettings (Maybe Text)

The name of the configuration template to describe. Conditional: You must specify either this parameter or an EnvironmentName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns a MissingRequiredParameter error.

dcsEnvironmentName :: Lens' DescribeConfigurationSettings (Maybe Text)

The name of the environment to describe. Condition: You must specify either this or a TemplateName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

dcsApplicationName :: Lens' DescribeConfigurationSettings Text

The application for the environment or configuration template.

Destructuring the Response

describeConfigurationSettingsResponse

Arguments

:: Int
dcsrsResponseStatus

-> DescribeConfigurationSettingsResponse

Creates a value of DescribeConfigurationSettingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsrsConfigurationSettings - A list of ConfigurationSettingsDescription .

dcsrsResponseStatus - -- | The response status code.

data DescribeConfigurationSettingsResponse

The results from a request to change the configuration settings of an environment.

See: describeConfigurationSettingsResponse smart constructor.

Response Lenses

dcsrsConfigurationSettings :: Lens' DescribeConfigurationSettingsResponse [ConfigurationSettingsDescription]

A list of ConfigurationSettingsDescription .

dcsrsResponseStatus :: Lens' DescribeConfigurationSettingsResponse Int

- | The response status code.