Network.AWS.ElasticBeanstalk.DescribeEnvironments

Description

Returns descriptions for existing environments.

Creating a Request

describeEnvironments :: DescribeEnvironments

Creates a value of DescribeEnvironments with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desEnvironmentIds - If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified IDs.

desEnvironmentNames - If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified names.

desNextToken - For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request. If no NextToken is specified, the first page is retrieved.

desVersionLabel - If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application version.

desMaxRecords - For a paginated request. Specify a maximum number of environments to include in each response. If no MaxRecords is specified, all available environments are retrieved in a single response.

desApplicationName - If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application.

desIncludedDeletedBackTo - If specified when IncludeDeleted is set to true , then environments deleted after this date are displayed.

desIncludeDeleted - Indicates whether to include deleted environments: true : Environments that have been deleted after IncludedDeletedBackTo are displayed. false : Do not include deleted environments.

data DescribeEnvironments

Request to describe one or more environments.

See: describeEnvironments smart constructor.

Request Lenses

desEnvironmentIds :: Lens' DescribeEnvironments [Text]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified IDs.

desEnvironmentNames :: Lens' DescribeEnvironments [Text]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified names.

desNextToken :: Lens' DescribeEnvironments (Maybe Text)

For a paginated request. Specify a token from a previous response page to retrieve the next response page. All other parameter values must be identical to the ones specified in the initial request. If no NextToken is specified, the first page is retrieved.

desVersionLabel :: Lens' DescribeEnvironments (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application version.

desMaxRecords :: Lens' DescribeEnvironments (Maybe Natural)

For a paginated request. Specify a maximum number of environments to include in each response. If no MaxRecords is specified, all available environments are retrieved in a single response.

desApplicationName :: Lens' DescribeEnvironments (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application.

desIncludedDeletedBackTo :: Lens' DescribeEnvironments (Maybe UTCTime)

If specified when IncludeDeleted is set to true , then environments deleted after this date are displayed.

desIncludeDeleted :: Lens' DescribeEnvironments (Maybe Bool)

Indicates whether to include deleted environments: true : Environments that have been deleted after IncludedDeletedBackTo are displayed. false : Do not include deleted environments.

Destructuring the Response

environmentDescriptionsMessage :: EnvironmentDescriptionsMessage

Creates a value of EnvironmentDescriptionsMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

edmNextToken - In a paginated request, the token that you can pass in a subsequent request to get the next response page.

edmEnvironments - Returns an EnvironmentDescription list.

data EnvironmentDescriptionsMessage

Result message containing a list of environment descriptions.

See: environmentDescriptionsMessage smart constructor.

Response Lenses

edmNextToken :: Lens' EnvironmentDescriptionsMessage (Maybe Text)

In a paginated request, the token that you can pass in a subsequent request to get the next response page.

edmEnvironments :: Lens' EnvironmentDescriptionsMessage [EnvironmentDescription]

Returns an EnvironmentDescription list.