Network.AWS.ElasticBeanstalk.DescribeEvents

Description

Returns list of event descriptions matching criteria up to the last 6 weeks.

This operation returns paginated results.

Creating a Request

describeEvents :: DescribeEvents

Creates a value of DescribeEvents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deRequestId - If specified, AWS Elastic Beanstalk restricts the described events to include only those associated with this request ID.

deTemplateName - If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that are associated with this environment configuration.

deStartTime - If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur on or after this time.

deSeverity - If specified, limits the events returned from this call to include only those with the specified severity or higher.

deNextToken - Pagination token. If specified, the events return the next batch of results.

deVersionLabel - If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this application version.

dePlatformARN - The ARN of the version of the custom platform.

deEnvironmentName - If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

deMaxRecords - Specifies the maximum number of events that can be returned, beginning with the most recent event.

deEndTime - If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur up to, but not including, the EndTime .

deApplicationName - If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those associated with this application.

deEnvironmentId - If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

data DescribeEvents

Request to retrieve a list of events for an environment.

See: describeEvents smart constructor.

Request Lenses

deRequestId :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the described events to include only those associated with this request ID.

deTemplateName :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that are associated with this environment configuration.

deStartTime :: Lens' DescribeEvents (Maybe UTCTime)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur on or after this time.

deSeverity :: Lens' DescribeEvents (Maybe EventSeverity)

If specified, limits the events returned from this call to include only those with the specified severity or higher.

deNextToken :: Lens' DescribeEvents (Maybe Text)

Pagination token. If specified, the events return the next batch of results.

deVersionLabel :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this application version.

dePlatformARN :: Lens' DescribeEvents (Maybe Text)

The ARN of the version of the custom platform.

deEnvironmentName :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

deMaxRecords :: Lens' DescribeEvents (Maybe Natural)

Specifies the maximum number of events that can be returned, beginning with the most recent event.

deEndTime :: Lens' DescribeEvents (Maybe UTCTime)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur up to, but not including, the EndTime .

deApplicationName :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those associated with this application.

deEnvironmentId :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

Destructuring the Response

describeEventsResponse

Arguments

:: Int
dersResponseStatus

-> DescribeEventsResponse

Creates a value of DescribeEventsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dersNextToken - If returned, this indicates that there are more results to obtain. Use this token in the next DescribeEvents call to get the next batch of events.

dersEvents - A list of EventDescription .

dersResponseStatus - -- | The response status code.

data DescribeEventsResponse

Result message wrapping a list of event descriptions.

See: describeEventsResponse smart constructor.

Response Lenses

dersNextToken :: Lens' DescribeEventsResponse (Maybe Text)

If returned, this indicates that there are more results to obtain. Use this token in the next DescribeEvents call to get the next batch of events.

dersEvents :: Lens' DescribeEventsResponse [EventDescription]

A list of EventDescription .

dersResponseStatus :: Lens' DescribeEventsResponse Int

- | The response status code.