Network.AWS.ElasticBeanstalk.DescribeEvents

Description

Returns list of event descriptions matching criteria up to the last 6 weeks.

This action returns the most recent 1,000 events from the specified NextToken.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_DescribeEvents.html

Request

data DescribeEvents

Request constructor

describeEvents :: DescribeEvents

DescribeEvents constructor.

The fields accessible through corresponding lenses are:

deApplicationName :: Maybe Text

deEndTime :: Maybe UTCTime

deEnvironmentId :: Maybe Text

deEnvironmentName :: Maybe Text

deMaxRecords :: Maybe Natural

deNextToken :: Maybe Text

deRequestId :: Maybe Text

deSeverity :: Maybe EventSeverity

deStartTime :: Maybe UTCTime

deTemplateName :: Maybe Text

deVersionLabel :: Maybe Text

Request lenses

deApplicationName :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those associated with this application.

deEndTime :: Lens' DescribeEvents (Maybe UTCTime)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur up to, but not including, the EndTime.

deEnvironmentId :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

deEnvironmentName :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

deMaxRecords :: Lens' DescribeEvents (Maybe Natural)

Specifies the maximum number of events that can be returned, beginning with the most recent event.

deNextToken :: Lens' DescribeEvents (Maybe Text)

Pagination token. If specified, the events return the next batch of results.

deRequestId :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the described events to include only those associated with this request ID.

deSeverity :: Lens' DescribeEvents (Maybe EventSeverity)

If specified, limits the events returned from this call to include only those with the specified severity or higher.

deStartTime :: Lens' DescribeEvents (Maybe UTCTime)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur on or after this time.

deTemplateName :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that are associated with this environment configuration.

deVersionLabel :: Lens' DescribeEvents (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this application version.

Response

data DescribeEventsResponse

Response constructor

describeEventsResponse :: DescribeEventsResponse

DescribeEventsResponse constructor.

The fields accessible through corresponding lenses are:

derEvents :: [EventDescription]

derNextToken :: Maybe Text

Response lenses

derEvents :: Lens' DescribeEventsResponse [EventDescription]

A list of EventDescription.

derNextToken :: Lens' DescribeEventsResponse (Maybe Text)

If returned, this indicates that there are more results to obtain. Use this token in the next DescribeEvents call to get the next batch of events.