Network.AWS.ElasticBeanstalk.RestartAppServer

Description

Causes the environment to restart the application container server running on each Amazon EC2 instance.

Creating a Request

restartAppServer :: RestartAppServer

Creates a value of RestartAppServer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rasEnvironmentName - The name of the environment to restart the server for. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

rasEnvironmentId - The ID of the environment to restart the server for. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

data RestartAppServer

See: restartAppServer smart constructor.

Request Lenses

rasEnvironmentName :: Lens' RestartAppServer (Maybe Text)

The name of the environment to restart the server for. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

rasEnvironmentId :: Lens' RestartAppServer (Maybe Text)

The ID of the environment to restart the server for. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

Destructuring the Response

restartAppServerResponse :: RestartAppServerResponse

Creates a value of RestartAppServerResponse with the minimum fields required to make a request.

data RestartAppServerResponse

See: restartAppServerResponse smart constructor.