Network.AWS.ElasticBeanstalk.RestartAppServer

Description

Causes the environment to restart the application container server running on each Amazon EC2 instance.

http://docs.aws.amazon.com/elasticbeanstalk/latest/api/API_RestartAppServer.html

Request

data RestartAppServer

Request constructor

restartAppServer :: RestartAppServer

RestartAppServer constructor.

The fields accessible through corresponding lenses are:

rasEnvironmentId :: Maybe Text

rasEnvironmentName :: Maybe Text

Request lenses

rasEnvironmentId :: Lens' RestartAppServer (Maybe Text)

The ID of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

rasEnvironmentName :: Lens' RestartAppServer (Maybe Text)

The name of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

Response

data RestartAppServerResponse

Response constructor

restartAppServerResponse :: RestartAppServerResponse

RestartAppServerResponse constructor.