Network.AWS.ElasticBeanstalk.TerminateEnvironment

Description

Terminates the specified environment.

Creating a Request

terminateEnvironment :: TerminateEnvironment

Creates a value of TerminateEnvironment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

teForceTerminate - Terminates the target environment even if another environment in the same group is dependent on it.

teTerminateResources - Indicates whether the associated AWS resources should shut down when the environment is terminated: * true : The specified environment as well as the associated AWS resources, such as Auto Scaling group and LoadBalancer, are terminated. * false : AWS Elastic Beanstalk resource management is removed from the environment, but the AWS resources continue to operate. For more information, see the AWS Elastic Beanstalk User Guide. Default: true Valid Values: true | false

teEnvironmentName - The name of the environment to terminate. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

teEnvironmentId - The ID of the environment to terminate. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

data TerminateEnvironment

Request to terminate an environment.

See: terminateEnvironment smart constructor.

Request Lenses

teForceTerminate :: Lens' TerminateEnvironment (Maybe Bool)

Terminates the target environment even if another environment in the same group is dependent on it.

teTerminateResources :: Lens' TerminateEnvironment (Maybe Bool)

Indicates whether the associated AWS resources should shut down when the environment is terminated: * true : The specified environment as well as the associated AWS resources, such as Auto Scaling group and LoadBalancer, are terminated. * false : AWS Elastic Beanstalk resource management is removed from the environment, but the AWS resources continue to operate. For more information, see the AWS Elastic Beanstalk User Guide. Default: true Valid Values: true | false

teEnvironmentName :: Lens' TerminateEnvironment (Maybe Text)

The name of the environment to terminate. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

teEnvironmentId :: Lens' TerminateEnvironment (Maybe Text)

The ID of the environment to terminate. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

Destructuring the Response

environmentDescription :: EnvironmentDescription

Creates a value of EnvironmentDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eStatus - The current operational status of the environment: * Launching : Environment is in the process of initial deployment. * Updating : Environment is in the process of updating its configuration settings or application version. * Ready : Environment is available to have an action performed on it, such as update or terminate. * Terminating : Environment is in the shut-down process. * Terminated : Environment is not running.

eCNAME - The URL to the CNAME for this environment.

eTemplateName - The name of the configuration template used to originally launch this environment.

eAbortableOperationInProgress - Indicates if there is an in-progress environment configuration update or application version deployment that you can cancel. true: There is an update in progress. false: There are no updates currently in progress.

eEndpointURL - For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

eResources - The description of the AWS resources used by this environment.

eDateUpdated - The last modified date for this environment.

eDateCreated - The creation date for this environment.

eHealth - Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment: * Red : Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. * Yellow : Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. * Green : Indicates the environment is healthy and fully functional. * Grey : Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey

eVersionLabel - The application version deployed in this environment.

ePlatformARN - The ARN of the platform.

eTier - Describes the current tier of this environment.

eEnvironmentName - The name of this environment.

eApplicationName - The name of the application associated with this environment.

eEnvironmentARN - The environment's Amazon Resource Name (ARN), which can be used in other API reuqests that require an ARN.

eSolutionStackName - The name of the SolutionStack deployed with this environment.

eEnvironmentId - The ID of this environment.

eHealthStatus - Returns the health status of the application running in your environment. For more information, see Health Colors and Statuses .

eEnvironmentLinks - A list of links to other environments in the same group.

eDescription - Describes this environment.

data EnvironmentDescription

Describes the properties of an environment.

See: environmentDescription smart constructor.

Response Lenses

eStatus :: Lens' EnvironmentDescription (Maybe EnvironmentStatus)

The current operational status of the environment: * Launching : Environment is in the process of initial deployment. * Updating : Environment is in the process of updating its configuration settings or application version. * Ready : Environment is available to have an action performed on it, such as update or terminate. * Terminating : Environment is in the shut-down process. * Terminated : Environment is not running.

eCNAME :: Lens' EnvironmentDescription (Maybe Text)

The URL to the CNAME for this environment.

eTemplateName :: Lens' EnvironmentDescription (Maybe Text)

The name of the configuration template used to originally launch this environment.

eAbortableOperationInProgress :: Lens' EnvironmentDescription (Maybe Bool)

Indicates if there is an in-progress environment configuration update or application version deployment that you can cancel. true: There is an update in progress. false: There are no updates currently in progress.

eEndpointURL :: Lens' EnvironmentDescription (Maybe Text)

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

eResources :: Lens' EnvironmentDescription (Maybe EnvironmentResourcesDescription)

The description of the AWS resources used by this environment.

eDateUpdated :: Lens' EnvironmentDescription (Maybe UTCTime)

The last modified date for this environment.

eDateCreated :: Lens' EnvironmentDescription (Maybe UTCTime)

The creation date for this environment.

eHealth :: Lens' EnvironmentDescription (Maybe EnvironmentHealth)

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment: * Red : Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. * Yellow : Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. * Green : Indicates the environment is healthy and fully functional. * Grey : Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey

eVersionLabel :: Lens' EnvironmentDescription (Maybe Text)

The application version deployed in this environment.

ePlatformARN :: Lens' EnvironmentDescription (Maybe Text)

The ARN of the platform.

eTier :: Lens' EnvironmentDescription (Maybe EnvironmentTier)

Describes the current tier of this environment.

eEnvironmentName :: Lens' EnvironmentDescription (Maybe Text)

The name of this environment.

eApplicationName :: Lens' EnvironmentDescription (Maybe Text)

The name of the application associated with this environment.

eEnvironmentARN :: Lens' EnvironmentDescription (Maybe Text)

The environment's Amazon Resource Name (ARN), which can be used in other API reuqests that require an ARN.

eSolutionStackName :: Lens' EnvironmentDescription (Maybe Text)

The name of the SolutionStack deployed with this environment.

eEnvironmentId :: Lens' EnvironmentDescription (Maybe Text)

The ID of this environment.

eHealthStatus :: Lens' EnvironmentDescription (Maybe EnvironmentHealthStatus)

Returns the health status of the application running in your environment. For more information, see Health Colors and Statuses .

eEnvironmentLinks :: Lens' EnvironmentDescription [EnvironmentLink]

A list of links to other environments in the same group.

eDescription :: Lens' EnvironmentDescription (Maybe Text)

Describes this environment.