Network.AWS.ElasticBeanstalk.UpdateApplicationVersion

Description

Updates the specified application version to have the specified properties.

Creating a Request

updateApplicationVersion

Arguments

:: Text
uavApplicationName

-> Text
uavVersionLabel

-> UpdateApplicationVersion

Creates a value of UpdateApplicationVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uavDescription - A new description for this version.

uavApplicationName - The name of the application associated with this version. If no application is found with this name, UpdateApplication returns an InvalidParameterValue error.

uavVersionLabel - The name of the version to update. If no application version is found with this label, UpdateApplication returns an InvalidParameterValue error.

data UpdateApplicationVersion

See: updateApplicationVersion smart constructor.

Request Lenses

uavDescription :: Lens' UpdateApplicationVersion (Maybe Text)

A new description for this version.

uavApplicationName :: Lens' UpdateApplicationVersion Text

The name of the application associated with this version. If no application is found with this name, UpdateApplication returns an InvalidParameterValue error.

uavVersionLabel :: Lens' UpdateApplicationVersion Text

The name of the version to update. If no application version is found with this label, UpdateApplication returns an InvalidParameterValue error.

Destructuring the Response

applicationVersionDescriptionMessage :: ApplicationVersionDescriptionMessage

Creates a value of ApplicationVersionDescriptionMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avdmApplicationVersion - The ApplicationVersionDescription of the application version.

data ApplicationVersionDescriptionMessage

Result message wrapping a single description of an application version.

See: applicationVersionDescriptionMessage smart constructor.

Response Lenses

avdmApplicationVersion :: Lens' ApplicationVersionDescriptionMessage (Maybe ApplicationVersionDescription)

The ApplicationVersionDescription of the application version.