Network.AWS.ElasticBeanstalk.UpdateConfigurationTemplate

Description

Updates the specified configuration template to have the specified properties or configuration option values.

Related Topics

DescribeConfigurationOptions

Creating a Request

updateConfigurationTemplate

Arguments

:: Text
uctApplicationName

-> Text
uctTemplateName

-> UpdateConfigurationTemplate

Creates a value of UpdateConfigurationTemplate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uctOptionsToRemove - A list of configuration options to remove from the configuration set. Constraint: You can remove only UserDefined configuration options.

uctOptionSettings - A list of configuration option settings to update with the new specified option value.

uctDescription - A new description for the configuration.

uctApplicationName - The name of the application associated with the configuration template to update. If no application is found with this name, UpdateConfigurationTemplate returns an InvalidParameterValue error.

uctTemplateName - The name of the configuration template to update. If no configuration template is found with this name, UpdateConfigurationTemplate returns an InvalidParameterValue error.

data UpdateConfigurationTemplate

The result message containing the options for the specified solution stack.

See: updateConfigurationTemplate smart constructor.

Request Lenses

uctOptionsToRemove :: Lens' UpdateConfigurationTemplate [OptionSpecification]

A list of configuration options to remove from the configuration set. Constraint: You can remove only UserDefined configuration options.

uctOptionSettings :: Lens' UpdateConfigurationTemplate [ConfigurationOptionSetting]

A list of configuration option settings to update with the new specified option value.

uctDescription :: Lens' UpdateConfigurationTemplate (Maybe Text)

A new description for the configuration.

uctApplicationName :: Lens' UpdateConfigurationTemplate Text

The name of the application associated with the configuration template to update. If no application is found with this name, UpdateConfigurationTemplate returns an InvalidParameterValue error.

uctTemplateName :: Lens' UpdateConfigurationTemplate Text

The name of the configuration template to update. If no configuration template is found with this name, UpdateConfigurationTemplate returns an InvalidParameterValue error.

Destructuring the Response

configurationSettingsDescription :: ConfigurationSettingsDescription

Creates a value of ConfigurationSettingsDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csdTemplateName - If not null , the name of the configuration template for this configuration set.

csdOptionSettings - A list of the configuration options and their values in this configuration set.

csdDateUpdated - The date (in UTC time) when this configuration set was last modified.

csdDateCreated - The date (in UTC time) when this configuration set was created.

csdPlatformARN - The ARN of the platform.

csdEnvironmentName - If not null , the name of the environment for this configuration set.

csdApplicationName - The name of the application associated with this configuration set.

csdDeploymentStatus - If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set: * null : This configuration is not associated with a running environment. * pending : This is a draft configuration that is not deployed to the associated environment but is in the process of deploying. * deployed : This is the configuration that is currently deployed to the associated running environment. * failed : This is a draft configuration that failed to successfully deploy.

csdSolutionStackName - The name of the solution stack this configuration set uses.

csdDescription - Describes this configuration set.

data ConfigurationSettingsDescription

Describes the settings for a configuration set.

See: configurationSettingsDescription smart constructor.

Response Lenses

csdTemplateName :: Lens' ConfigurationSettingsDescription (Maybe Text)

If not null , the name of the configuration template for this configuration set.

csdOptionSettings :: Lens' ConfigurationSettingsDescription [ConfigurationOptionSetting]

A list of the configuration options and their values in this configuration set.

csdDateUpdated :: Lens' ConfigurationSettingsDescription (Maybe UTCTime)

The date (in UTC time) when this configuration set was last modified.

csdDateCreated :: Lens' ConfigurationSettingsDescription (Maybe UTCTime)

The date (in UTC time) when this configuration set was created.

csdPlatformARN :: Lens' ConfigurationSettingsDescription (Maybe Text)

The ARN of the platform.

csdEnvironmentName :: Lens' ConfigurationSettingsDescription (Maybe Text)

If not null , the name of the environment for this configuration set.

csdApplicationName :: Lens' ConfigurationSettingsDescription (Maybe Text)

The name of the application associated with this configuration set.

csdDeploymentStatus :: Lens' ConfigurationSettingsDescription (Maybe ConfigurationDeploymentStatus)

If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set: * null : This configuration is not associated with a running environment. * pending : This is a draft configuration that is not deployed to the associated environment but is in the process of deploying. * deployed : This is the configuration that is currently deployed to the associated running environment. * failed : This is a draft configuration that failed to successfully deploy.

csdSolutionStackName :: Lens' ConfigurationSettingsDescription (Maybe Text)

The name of the solution stack this configuration set uses.

csdDescription :: Lens' ConfigurationSettingsDescription (Maybe Text)

Describes this configuration set.