Network.AWS.ElasticBeanstalk.UpdateEnvironment

Description

Updates the environment description, deploys a new application version, updates the configuration settings to an entirely new configuration template, or updates select configuration option values in the running environment.

Attempting to update both the release and configuration is not allowed and AWS Elastic Beanstalk returns an InvalidParameterCombination error.

When updating the configuration settings to a new template or individual settings, a draft configuration is created and DescribeConfigurationSettings for this environment returns two setting descriptions with different DeploymentStatus values.

Creating a Request

updateEnvironment :: UpdateEnvironment

Creates a value of UpdateEnvironment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ueTemplateName - If this parameter is specified, AWS Elastic Beanstalk deploys this configuration template to the environment. If no such configuration template is found, AWS Elastic Beanstalk returns an InvalidParameterValue error.

ueOptionsToRemove - A list of custom user-defined configuration options to remove from the configuration set for this environment.

ueOptionSettings - If specified, AWS Elastic Beanstalk updates the configuration set associated with the running environment and sets the specified configuration options to the requested value.

ueVersionLabel - If this parameter is specified, AWS Elastic Beanstalk deploys the named application version to the environment. If no such application version is found, returns an InvalidParameterValue error.

uePlatformARN - The ARN of the platform, if used.

ueTier - This specifies the tier to use to update the environment. Condition: At this time, if you change the tier version, name, or type, AWS Elastic Beanstalk returns InvalidParameterValue error.

ueEnvironmentName - The name of the environment to update. If no environment with this name exists, AWS Elastic Beanstalk returns an InvalidParameterValue error. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

ueApplicationName - The name of the application with which the environment is associated.

ueSolutionStackName - This specifies the platform version that the environment will run after the environment is updated.

ueEnvironmentId - The ID of the environment to update. If no environment with this ID exists, AWS Elastic Beanstalk returns an InvalidParameterValue error. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

ueGroupName - The name of the group to which the target environment belongs. Specify a group name only if the environment's name is specified in an environment manifest and not with the environment name or environment ID parameters. See Environment Manifest (env.yaml) for details.

ueDescription - If this parameter is specified, AWS Elastic Beanstalk updates the description of this environment.

data UpdateEnvironment

Request to update an environment.

See: updateEnvironment smart constructor.

Request Lenses

ueTemplateName :: Lens' UpdateEnvironment (Maybe Text)

If this parameter is specified, AWS Elastic Beanstalk deploys this configuration template to the environment. If no such configuration template is found, AWS Elastic Beanstalk returns an InvalidParameterValue error.

ueOptionsToRemove :: Lens' UpdateEnvironment [OptionSpecification]

A list of custom user-defined configuration options to remove from the configuration set for this environment.

ueOptionSettings :: Lens' UpdateEnvironment [ConfigurationOptionSetting]

If specified, AWS Elastic Beanstalk updates the configuration set associated with the running environment and sets the specified configuration options to the requested value.

ueVersionLabel :: Lens' UpdateEnvironment (Maybe Text)

If this parameter is specified, AWS Elastic Beanstalk deploys the named application version to the environment. If no such application version is found, returns an InvalidParameterValue error.

uePlatformARN :: Lens' UpdateEnvironment (Maybe Text)

The ARN of the platform, if used.

ueTier :: Lens' UpdateEnvironment (Maybe EnvironmentTier)

This specifies the tier to use to update the environment. Condition: At this time, if you change the tier version, name, or type, AWS Elastic Beanstalk returns InvalidParameterValue error.

ueEnvironmentName :: Lens' UpdateEnvironment (Maybe Text)

The name of the environment to update. If no environment with this name exists, AWS Elastic Beanstalk returns an InvalidParameterValue error. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

ueApplicationName :: Lens' UpdateEnvironment (Maybe Text)

The name of the application with which the environment is associated.

ueSolutionStackName :: Lens' UpdateEnvironment (Maybe Text)

This specifies the platform version that the environment will run after the environment is updated.

ueEnvironmentId :: Lens' UpdateEnvironment (Maybe Text)

The ID of the environment to update. If no environment with this ID exists, AWS Elastic Beanstalk returns an InvalidParameterValue error. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

ueGroupName :: Lens' UpdateEnvironment (Maybe Text)

The name of the group to which the target environment belongs. Specify a group name only if the environment's name is specified in an environment manifest and not with the environment name or environment ID parameters. See Environment Manifest (env.yaml) for details.

ueDescription :: Lens' UpdateEnvironment (Maybe Text)

If this parameter is specified, AWS Elastic Beanstalk updates the description of this environment.

Destructuring the Response

environmentDescription :: EnvironmentDescription

Creates a value of EnvironmentDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eStatus - The current operational status of the environment: * Launching : Environment is in the process of initial deployment. * Updating : Environment is in the process of updating its configuration settings or application version. * Ready : Environment is available to have an action performed on it, such as update or terminate. * Terminating : Environment is in the shut-down process. * Terminated : Environment is not running.

eCNAME - The URL to the CNAME for this environment.

eTemplateName - The name of the configuration template used to originally launch this environment.

eAbortableOperationInProgress - Indicates if there is an in-progress environment configuration update or application version deployment that you can cancel. true: There is an update in progress. false: There are no updates currently in progress.

eEndpointURL - For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

eResources - The description of the AWS resources used by this environment.

eDateUpdated - The last modified date for this environment.

eDateCreated - The creation date for this environment.

eHealth - Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment: * Red : Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. * Yellow : Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. * Green : Indicates the environment is healthy and fully functional. * Grey : Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey

eVersionLabel - The application version deployed in this environment.

ePlatformARN - The ARN of the platform.

eTier - Describes the current tier of this environment.

eEnvironmentName - The name of this environment.

eApplicationName - The name of the application associated with this environment.

eEnvironmentARN - The environment's Amazon Resource Name (ARN), which can be used in other API reuqests that require an ARN.

eSolutionStackName - The name of the SolutionStack deployed with this environment.

eEnvironmentId - The ID of this environment.

eHealthStatus - Returns the health status of the application running in your environment. For more information, see Health Colors and Statuses .

eEnvironmentLinks - A list of links to other environments in the same group.

eDescription - Describes this environment.

data EnvironmentDescription

Describes the properties of an environment.

See: environmentDescription smart constructor.

Response Lenses

eStatus :: Lens' EnvironmentDescription (Maybe EnvironmentStatus)

The current operational status of the environment: * Launching : Environment is in the process of initial deployment. * Updating : Environment is in the process of updating its configuration settings or application version. * Ready : Environment is available to have an action performed on it, such as update or terminate. * Terminating : Environment is in the shut-down process. * Terminated : Environment is not running.

eCNAME :: Lens' EnvironmentDescription (Maybe Text)

The URL to the CNAME for this environment.

eTemplateName :: Lens' EnvironmentDescription (Maybe Text)

The name of the configuration template used to originally launch this environment.

eAbortableOperationInProgress :: Lens' EnvironmentDescription (Maybe Bool)

Indicates if there is an in-progress environment configuration update or application version deployment that you can cancel. true: There is an update in progress. false: There are no updates currently in progress.

eEndpointURL :: Lens' EnvironmentDescription (Maybe Text)

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

eResources :: Lens' EnvironmentDescription (Maybe EnvironmentResourcesDescription)

The description of the AWS resources used by this environment.

eDateUpdated :: Lens' EnvironmentDescription (Maybe UTCTime)

The last modified date for this environment.

eDateCreated :: Lens' EnvironmentDescription (Maybe UTCTime)

The creation date for this environment.

eHealth :: Lens' EnvironmentDescription (Maybe EnvironmentHealth)

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment: * Red : Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. * Yellow : Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. * Green : Indicates the environment is healthy and fully functional. * Grey : Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey

eVersionLabel :: Lens' EnvironmentDescription (Maybe Text)

The application version deployed in this environment.

ePlatformARN :: Lens' EnvironmentDescription (Maybe Text)

The ARN of the platform.

eTier :: Lens' EnvironmentDescription (Maybe EnvironmentTier)

Describes the current tier of this environment.

eEnvironmentName :: Lens' EnvironmentDescription (Maybe Text)

The name of this environment.

eApplicationName :: Lens' EnvironmentDescription (Maybe Text)

The name of the application associated with this environment.

eEnvironmentARN :: Lens' EnvironmentDescription (Maybe Text)

The environment's Amazon Resource Name (ARN), which can be used in other API reuqests that require an ARN.

eSolutionStackName :: Lens' EnvironmentDescription (Maybe Text)

The name of the SolutionStack deployed with this environment.

eEnvironmentId :: Lens' EnvironmentDescription (Maybe Text)

The ID of this environment.

eHealthStatus :: Lens' EnvironmentDescription (Maybe EnvironmentHealthStatus)

Returns the health status of the application running in your environment. For more information, see Health Colors and Statuses .

eEnvironmentLinks :: Lens' EnvironmentDescription [EnvironmentLink]

A list of links to other environments in the same group.

eDescription :: Lens' EnvironmentDescription (Maybe Text)

Describes this environment.