Network.AWS.ElasticBeanstalk.V2010_12_01.DescribeApplicationVersions

Description

Returns descriptions for existing application versions. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &Operation=DescribeApplicationVersions &AuthParams amazonaws.com sample.war Version1 description SampleApp 2010-11-17T03:21:59.161Z 2010-11-17T03:21:59.161Z 773cd80a-f26c-11df-8a78-9f77047e0d0c.

Request

data DescribeApplicationVersions

Result message containing a list of configuration descriptions.

Request constructor

mkDescribeApplicationVersions :: DescribeApplicationVersions

Smart constructor for the minimum required parameters to construct a valid DescribeApplicationVersions request.

Request lenses

dav1ApplicationName :: Lens' DescribeApplicationVersions (Maybe Text)

If specified, AWS Elastic Beanstalk restricts the returned descriptions to only include ones that are associated with the specified application.

dav1VersionLabels :: Lens' DescribeApplicationVersions [Text]

If specified, restricts the returned descriptions to only include ones that have the specified version labels.

Response

data DescribeApplicationVersionsResponse

Result message wrapping a list of application version descriptions.

Response constructor

mkDescribeApplicationVersionsResponse :: DescribeApplicationVersionsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeApplicationVersionsResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

davrApplicationVersions :: Lens' DescribeApplicationVersionsResponse [ApplicationVersionDescription]

A list of ApplicationVersionDescription .