Network.AWS.ElasticBeanstalk.V2010_12_01.ListAvailableSolutionStacks

Description

Returns a list of the available solution stack names. https:/elasticbeanstalk.us-east-1.amazon.com?Operation=ListAvailableSolutionStacks &AuthParams 64bit Amazon Linux running Tomcat 6 32bit Amazon Linux running Tomcat 6 64bit Amazon Linux running Tomcat 7 32bit Amazon Linux running Tomcat 7 f21e2a92-f1fc-11df-8a78-9f77047e0d0c.

Request

data ListAvailableSolutionStacks

Request constructor

mkListAvailableSolutionStacks :: ListAvailableSolutionStacks

Smart constructor for the minimum required parameters to construct a valid ListAvailableSolutionStacks request.

Response

data ListAvailableSolutionStacksResponse

A list of available AWS Elastic Beanstalk solution stacks.

Response constructor

mkListAvailableSolutionStacksResponse :: ListAvailableSolutionStacksResponse

Smart constructor for the minimum required parameters to construct a valid ListAvailableSolutionStacksResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

SolutionStacks :: [Text]

SolutionStackDetails :: [SolutionStackDescription]

Response lenses

lassrSolutionStacks :: Lens' ListAvailableSolutionStacksResponse [Text]

A list of available solution stacks.

lassrSolutionStackDetails :: Lens' ListAvailableSolutionStacksResponse [SolutionStackDescription]

A list of available solution stacks and their SolutionStackDescription.