Network.AWS.ElasticBeanstalk.V2010_12_01.RetrieveEnvironmentInfo

Description

Retrieves the compiled information from a RequestEnvironmentInfo request. Related Topics RequestEnvironmentInfo https:/elasticbeanstalk.us-east-1.amazon.com?EnvironmentId=e-hc8mvnayrx &EnvironmentName=SampleAppVersion &InfoType=tail &Operation=RetrieveEnvironmentInfo &AuthParams https:/elasticbeanstalk.us-east-1.s3.amazonaws.comenvironments%2Fa514386a-709f-4888-9683-068c38d744b4%2Flogs%2Fi-92a3ceff%2F278756a8-7d83-4bc1-93db-b1763163705a.log?Expires=1291236023 &AuthParams 2010-11-17T20:40:23.210Z tail i-92a3ceff e8e785c9-f28a-11df-8a78-9f77047e0d0c.

Documentation

retrieveEnvironmentInfo

Arguments

:: EnvironmentInfoType
_reinInfoType

-> RetrieveEnvironmentInfo

Minimum specification for a RetrieveEnvironmentInfo request.

data RetrieveEnvironmentInfo

Constructors

RetrieveEnvironmentInfo

Fields

_reinInfoType :: EnvironmentInfoType

The type of information to retrieve.

_reinEnvironmentId :: Maybe Text

The ID of the data's environment. If no such environment is found, returns an InvalidParameterValue error. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

_reinEnvironmentName :: Maybe Text

The name of the data's environment. If no such environment is found, returns an InvalidParameterValue error. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

reinInfoType :: Lens' RetrieveEnvironmentInfo EnvironmentInfoType

reinEnvironmentName :: Lens' RetrieveEnvironmentInfo (Maybe Text)

reinEnvironmentId :: Lens' RetrieveEnvironmentInfo (Maybe Text)

data RetrieveEnvironmentInfoResponse

Constructors

RetrieveEnvironmentInfoResponse

Fields

_reirmEnvironmentInfo :: [EnvironmentInfoDescription]

The EnvironmentInfoDescription of the environment.

reirmEnvironmentInfo :: Iso' RetrieveEnvironmentInfoResponse [EnvironmentInfoDescription]