Network.AWS.ElasticBeanstalk.V2010_12_01.Types

Description

AWS Elastic Beanstalk is an easy-to-use service for deploying and scaling web applications and services developed with Java, .NET, PHP, Node.js, Python, Ruby, and Docker on familiar servers such as Apache HTTP Server, Apache Tomcat, Nginx, Passenger, and IIS 7.5/8. You can simply upload your code and Elastic Beanstalk automatically handles the deployment, from capacity provisioning, load balancing, auto-scaling to application health monitoring. At the same time, you retain full control over the AWS resources powering your application and can access the underlying resources at any time. There is no additional charge for Elastic Beanstalk - you pay only for the AWS resources needed to store and run your applications.

Service

data ElasticBeanstalk

Supported version (2010-12-01) of the AWS Elastic Beanstalk service.

Errors

data family Er a

XML

xmlOptions :: Tagged a XMLOptions

ConfigurationDeploymentStatus

data ConfigurationDeploymentStatus

Constructors

ConfigurationDeploymentStatusDeployed
deployed

ConfigurationDeploymentStatusFailed
failed

ConfigurationDeploymentStatusPending
pending

ConfigurationOptionValueType

data ConfigurationOptionValueType

Constructors

ConfigurationOptionValueTypeList
List

ConfigurationOptionValueTypeScalar
Scalar

EnvironmentHealth

data EnvironmentHealth

Constructors

EnvironmentHealthGreen
Green

EnvironmentHealthGrey
Grey

EnvironmentHealthRed
Red

EnvironmentHealthYellow
Yellow

EnvironmentInfoType

data EnvironmentInfoType

Constructors

EnvironmentInfoTypeTail
tail

EnvironmentStatus

data EnvironmentStatus

Constructors

EnvironmentStatusLaunching
Launching

EnvironmentStatusReady
Ready

EnvironmentStatusTerminated
Terminated

EnvironmentStatusTerminating
Terminating

EnvironmentStatusUpdating
Updating

EventSeverity

data EventSeverity

Constructors

EventSeverityDebug
DEBUG

EventSeverityError
ERROR

EventSeverityFatal
FATAL

EventSeverityInfo
INFO

EventSeverityTrace
TRACE

EventSeverityWarn
WARN

ValidationSeverity

data ValidationSeverity

Constructors

ValidationSeverityError
error

ValidationSeverityWarning
warning

AutoScalingGroup

data AutoScalingGroup

Describes an Auto Scaling launch configuration.

mkAutoScalingGroup :: AutoScalingGroup

Smart constructor for the minimum required fields to construct a valid AutoScalingGroup data type to populate a request.

asgName :: Lens' AutoScalingGroup (Maybe Text)

The name of the AutoScalingGroup .

EnvironmentResourcesDescription

data EnvironmentResourcesDescription

The description of the AWS resources used by this environment.

mkEnvironmentResourcesDescription :: EnvironmentResourcesDescription

Smart constructor for the minimum required fields to construct a valid EnvironmentResourcesDescription data type to populate a request.

erdLoadBalancer :: Lens' EnvironmentResourcesDescription (Maybe LoadBalancerDescription)

Describes the LoadBalancer.

Instance

data Instance

The description of an Amazon EC2 instance.

mkInstance :: Instance

Smart constructor for the minimum required fields to construct a valid Instance data type to populate a request.

iId :: Lens' Instance (Maybe Text)

The ID of the Amazon EC2 instance.

LaunchConfiguration

data LaunchConfiguration

Describes an Auto Scaling launch configuration.

mkLaunchConfiguration :: LaunchConfiguration

Smart constructor for the minimum required fields to construct a valid LaunchConfiguration data type to populate a request.

lcName :: Lens' LaunchConfiguration (Maybe Text)

The name of the launch configuration.

LoadBalancer

data LoadBalancer

Describes a LoadBalancer.

mkLoadBalancer :: LoadBalancer

Smart constructor for the minimum required fields to construct a valid LoadBalancer data type to populate a request.

lbName :: Lens' LoadBalancer (Maybe Text)

The name of the LoadBalancer.

Trigger

data Trigger

Describes a trigger.

mkTrigger :: Trigger

Smart constructor for the minimum required fields to construct a valid Trigger data type to populate a request.

trsName :: Lens' Trigger (Maybe Text)

The name of the trigger.

ApplicationDescription

data ApplicationDescription

The ApplicationDescription of the application.

mkApplicationDescription :: ApplicationDescription

Smart constructor for the minimum required fields to construct a valid ApplicationDescription data type to populate a request.

adApplicationName :: Lens' ApplicationDescription (Maybe Text)

The name of the application.

adDescription :: Lens' ApplicationDescription (Maybe Text)

User-defined description of the application.

adDateCreated :: Lens' ApplicationDescription (Maybe ISO8601)

The date when the application was created.

adDateUpdated :: Lens' ApplicationDescription (Maybe ISO8601)

The date when the application was last modified.

adVersions :: Lens' ApplicationDescription [Text]

The names of the versions for this application.

adConfigurationTemplates :: Lens' ApplicationDescription [Text]

The names of the configuration templates associated with this application.

ApplicationVersionDescription

data ApplicationVersionDescription

The ApplicationVersionDescription of the application version.

mkApplicationVersionDescription :: ApplicationVersionDescription

Smart constructor for the minimum required fields to construct a valid ApplicationVersionDescription data type to populate a request.

avdApplicationName :: Lens' ApplicationVersionDescription (Maybe Text)

The name of the application associated with this release.

avdDescription :: Lens' ApplicationVersionDescription (Maybe Text)

The description of this application version.

avdVersionLabel :: Lens' ApplicationVersionDescription (Maybe Text)

A label uniquely identifying the version for the associated application.

avdSourceBundle :: Lens' ApplicationVersionDescription (Maybe S3Location)

The location where the source bundle is located for this version.

avdDateCreated :: Lens' ApplicationVersionDescription (Maybe ISO8601)

The creation date of the application version.

avdDateUpdated :: Lens' ApplicationVersionDescription (Maybe ISO8601)

The last modified date of the application version.

ConfigurationOptionDescription

data ConfigurationOptionDescription

Describes the possible values for a configuration option.

mkConfigurationOptionDescription :: ConfigurationOptionDescription

Smart constructor for the minimum required fields to construct a valid ConfigurationOptionDescription data type to populate a request.

codNamespace :: Lens' ConfigurationOptionDescription (Maybe Text)

A unique namespace identifying the option's associated AWS resource.

codName :: Lens' ConfigurationOptionDescription (Maybe Text)

The name of the configuration option.

codDefaultValue :: Lens' ConfigurationOptionDescription (Maybe Text)

The default value for this configuration option.

codChangeSeverity :: Lens' ConfigurationOptionDescription (Maybe Text)

An indication of which action is required if the value for this configuration option changes: NoInterruption - There is no interruption to the environment or application availability. RestartEnvironment - The environment is restarted, all AWS resources are deleted and recreated, and the environment is unavailable during the process. RestartApplicationServer - The environment is available the entire time. However, a short application outage occurs when the application servers on the running Amazon EC2 instances are restarted. NoInterruption : There is no interruption to the environment or application availability. RestartEnvironment : The environment is entirely restarted, all AWS resources are deleted and recreated, and the environment is unavailable during the process. RestartApplicationServer : The environment is available the entire time. However, a short application outage occurs when the application servers on the running Amazon EC2 instances are restarted.

codUserDefined :: Lens' ConfigurationOptionDescription (Maybe Bool)

An indication of whether the user defined this configuration option: true : This configuration option was defined by the user. It is a valid choice for specifying this as an Option to Remove when updating configuration settings. false : This configuration was not defined by the user. true : This configuration option was defined by the user. It is a valid choice for specifying if this as an Option to Remove when updating configuration settings. false : This configuration was not defined by the user. Constraint: You can remove only UserDefined options from a configuration. Valid Values: true | false.

codValueType :: Lens' ConfigurationOptionDescription (Maybe ConfigurationOptionValueType)

An indication of which type of values this option has and whether it is allowable to select one or more than one of the possible values: Scalar : Values for this option are a single selection from the possible values, or a unformatted string or numeric value governed by the MINMAXRegex constraints: List : Values for this option are multiple selections of the possible values. Boolean : Values for this option are either true or false . Scalar : Values for this option are a single selection from the possible values, or an unformatted string, or numeric value governed by the MINMAXRegex constraints. List : Values for this option are multiple selections from the possible values. Boolean : Values for this option are either true or false .

codValueOptions :: Lens' ConfigurationOptionDescription [Text]

If specified, values for the configuration option are selected from this list.

codMinValue :: Lens' ConfigurationOptionDescription (Maybe Integer)

If specified, the configuration option must be a numeric value greater than this value.

codMaxValue :: Lens' ConfigurationOptionDescription (Maybe Integer)

If specified, the configuration option must be a numeric value less than this value.

codMaxLength :: Lens' ConfigurationOptionDescription (Maybe Integer)

If specified, the configuration option must be a string value no longer than this value.

codRegex :: Lens' ConfigurationOptionDescription (Maybe OptionRestrictionRegex)

If specified, the configuration option must be a string value that satisfies this regular expression.

ConfigurationOptionSetting

data ConfigurationOptionSetting

A specification identifying an individual configuration option along with its current value. For a list of possible option values, go to Option Values in the AWS Elastic Beanstalk Developer Guide.

mkConfigurationOptionSetting :: ConfigurationOptionSetting

Smart constructor for the minimum required fields to construct a valid ConfigurationOptionSetting data type to populate a request.

cosNamespace :: Lens' ConfigurationOptionSetting (Maybe Text)

A unique namespace identifying the option's associated AWS resource.

cosOptionName :: Lens' ConfigurationOptionSetting (Maybe Text)

The name of the configuration option.

cosValue :: Lens' ConfigurationOptionSetting (Maybe Text)

The current value for the configuration option.

ConfigurationSettingsDescription

data ConfigurationSettingsDescription

Describes the settings for a configuration set.

mkConfigurationSettingsDescription :: ConfigurationSettingsDescription

Smart constructor for the minimum required fields to construct a valid ConfigurationSettingsDescription data type to populate a request.

csdSolutionStackName :: Lens' ConfigurationSettingsDescription (Maybe Text)

The name of the solution stack this configuration set uses.

csdApplicationName :: Lens' ConfigurationSettingsDescription (Maybe Text)

The name of the application associated with this configuration set.

csdTemplateName :: Lens' ConfigurationSettingsDescription (Maybe Text)

If not null, the name of the configuration template for this configuration set.

csdDescription :: Lens' ConfigurationSettingsDescription (Maybe Text)

Describes this configuration set.

csdEnvironmentName :: Lens' ConfigurationSettingsDescription (Maybe Text)

If not null, the name of the environment for this configuration set.

csdDeploymentStatus :: Lens' ConfigurationSettingsDescription (Maybe ConfigurationDeploymentStatus)

If this configuration set is associated with an environment, the DeploymentStatus parameter indicates the deployment status of this configuration set: null: This configuration is not associated with a running environment. pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying. deployed: This is the configuration that is currently deployed to the associated running environment. failed: This is a draft configuration, that failed to successfully deploy. null: This configuration is not associated with a running environment. pending: This is a draft configuration that is not deployed to the associated environment but is in the process of deploying. deployed: This is the configuration that is currently deployed to the associated running environment. failed: This is a draft configuration that failed to successfully deploy.

csdDateCreated :: Lens' ConfigurationSettingsDescription (Maybe ISO8601)

The date (in UTC time) when this configuration set was created.

csdDateUpdated :: Lens' ConfigurationSettingsDescription (Maybe ISO8601)

The date (in UTC time) when this configuration set was last modified.

csdOptionSettings :: Lens' ConfigurationSettingsDescription [ConfigurationOptionSetting]

A list of the configuration options and their values in this configuration set.

EnvironmentDescription

data EnvironmentDescription

Describes the properties of an environment.

mkEnvironmentDescription :: EnvironmentDescription

Smart constructor for the minimum required fields to construct a valid EnvironmentDescription data type to populate a request.

edEnvironmentName :: Lens' EnvironmentDescription (Maybe Text)

The name of this environment.

edEnvironmentId :: Lens' EnvironmentDescription (Maybe Text)

The ID of this environment.

edApplicationName :: Lens' EnvironmentDescription (Maybe Text)

The name of the application associated with this environment.

edVersionLabel :: Lens' EnvironmentDescription (Maybe Text)

The application version deployed in this environment.

edSolutionStackName :: Lens' EnvironmentDescription (Maybe Text)

The name of the SolutionStack deployed with this environment.

edTemplateName :: Lens' EnvironmentDescription (Maybe Text)

The name of the configuration template used to originally launch this environment.

edDescription :: Lens' EnvironmentDescription (Maybe Text)

Describes this environment.

edEndpointURL :: Lens' EnvironmentDescription (Maybe Text)

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

edCNAME :: Lens' EnvironmentDescription (Maybe Text)

The URL to the CNAME for this environment.

edDateCreated :: Lens' EnvironmentDescription (Maybe ISO8601)

The creation date for this environment.

edDateUpdated :: Lens' EnvironmentDescription (Maybe ISO8601)

The last modified date for this environment.

edStatus :: Lens' EnvironmentDescription (Maybe EnvironmentStatus)

The current operational status of the environment: Launching: Environment is in the process of initial deployment. Updating: Environment is in the process of updating its configuration settings or application version. Ready: Environment is available to have an action performed on it, such as update or terminate. Terminating: Environment is in the shut-down process. Terminated: Environment is not running.

edHealth :: Lens' EnvironmentDescription (Maybe EnvironmentHealth)

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment: Red : Indicates the environment is not working. Yellow: Indicates that something is wrong, the application might not be available, but the instances appear running. Green: Indicates the environment is healthy and fully functional. Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. Green: Indicates the environment is healthy and fully functional. Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey.

edResources :: Lens' EnvironmentDescription (Maybe EnvironmentResourcesDescription)

The description of the AWS resources used by this environment.

edTier :: Lens' EnvironmentDescription (Maybe EnvironmentTier)

Describes the current tier of this environment.

EnvironmentInfoDescription

data EnvironmentInfoDescription

The information retrieved from the Amazon EC2 instances.

mkEnvironmentInfoDescription :: EnvironmentInfoDescription

Smart constructor for the minimum required fields to construct a valid EnvironmentInfoDescription data type to populate a request.

eidInfoType :: Lens' EnvironmentInfoDescription (Maybe EnvironmentInfoType)

The type of information retrieved.

eidEc2InstanceId :: Lens' EnvironmentInfoDescription (Maybe Text)

The Amazon EC2 Instance ID for this information.

eidSampleTimestamp :: Lens' EnvironmentInfoDescription (Maybe ISO8601)

The time stamp when this information was retrieved.

eidMessage :: Lens' EnvironmentInfoDescription (Maybe Text)

The retrieved information.

EnvironmentResourceDescription

data EnvironmentResourceDescription

A list of EnvironmentResourceDescription.

mkEnvironmentResourceDescription :: EnvironmentResourceDescription

Smart constructor for the minimum required fields to construct a valid EnvironmentResourceDescription data type to populate a request.

erdrsEnvironmentName :: Lens' EnvironmentResourceDescription (Maybe Text)

The name of the environment.

erdrsAutoScalingGroups :: Lens' EnvironmentResourceDescription [AutoScalingGroup]

The AutoScalingGroups used by this environment.

erdrsInstances :: Lens' EnvironmentResourceDescription [Instance]

The Amazon EC2 instances used by this environment.

erdrsLaunchConfigurations :: Lens' EnvironmentResourceDescription [LaunchConfiguration]

The Auto Scaling launch configurations in use by this environment.

erdrsLoadBalancers :: Lens' EnvironmentResourceDescription [LoadBalancer]

The LoadBalancers in use by this environment.

erdrsTriggers :: Lens' EnvironmentResourceDescription [Trigger]

The AutoScaling triggers in use by this environment.

erdrsQueues :: Lens' EnvironmentResourceDescription [Queue]

The queues used by this environment.

EnvironmentTier

data EnvironmentTier

This specifies the tier to use for creating this environment.

mkEnvironmentTier :: EnvironmentTier

Smart constructor for the minimum required fields to construct a valid EnvironmentTier data type to populate a request.

etName :: Lens' EnvironmentTier (Maybe Text)

The name of this environment tier.

etType :: Lens' EnvironmentTier (Maybe Text)

The type of this environment tier.

etVersion :: Lens' EnvironmentTier (Maybe Text)

The version of this environment tier.

EventDescription

data EventDescription

Describes an event.

mkEventDescription :: EventDescription

Smart constructor for the minimum required fields to construct a valid EventDescription data type to populate a request.

edrsEventDate :: Lens' EventDescription (Maybe ISO8601)

The date when the event occurred.

edrsMessage :: Lens' EventDescription (Maybe Text)

The event message.

edrsApplicationName :: Lens' EventDescription (Maybe Text)

The application associated with the event.

edrsVersionLabel :: Lens' EventDescription (Maybe Text)

The release label for the application version associated with this event.

edrsTemplateName :: Lens' EventDescription (Maybe Text)

The name of the configuration associated with this event.

edrsEnvironmentName :: Lens' EventDescription (Maybe Text)

The name of the environment associated with this event.

edrsRequestId :: Lens' EventDescription (Maybe Text)

The web service request ID for the activity of this event.

edrsSeverity :: Lens' EventDescription (Maybe EventSeverity)

The severity level of this event.

Listener

data Listener

Describes the properties of a Listener for the LoadBalancer.

mkListener :: Listener

Smart constructor for the minimum required fields to construct a valid Listener data type to populate a request.

lProtocol :: Lens' Listener (Maybe Text)

The protocol that is used by the Listener.

lPort :: Lens' Listener (Maybe Integer)

The port that is used by the Listener.

LoadBalancerDescription

data LoadBalancerDescription

Describes the LoadBalancer.

mkLoadBalancerDescription :: LoadBalancerDescription

Smart constructor for the minimum required fields to construct a valid LoadBalancerDescription data type to populate a request.

lbdLoadBalancerName :: Lens' LoadBalancerDescription (Maybe Text)

The name of the LoadBalancer.

lbdDomain :: Lens' LoadBalancerDescription (Maybe Text)

The domain name of the LoadBalancer.

lbdListeners :: Lens' LoadBalancerDescription [Listener]

A list of Listeners used by the LoadBalancer.

OptionRestrictionRegex

data OptionRestrictionRegex

If specified, the configuration option must be a string value that satisfies this regular expression.

mkOptionRestrictionRegex :: OptionRestrictionRegex

Smart constructor for the minimum required fields to construct a valid OptionRestrictionRegex data type to populate a request.

orrPattern :: Lens' OptionRestrictionRegex (Maybe Text)

The regular expression pattern that a string configuration option value with this restriction must match.

orrLabel :: Lens' OptionRestrictionRegex (Maybe Text)

A unique name representing this regular expression.

OptionSpecification

data OptionSpecification

A specification identifying an individual configuration option.

mkOptionSpecification :: OptionSpecification

Smart constructor for the minimum required fields to construct a valid OptionSpecification data type to populate a request.

osNamespace :: Lens' OptionSpecification (Maybe Text)

A unique namespace identifying the option's associated AWS resource.

osOptionName :: Lens' OptionSpecification (Maybe Text)

The name of the configuration option.

Queue

data Queue

Describes a queue.

mkQueue :: Queue

Smart constructor for the minimum required fields to construct a valid Queue data type to populate a request.

qName :: Lens' Queue (Maybe Text)

The name of the queue.

qURL :: Lens' Queue (Maybe Text)

The URL of the queue.

S3Location

data S3Location

The Amazon S3 bucket and key that identify the location of the source bundle for this version. If data found at the Amazon S3 location exceeds the maximum allowed source bundle size, AWS Elastic Beanstalk returns an InvalidParameterValue error. The maximum size allowed is 512 MB. Default: If not specified, AWS Elastic Beanstalk uses a sample application. If only partially specified (for example, a bucket is provided but not the key) or if no data is found at the Amazon S3 location, AWS Elastic Beanstalk returns an InvalidParameterCombination error.

mkS3Location :: S3Location

Smart constructor for the minimum required fields to construct a valid S3Location data type to populate a request.

slS3Bucket :: Lens' S3Location (Maybe Text)

The Amazon S3 bucket where the data is located.

slS3Key :: Lens' S3Location (Maybe Text)

The Amazon S3 key where the data is located.

SolutionStackDescription

data SolutionStackDescription

Describes the solution stack.

mkSolutionStackDescription :: SolutionStackDescription

Smart constructor for the minimum required fields to construct a valid SolutionStackDescription data type to populate a request.

ssdSolutionStackName :: Lens' SolutionStackDescription (Maybe Text)

The name of the solution stack.

ssdPermittedFileTypes :: Lens' SolutionStackDescription [Text]

The permitted file types allowed for a solution stack.

SourceConfiguration

data SourceConfiguration

If specified, AWS Elastic Beanstalk uses the configuration values from the specified configuration template to create a new configuration. Values specified in the OptionSettings parameter of this call overrides any values obtained from the SourceConfiguration. If no configuration template is found, returns an InvalidParameterValue error. Constraint: If both the solution stack name parameter and the source configuration parameters are specified, the solution stack of the source configuration template must match the specified solution stack name or else AWS Elastic Beanstalk returns an InvalidParameterCombination error.

mkSourceConfiguration :: SourceConfiguration

Smart constructor for the minimum required fields to construct a valid SourceConfiguration data type to populate a request.

scApplicationName :: Lens' SourceConfiguration (Maybe Text)

The name of the application associated with the configuration.

scTemplateName :: Lens' SourceConfiguration (Maybe Text)

The name of the configuration template.

Tag

data Tag

Describes a tag applied to a resource in an environment.

mkTag :: Tag

Smart constructor for the minimum required fields to construct a valid Tag data type to populate a request.

tKey :: Lens' Tag (Maybe Text)

The key of the tag.

tValue :: Lens' Tag (Maybe Text)

The value of the tag.

ValidationMessage

data ValidationMessage

An error or warning for a desired configuration option value.

mkValidationMessage :: ValidationMessage

Smart constructor for the minimum required fields to construct a valid ValidationMessage data type to populate a request.

vmMessage :: Lens' ValidationMessage (Maybe Text)

A message describing the error or warning.

vmSeverity :: Lens' ValidationMessage (Maybe ValidationSeverity)

An indication of the severity of this message: error: This message indicates that this is not a valid setting for an option. warning: This message is providing information you should take into account. error: This message indicates that this is not a valid setting for an option. warning: This message is providing information you should take into account.

vmNamespace :: Lens' ValidationMessage (Maybe Text)

vmOptionName :: Lens' ValidationMessage (Maybe Text)