Network.AWS.ElasticBeanstalk.V2010_12_01.UpdateApplicationVersion

Description

Updates the specified application version to have the specified properties. If a property (for example, description) is not provided, the value remains unchanged. To clear properties, specify an empty string. https:/elasticbeanstalk.us-east-1.amazon.com?ApplicationName=SampleApp &VersionLabel=New%20Version &Description=New%20Release%20Description &Operation=UpdateApplicationVersion &AuthParams awsemr sample.war New Version New Release Description SampleApp 2010-11-17T19:26:20.699Z 2010-11-17T20:48:16.632Z 00b10aa1-f28c-11df-8a78-9f77047e0d0c.

Request

data UpdateApplicationVersion

Request constructor

mkUpdateApplicationVersion

Arguments

:: Text
uavApplicationName

-> Text
uavVersionLabel

-> UpdateApplicationVersion

Smart constructor for the minimum required parameters to construct a valid UpdateApplicationVersion request.

Request lenses

uavApplicationName :: Lens' UpdateApplicationVersion Text

The name of the application associated with this version. If no application is found with this name, UpdateApplication returns an InvalidParameterValue error.

uavVersionLabel :: Lens' UpdateApplicationVersion Text

The name of the version to update. If no application version is found with this label, UpdateApplication returns an InvalidParameterValue error.

uavDescription :: Lens' UpdateApplicationVersion (Maybe Text)

A new description for this release.

Response

data UpdateApplicationVersionResponse

Result message wrapping a single description of an application version.

Response lenses

uavrApplicationVersion :: Lens' UpdateApplicationVersionResponse (Maybe ApplicationVersionDescription)

The ApplicationVersionDescription of the application version.