Network.AWS.ElasticBeanstalk.V2010_12_01.UpdateEnvironment

Description

Updates the environment description, deploys a new application version, updates the configuration settings to an entirely new configuration template, or updates select configuration option values in the running environment. Attempting to update both the release and configuration is not allowed and AWS Elastic Beanstalk returns an InvalidParameterCombination error. When updating the configuration settings to a new template or individual settings, a draft configuration is created and DescribeConfigurationSettings for this environment returns two setting descriptions with different DeploymentStatus values. https:/elasticbeanstalk.us-east-1.amazon.com?EnvironmentId=e-hc8mvnayrx &EnvironmentName=SampleAppVersion &TemplateName=default &OptionsToRemove.member.1.Namespace=aws%3Aautoscaling%3Atrigger &OptionsToRemove.member.1.OptionName=MeasureName &Operation=UpdateEnvironment &AuthParams New Version Deploying SampleApp elasticbeanstalk-SampleAppVersion-246126201.us-east-1.elb.amazonaws.com SampleApp.elasticbeanstalk.amazonaws.com Grey e-hc8mvnayrx 2010-11-17T21:05:55.251Z 32bit Amazon Linux running Tomcat 7 SampleAppDescription SampleAppVersion 2010-11-17T20:17:42.339Z 7705f0bc-f28e-11df-8a78-9f77047e0d0c.

Documentation

updateEnvironment :: UpdateEnvironment

Minimum specification for a UpdateEnvironment request.

data UpdateEnvironment

Constructors

UpdateEnvironment

Fields

_uemOptionSettings :: [ConfigurationOptionSetting]

If specified, AWS Elastic Beanstalk updates the configuration set associated with the running environment and sets the specified configuration options to the requested value.

_uemTemplateName :: Maybe Text

If this parameter is specified, AWS Elastic Beanstalk deploys this configuration template to the environment. If no such configuration template is found, AWS Elastic Beanstalk returns an InvalidParameterValue error.

_uemDescription :: Maybe Text

If this parameter is specified, AWS Elastic Beanstalk updates the description of this environment.

_uemEnvironmentId :: Maybe Text

The ID of the environment to update. If no environment with this ID exists, AWS Elastic Beanstalk returns an InvalidParameterValue error. Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

_uemEnvironmentName :: Maybe Text

The name of the environment to update. If no environment with this name exists, AWS Elastic Beanstalk returns an InvalidParameterValue error. Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

_uemTier :: Maybe EnvironmentTier

This specifies the tier to use to update the environment. Condition: You can only update the tier version for an environment. If you change the name of the type, AWS Elastic Beanstalk returns InvalidParameterValue error.

_uemOptionsToRemove :: [OptionSpecification]

A list of custom user-defined configuration options to remove from the configuration set for this environment.

_uemVersionLabel :: Maybe Text

If this parameter is specified, AWS Elastic Beanstalk deploys the named application version to the environment. If no such application version is found, returns an InvalidParameterValue error.

uemVersionLabel :: Lens' UpdateEnvironment (Maybe Text)

uemTier :: Lens' UpdateEnvironment (Maybe EnvironmentTier)

uemTemplateName :: Lens' UpdateEnvironment (Maybe Text)

uemOptionsToRemove :: Lens' UpdateEnvironment [OptionSpecification]

uemOptionSettings :: Lens' UpdateEnvironment [ConfigurationOptionSetting]

uemEnvironmentName :: Lens' UpdateEnvironment (Maybe Text)

uemEnvironmentId :: Lens' UpdateEnvironment (Maybe Text)

uemDescription :: Lens' UpdateEnvironment (Maybe Text)

data UpdateEnvironmentResponse

Constructors

UpdateEnvironmentResponse

Fields

_edApplicationName :: Maybe Text

The name of the application associated with this environment.

_edTemplateName :: Maybe Text

The name of the configuration template used to originally launch this environment.

_edDateCreated :: Maybe ISO8601

The creation date for this environment.

_edCNAME :: Maybe Text

The URL to the CNAME for this environment.

_edDescription :: Maybe Text

Describes this environment.

_edEndpointURL :: Maybe Text

For load-balanced, autoscaling environments, the URL to the LoadBalancer. For single-instance environments, the IP address of the instance.

_edHealth :: Maybe EnvironmentHealth

Describes the health status of the environment. AWS Elastic Beanstalk indicates the failure levels for a running environment: Red : Indicates the environment is not working. Yellow: Indicates that something is wrong, the application might not be available, but the instances appear running. Green: Indicates the environment is healthy and fully functional. Red: Indicates the environment is not responsive. Occurs when three or more consecutive failures occur for an environment. Yellow: Indicates that something is wrong. Occurs when two consecutive failures occur for an environment. Green: Indicates the environment is healthy and fully functional. Grey: Default health for a new environment. The environment is not fully launched and health checks have not started or health checks are suspended during an UpdateEnvironment or RestartEnvironement request. Default: Grey.

_edEnvironmentId :: Maybe Text

The ID of this environment.

_edEnvironmentName :: Maybe Text

The name of this environment.

_edResources :: Maybe EnvironmentResourcesDescription

The description of the AWS resources used by this environment.

_edStatus :: Maybe EnvironmentStatus

The current operational status of the environment: Launching: Environment is in the process of initial deployment. Updating: Environment is in the process of updating its configuration settings or application version. Ready: Environment is available to have an action performed on it, such as update or terminate. Terminating: Environment is in the shut-down process. Terminated: Environment is not running.

_edTier :: Maybe EnvironmentTier

Describes the current tier of this environment.

_edSolutionStackName :: Maybe Text

The name of the SolutionStack deployed with this environment.

_edDateUpdated :: Maybe ISO8601

The last modified date for this environment.

_edVersionLabel :: Maybe Text

The application version deployed in this environment.

edVersionLabel :: Lens' UpdateEnvironmentResponse (Maybe Text)

edTier :: Lens' UpdateEnvironmentResponse (Maybe EnvironmentTier)

edTemplateName :: Lens' UpdateEnvironmentResponse (Maybe Text)

edStatus :: Lens' UpdateEnvironmentResponse (Maybe EnvironmentStatus)

edSolutionStackName :: Lens' UpdateEnvironmentResponse (Maybe Text)

edResources :: Lens' UpdateEnvironmentResponse (Maybe EnvironmentResourcesDescription)

edHealth :: Lens' UpdateEnvironmentResponse (Maybe EnvironmentHealth)

edEnvironmentName :: Lens' UpdateEnvironmentResponse (Maybe Text)

edEnvironmentId :: Lens' UpdateEnvironmentResponse (Maybe Text)

edEndpointURL :: Lens' UpdateEnvironmentResponse (Maybe Text)

edDescription :: Lens' UpdateEnvironmentResponse (Maybe Text)

edDateUpdated :: Lens' UpdateEnvironmentResponse (Maybe ISO8601)

edDateCreated :: Lens' UpdateEnvironmentResponse (Maybe ISO8601)

edCNAME :: Lens' UpdateEnvironmentResponse (Maybe Text)

edApplicationName :: Lens' UpdateEnvironmentResponse (Maybe Text)