Network.AWS.ElasticBeanstalk.ValidateConfigurationSettings

Description

Takes a set of configuration settings and either a configuration template or environment, and determines whether those values are valid.

This action returns a list of messages indicating any errors or warnings associated with the selection of option values.

Creating a Request

validateConfigurationSettings

Arguments

:: Text
vcsApplicationName

-> ValidateConfigurationSettings

Creates a value of ValidateConfigurationSettings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vcsTemplateName - The name of the configuration template to validate the settings against. Condition: You cannot specify both this and an environment name.

vcsEnvironmentName - The name of the environment to validate the settings against. Condition: You cannot specify both this and a configuration template name.

vcsApplicationName - The name of the application that the configuration template or environment belongs to.

vcsOptionSettings - A list of the options and desired values to evaluate.

data ValidateConfigurationSettings

A list of validation messages for a specified configuration template.

See: validateConfigurationSettings smart constructor.

Request Lenses

vcsTemplateName :: Lens' ValidateConfigurationSettings (Maybe Text)

The name of the configuration template to validate the settings against. Condition: You cannot specify both this and an environment name.

vcsEnvironmentName :: Lens' ValidateConfigurationSettings (Maybe Text)

The name of the environment to validate the settings against. Condition: You cannot specify both this and a configuration template name.

vcsApplicationName :: Lens' ValidateConfigurationSettings Text

The name of the application that the configuration template or environment belongs to.

vcsOptionSettings :: Lens' ValidateConfigurationSettings [ConfigurationOptionSetting]

A list of the options and desired values to evaluate.

Destructuring the Response

validateConfigurationSettingsResponse

Arguments

:: Int
vcsrsResponseStatus

-> ValidateConfigurationSettingsResponse

Creates a value of ValidateConfigurationSettingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vcsrsMessages - A list of ValidationMessage .

vcsrsResponseStatus - -- | The response status code.

data ValidateConfigurationSettingsResponse

Provides a list of validation messages.

See: validateConfigurationSettingsResponse smart constructor.

Response Lenses

vcsrsMessages :: Lens' ValidateConfigurationSettingsResponse [ValidationMessage]

A list of ValidationMessage .

vcsrsResponseStatus :: Lens' ValidateConfigurationSettingsResponse Int

- | The response status code.